### 2023-01-29 #5112 @editorialbot
[PRE REVIEW]: ms3: A parser for MuseScore files, serving as data factory for annotated music corpora
**Submitting author:** @johentsch (<a href="http://orcid.org/0000-0002-1986-9545">Johannes Hentschel</a>)
**Repository:** https://github.com/johentsch/ms3
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.1.2
