### 2023-01-30 #5112 @johentsch
Thanks @arfon.

> MISSING DOIs
> - 10.2307/895496 may be a valid DOI for title: Music Notation: A Manual of Modern Practice

To save everyone's time: the DOI suggested by the bot is a book review.
