### 2023-01-31 90c7150
adds 'filter_other_fnames' argument to Corpus.add_dir()
