### 2023-01-31 2b12be5
enables hiding the info prints in operations.insert_labels_into_score()
