+++ src/ms3/operations.py
@@ -12,7 +12,8 @@ def insert_labels_into_score(ms3_object: Union[Parse, Corpus],
                              staff: int = None,
                              voice: Optional[Literal[1, 2, 3, 4]] = None,
                              harmony_layer: Optional[Literal[0, 1, 2]] = None,
-                             check_for_clashes: bool = True
+                             check_for_clashes: bool = True,
+                             print_info: bool = True,
                              ) -> None:
     """ Write labels into the <Harmony> tags of the corresponding MuseScore files.
 
@@ -43,6 +44,9 @@ def insert_labels_into_score(ms3_object: Union[Parse, Corpus],
         check_for_clashes
             By default, warnings are thrown when there already exists a label at a position (and in a notational
             layer) where a new one is attached. Pass False to deactivate these warnings.
+        print_info:
+            By default, the ms3_object is displayed before and after parsing. Pass False to prevent this,
+            for example when the object has many, many files.
     """
     logger = get_logger('ms3.add')
     facet = check_argument_against_literal_type(facet, AnnotationsFacet, logger=logger)
@@ -51,8 +55,10 @@ def insert_labels_into_score(ms3_object: Union[Parse, Corpus],
     ms3_object.disambiguate_facet('scores', ask_for_input=ask_for_input)
     ms3_object.view.fnames_with_incomplete_facets = False
     obj_name = type(ms3_object).__name__.upper()
-    print(f"VIEW ON THE {obj_name} BEFORE PARSING:")
-    ms3_object.info()
+    if print_info:
+        print(f"VIEW ON THE {obj_name} BEFORE PARSING:")
+        ms3_object.info()
+    print(f"PARSING SCORES...")
     ms3_object.parse(parallel=False)
     if replace:
         print("REMOVING LABELS FROM PARSED SCORES...")
@@ -60,8 +66,10 @@ def insert_labels_into_score(ms3_object: Union[Parse, Corpus],
     print("INSERTING LABELS INTO SCORES...")
     ms3_object.load_facet_into_scores(facet)
     ms3_object.insert_detached_labels(staff=staff, voice=voice, harmony_layer=harmony_layer, check_for_clashes=check_for_clashes)
-    print(f"{obj_name} OBJECT AFTER THE OPERATION:")
-    ms3_object.info()
+    if print_info:
+        print(f"{obj_name} OBJECT AFTER THE OPERATION:")
+        ms3_object.info()
+    print("DONE INSERTING.")
 
 def extract(parse_obj: Parse,
             root_dir: Optional[str] = None,
