### 2023-02-03 d10a4c0
removing deprecated elements from tab completion and having them throw DeprecationWarning
