+++ src/ms3/parse.py
@@ -20,6 +20,19 @@ class Parse(LoggedClass):
     Class for creating one or several :class:`~.corpus.Corpus` objects and performing actions on all of them.
     """
 
+    _deprecated_elements = [
+        "parsed_mscx",
+        "parsed_tsv",
+        "add_detached_annotations",
+        "count_annotation_layers",
+        "count_labels",
+        "get_lists",
+        "parse_mscx",
+        "pieces",
+        "store_scores",
+        "update_metadata",
+    ]
+
     def __init__(self,
                  directory: Optional[Union[str, Collection[str]]] = None,
                  recursive: bool = True,
@@ -176,15 +189,7 @@ class Parse(LoggedClass):
         return sum(corpus.n_unparsed_tsvs for _, corpus in self)
 
 
-    @property
-    def parsed_mscx(self) -> pd.DataFrame:
-        """Deprecated property. Replaced by :attr:`n_parsed_scores`"""
-        raise AttributeError(f"Property has been renamed to n_parsed_scores.")
 
-    @property
-    def parsed_tsv(self) -> pd.DataFrame:
-        """Deprecated property. Replaced by :attr:`n_parsed_tsvs`"""
-        raise AttributeError(f"Property has been renamed to n_parsed_tsvs.")
 
     @property
     def view(self) -> View:
@@ -264,104 +269,9 @@ class Parse(LoggedClass):
                                     f"was replaced by {directory}")
         self.corpus_paths[corpus_name] = directory
         self.corpus_objects[corpus_name] = corpus
-        # convertible = self.ms is not None
-        # if file_re is None:
-        #     file_re = Score._make_extension_regex(tsv=True, convertible=convertible)
-        # if exclude_re is None:
-        #     exclude_re = r'(^(\.|_|concatenated_)|_reviewed)'
-        # directory = resolve_dir(directory)
-        # self.last_scanned_dir = directory
-        # if key is None:
-        #     key = os.path.basename(directory)
-        # if key not in self.files:
-        #     self.logger.debug(f"Adding {directory} as new corpus with key '{key}'.")
-        #     self.files[key] = []
-        #     self.corpus_paths[key] = resolve_dir(directory)
-        # else:
-        #     self.logger.info(f"Adding {directory} to existing corpus with key '{key}'.")
-        #
-        # top_level_folders, top_level_files = first_level_files_and_subdirs(directory)
-        # self.logger.debug(f"Top level folders: {top_level_folders}\nTop level files: {top_level_files}")
-        #
-        # added_ids = []
-        #
-        # # look for scores
-        # scores_folder = None
-        # if 'scores' in kwargs and kwargs['scores'] in top_level_folders:
-        #     scores_folder = kwargs['scores']
-        # elif 'MS3' in top_level_folders:
-        #     scores_folder = 'MS3'
-        # elif 'scores' in top_level_folders:
-        #     scores_folder = 'scores'
-        # else:
-        #     msg = f"No scores folder found among {top_level_folders}."
-        #     if 'scores' not in kwargs:
-        #         msg += " If one of them has MuseScore files, indicate it by passing scores='scores_folder'."
-        #     self.logger.info(msg)
-        # if scores_folder is not None:
-        #     score_re = Score._make_extension_regex(convertible=convertible)
-        #     scores_path = os.path.join(directory, scores_folder)
-        #     score_paths = sorted(scan_directory(scores_path, file_re=score_re, recursive=recursive))
-        #     score_ids = self.add_files(paths=score_paths, key=key)
-        #     added_ids += score_ids
-        #     score_fnames = self._get_unambiguous_fnames_from_ids(score_ids, key=key)
-        #
-        #     for fname, id in score_fnames.items():
[…]
