### 2023-02-03 ea17af8
enables view settings when adding new corpora to Parse object
