### 2023-02-03 892a48b
enables setting property .ms everywhere
