+++ src/ms3/parse.py
@@ -145,8 +145,14 @@ class Parse(LoggedClass):
         return self._ms
 
     @ms.setter
-    def ms(self, ms):
-        self._ms = get_musescore(ms, logger=self.logger)
+    def ms(self, ms: Union[str, Literal['auto', 'win', 'mac']]):
+        executable = get_musescore(ms, logger=self.logger)
+        if executable is None:
+            raise FileNotFoundError(f"'{ms}' did not lead me to a MuseScore executable.")
+        if executable is not None:
+            self._ms = executable
+            for _, corpus in self.__iter__():
+                corpus.ms = executable
 
     @property
     def n_detected(self) -> int:
