### 2023-02-07 #5112 @faroit
👋  Hi @johentsch, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

As this toolbox deals with a software useful for music information retrieval research it would be great to have reviewers with domain knowledge. I already asked `@TGabor`, if he could serve as a reviewer.

Furthermore, @johentsch, I am also interested in your input on potential other reviewers so any suggestions you have are also welcome.
For this, please check the JOSS [COI policy](<https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy>) and mention them with a space between the @ and their handle or as a `codeblock` (so you would refer to me as `@faroit`).
