### 2023-02-07 f8bc8d5
catching pandas FutureWarnings, adding type annotations
