+++ src/ms3/utils.py
@@ -681,7 +681,7 @@ def convert_to_ms4(old, new, MS='mscore'):
         if version is None:
             msg = f"{MS} is not MuseScore 4.x.x."
         else:
-            msg = f"{MS3} is MuseScore {version}, not 4. Try convert()."
+            msg = f"{MS} is MuseScore {version}, not 4. Try convert()."
         raise RuntimeError(msg)
     old_path, old_file = os.path.split(old)
     old_fname, old_ext = os.path.splitext(old_file)
@@ -1191,12 +1191,13 @@ def add_collections(left: tuple, right: Collection, dtype: Dtype) -> tuple:
     ...
 def add_collections(left: Union[pd.Series, NDArray, list, tuple],
                     right: Collection,
-                    dtype: Dtype = str):
+                    dtype: Dtype = 'string') -> Union[pd.Series, NDArray, list, tuple]:
+    """Zip-adds together the strings (by default) contained in two collections regardless of their types (think of adding
+    two columns together element-wise). Pass another ``dtype`` if you want the values to be converted to another
+    datatype before adding them together.
+    """
     if isinstance(left, pd.Series) and isinstance(right, pd.Series):
-        try:
-            return left + right
-        except TypeError:
-            return left.astype(dtype) + right.astype(dtype)
+        left.astype(dtype) + right.astype(dtype)
     result_series = pd.Series(left, dtype=dtype) + pd.Series(right, dtype=dtype)
     try:
         return left.__class__(result_series.to_list())
@@ -1218,7 +1219,16 @@ def cast2collection(coll: tuple, func: Callable, *args, **kwargs) -> tuple:
 def cast2collection(coll: Union[pd.Series, NDArray, list, tuple], func: Callable, *args, **kwargs) -> Union[pd.Series, NDArray, list, tuple]:
     if isinstance(coll, pd.Series):
         return transform(coll, func, *args, **kwargs)
-    result_series = func(pd.Series(coll), *args, **kwargs)
+    with warnings.catch_warnings():
+        # pandas developers doing their most annoying thing >:(
+        warnings.filterwarnings(
+            "ignore",
+            category=FutureWarning,
+            message=(
+                ".*The default dtype for empty Series.*"
+            ),
+        )
+        result_series = func(pd.Series(coll), *args, **kwargs)
     try:
         return coll.__class__(result_series.to_list())
     except TypeError:
@@ -1296,33 +1306,38 @@ def fifths2iv(fifths: int,
     return quality + int_num
 
 @overload
-def tpc2name(tpc: int, ms: bool = False, minor: bool = False) -> str:
+def tpc2name(tpc: int, ms: bool = False, minor: bool = False) -> Optional[str]:
     ...
 @overload
-def tpc2name(tpc: pd.Series, ms: bool = False, minor: bool = False) -> pd.Series:
+def tpc2name(tpc: pd.Series, ms: bool = False, minor: bool = False) -> Optional[pd.Series]:
     ...
 @overload
-def tpc2name(tpc: NDArray[int], ms: bool = False, minor: bool = False) -> NDArray[str]:
+def tpc2name(tpc: NDArray[int], ms: bool = False, minor: bool = False) -> Optional[NDArray[str]]:
     ...
 @overload
-def tpc2name(tpc: List[int], ms: bool = False, minor: bool = False) -> List[str]:
+def tpc2name(tpc: List[int], ms: bool = False, minor: bool = False) -> Optional[List[str]]:
     ...
 @overload
-def tpc2name(tpc:  Tuple[int], ms: bool = False, minor: bool = False) -> Tuple[str]:
[…]
