+++ paper/paper.bib
@@ -86,6 +86,23 @@
   file = {/home/hentsche/Zotero/storage/F59YP4JR/London - 2013 - Building a Representative Corpus of Classical Musi.pdf}
 }
 
+@inproceedings{McLeod:21,
+  title={A Modular System for the Harmonic Analysis of Musical Scores using a large vocabulary},
+  author={McLeod, Andrew and Rohrmeier, Martin},
+  booktitle= {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},,
+  year={2021},
+  pages={435--442}
+}
+
+@article{McLeod,
+  author  = {McLeod, Andrew and Rohrmeier, Martin},
+  title   = {An Integrated System for Harmonic Analysis Including Chord Tone Alterations},
+  journal = {Transactions of the International Society for Music Information Retrieval},
+  year    = {To be submitted},
+}
+
+
+
 @phdthesis{Moss2019_TransitionsTonalityModelBased,
   title = {Transitions of {{Tonality}}: {{A Model-Based Corpus Study}}},
   author = {Moss, Fabian Claude},
