### 2023-02-07 #5112 @johentsch
> please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and mention them with a space between the @ and their handle or as a codeblock

Sorry, I had not been very thorough in following your instructions and have now removed the mentions from my previous answer. Also, I have read the COI policy more carefully and have to admit that @allorens might have been a premature suggestion because we are currently preparing a publication together (thanks for volunteering, Ana!). I thought she would be suited because `ms3` was used in her last research project. The decision will be @faroit's but I think it's appropriate to bring this up.
