+++ paper/paper.bib
@@ -89,7 +89,7 @@
 @inproceedings{McLeod:21,
   title={A Modular System for the Harmonic Analysis of Musical Scores using a large vocabulary},
   author={McLeod, Andrew and Rohrmeier, Martin},
-  booktitle= {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},,
+  booktitle= {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
   year={2021},
   pages={435--442}
 }
