+++ paper/paper.bib
@@ -91,7 +91,8 @@
   author={McLeod, Andrew and Rohrmeier, Martin},
   booktitle= {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
   year={2021},
-  pages={435--442}
+  pages={435--442},
+  doi = {10.5281/zenodo.5624432}
 }
 
 @article{McLeod,
