### 2023-02-08 1497476
prevents including 'volta_mcs' in metadata of pieces that don't have voltas
