+++ src/ms3/bs4_parser.py
@@ -1210,8 +1210,6 @@ The first ending MC {mc} is being used. Suppress this warning by using disambigu
             data['length_qb_unfolded'] = round(unfolded_measures.duration_qb.sum(), 2)
         if self.has_voltas:
             data['volta_mcs'] = tuple(tuple(tuple(mcs) for mcs in group.values()) for group in self.volta_structure.values())
-        else:
-            data['volta_mcs'] = None
 
         # labels
         all_labels = self.get_raw_labels()
