### 2023-02-08 7b037fa
adapts utils.infer_tsv_type() to the fact that chord tables now come with 'event' column
