### 2023-02-09 #5112 @malcolmsailor
Hi @faroit I'm open to acting as a reviewer but first I'd like to get some idea what the time frame for completing the review would be.

I also feel I should mention that I know and like @johentsch from corresponding with him by email frequently about our mutual projects and (once) via Zoom. I don't know if that rises to the level of a conflict of interest, but I wanted to disclose it.
