### 2023-02-09 11c834d
makes unknown dtype default to 'string' in CSVW metadata
