### 2023-02-10 b45016d
Piece.get_facet() gets parameter 'force' defaulting to False to avoid unsolicited score parsing
