### 2023-02-11 fdce948
when iterating through pieces, skip fnames that don't have any files under the current view
