### 2023-02-12 #5112 @faroit
@malcolmsailor We usually aim for completing initial reviews within 4-6 weeks.

> I also feel I should mention that I know and like @johentsch from corresponding with him by email frequently about our mutual projects and (once) via Zoom. I don't know if that rises to the level of a conflict of interest, but I wanted to disclose it.

I don't think this is a conflict of interest so I am happy to proceed.
