### 2023-02-18 d260261
segment_by_criterion warns if not IntervalIndex is present
