### 2023-02-19 4b39d17
first version of new docs using notebooks and Pydata-Sphinx-Theme
