+++ docs/build/html/_sources/build/jupyter_execute/build/jupyter_execute/manual.ipynb.txt
@@ -1,0 +1,4163 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "fff39cf4",
+   "metadata": {},
+   "source": [
+    "# Manual\n",
+    "\n",
+    "![](ms3_architecture.png)\n",
+    "\n",
+    "This page is a detailed guide for using ms3 for different tasks. It\n",
+    "supposes you are working in an interactive Python interpreter such as\n",
+    "IPython, Jupyter, Google Colab, or just the console."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "9124ab0b",
+   "metadata": {},
+   "source": [
+    "## Good to know\n",
+    "\n",
+    "(corpss)=\n",
+    "### Corpss structure \n",
+    "\n",
+    "ms3 extracts various aspects (e.g. notes, chord labels, dynamics) from a\n",
+    "single MuseScore file and stores them in separate TSV files. In order to\n",
+    "jointly evaluate this information (e.g. to combine harmony labels with\n",
+    "the corresponding notes in the score), ms3 associates files by their\n",
+    "names. Therefore it is important to not rename files at a later point\n",
+    "and it is recommended to stick to one organizational principle for all\n",
+    "corpora in use. The two main principles are suffixes and subfolders.\n",
+    "\n",
+    "[DCML corpora](https://github.com/DCMLab/dcml_corpora) use the same\n",
+    "subfolder structure: MuseScore files are stored in the `MS3` folder of a\n",
+    "corpus and all other aspects are stored with identical filenames (but\n",
+    "different extension) in sibling folders. This structure results\n",
+    "naturally when using default parameters such as `ms3 extract -N -M -X`:\n",
+    "\n",
+    "``` console\n",
+    "corpus1\n",
+    "├── harmonies\n",
+    "├── measures\n",
+    "├── MS3\n",
+    "└── notes\n",
+    "corpus2\n",
+    "├── harmonies\n",
+    "├── measures\n",
+    "├── MS3\n",
+    "└── notes\n",
+    "```\n",
+    "\n",
+    "When loading corpora, ms3 looks for the standard folder names (and\n",
+    "suffixes) to indentify individual corpora and assign them keys\n",
+    "automatically (e.g. in the example above, \\'corpus1\\' and \\'corpus2\\').\n",
+    "Using the default names will therefore facilitate the use of the library\n",
+    "considerably."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "317f9358",
+   "metadata": {},
+   "source": [
+    "(key_and_id)=\n",
+    "### Keys and IDs\n",
+    "\n",
+    "ms3 uses keys for grouping files. The way how these keys are being used\n",
+    "is transitioning in the 0.5.x versions:\n",
+    "\n",
+    "-   **\\< 0.5** keys were arbitrary and used by some methods to bring\n",
+    "    groups of files together. For example, `Parse.attach_labels()` would\n",
+    "    take a key with a group of scores and, as the parameter\n",
+    "    `annotation_key`, the key with a group of annotations and then use\n",
+    "    `Parse.match_files()` to match the files from the given keys.\n",
+    "-   **\\>= 0.6.x** keys are used to address sub-corpora that are assumed\n",
+    "    to have a particular `corpus_structure`{.interpreted-text\n",
+    "    role=\"ref\"}. The method `Parse.attach_labels()` then takes only one\n",
[…]
