+++ docs/build/html/_sources/index.rst.txt
@@ -5,7 +5,7 @@ Table of Contents
 =================
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
 
    Installation <install>
    Quick Reference <quick>
