+++ docs/build/html/_sources/README.md.txt
@@ -1,0 +1,5 @@
+# Building the docs
+
+1. In the folder `docs`, run `pip install -r requirements.txt`.
+1. Then you can build the current docs using `make html`.
+1. See the output by opening `docs/build/html/index.html`.

[…]
