### 2023-02-22 7b8f70e
end-of-day commig of dezrann.py
collab with Louis Couturier
