+++ src/ms3/dezrann.py
@@ -1,0 +1,243 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+""" 
+DCML to Dezrann
+===============
+
+Script to convert contiguous score annotations from a tabular format (one line per label) into
+the JSON-LD format .dez, used by the Dezrann annotation tool developed at the Algomus group.
+
+# Intro
+
+The script presents a first application of what is to become a formal standard of a "measure map";
+see first discussion points at 
+
+* https://gitlab.com/algomus.fr/dezrann/dezrann/-/issues/1030#note_1122509147)
+* https://github.com/MarkGotham/bar-measure/
+
+As an early proxy of a measure map, the current version uses the measure tables that each
+DCML corpus provides in its `measures` folder. This is beneficial in the current context because:
+
+1. The files are required for correct, actionable quarter-note positions without having to re-parse 
+  the entire score.
+2. The files play an essential role for validating the conversion output.
+3. They help avoiding the confusion that necessarily arises when several addressing schemes are 
+  at play.
+
+In detail:
+
+## 1. Quarterbeats
+
+From a technical perspective, offsets in the sense of "distance from the origin" represent the
+primary mechanism of referencing positions in a text (character counts being the default in NLP).
+Music scores are typically aligned with a time line of "musical time", an alignment which is 
+frequently expressed as float values representing an event's distance from the score's beginning, 
+measured in quarter notes, here referred to as quarterbeats. The fundamental problem, however, is
+ensuring that quarterbeat positions refer to the same time line. The commonplace
+score encoding formats do not indicate quarterbeat positions. Instead, they structure
+musical time in a sequence of containers, generally called "measures", each of which represents 
+a time line starting from 0. Counting measure units (of some kind) therefore represents the second 
+prevalent way of indicating positions in a score, together with an event onset indicating an
+event's distance from the container's beginning. To avoid terminological confusion, we call 
+the distance from the beginning of a measure container "onset".
+
+Looking at a single score, there is an unambiguous mapping between the two types of positions:
+`event_offset = measure_offset + event_onset`. Problems arise, however when information from one
+score is to be set into relation with timed information from another source. This is a wide-spread
+problem in the context of music research and musical corpus studies where data from different
+sources with different ways of expressing timestamps frequently needs to be aligned, often in 
+absence of the original score that one of the source is aligned to. Currently, there is no
+standardized way of storing such alignments for later re-use. Hence the idea of a central
+mapping file for storing alignments between positions given as quarterbeats, measure+onset,
+recording timestamps in seconds, IDs, and other data relevant for score addressability.
+
+**Different types of quarterbeats**
+
+All TSV files issued by the DCML come with the column `quarterbeats` indicating every event's
+offset from the score's beginning (position 0). With the caveat that, in the case of first/second endings 
+("voltas"), the indicated values do not take into account any but the second ending, with the
+rationale that they should represent the temporal proportion of a single playthrough without any
+repetitions. For correct conversion, therefore, using a strict, measuring-stick-based variant
[…]
