+++ src/ms3/dezrann.py
@@ -120,6 +120,7 @@ class DezrannLabel(TypedDict):
     duration: float
     line: str #= "top.3" #Literal?
     tag: str
+    layers: List[str]
 
 class DezrannDict(TypedDict):
     labels: List[DezrannLabel]
@@ -157,23 +158,32 @@ def transform_df(labels: pd.DataFrame,
     transformed_df = pd.concat([quarterbeats.rename('quarterbeats'), labels.duration_qb.rename('duration'), labels[label_column].rename('label')], axis=1)
     return transformed_df.to_dict(orient='records')
     
-def make_dezrann_label(quarterbeats: float, duration: float, label: str) -> DezrannLabel:
-    return DezrannLabel(type="Harmony", start=quarterbeats, duration=duration, line="top.3", tag=label)
+def make_dezrann_label(
+            quarterbeats: float, duration: float, label: str, origin: List[str]) -> DezrannLabel:
+    return DezrannLabel(
+        type="Harmony",
+        start=quarterbeats,
+        duration=duration,
+        line="top.3",
+        tag=label,
+        layers=origin
+    )
 
-def convert_dcml_list_to_dezrann_list(values_dict: List[DcmlLabel]) -> List[DezrannDict]:
+def convert_dcml_list_to_dezrann_list(values_dict: List[DcmlLabel], origin: List[str]) -> List[DezrannDict]:
     label_list = []
     for e in values_dict:
         label_list.append(
             make_dezrann_label(
                 quarterbeats=e["quarterbeats"],
                 duration=e["duration"],
-                label=e["label"]
+                label=e["label"],
+                origin=origin
             )
         )
     return DezrannDict(labels=label_list, meta={"layout": []})
     
 
-def generate_dez(path_measures, path_labels, output_path="labels.dez"): # need paths for harmony.TSV + paths for measures.TSV
+def generate_dez(path_measures, path_labels, output_path="labels.dez", origin: List[str] = ["DCML"]):
     """
     path_measures : :obj:`str`
         Path to a TSV file as output by format_data().
@@ -181,6 +191,8 @@ def generate_dez(path_measures, path_labels, output_path="labels.dez"): # need p
         Path to a TSV file as output by format_data().
     output_labels : :obj:`str`
         Path to a TSV file as output by format_data().
+    origin : :obj:`list`
+        List of source(s) from which the labels originate. Defaults to ["DCML"].
     """
     harmonies = pd.read_csv(
         path_labels, sep='\t',
@@ -193,7 +205,7 @@ def generate_dez(path_measures, path_labels, output_path="labels.dez"): # need p
         converters={'quarterbeats_all_endings': safe_frac}
     )
     dcml_labels = transform_df(labels=harmonies, measures=measures)
-    dezrann_content = convert_dcml_list_to_dezrann_list(dcml_labels)
+    dezrann_content = convert_dcml_list_to_dezrann_list(dcml_labels, origin=origin)
     
     # Manual post-processing  #TODO: improve these cases
     # 1) Avoid NaN values in "duration" (happens in second endings)
