### 2023-02-23 3321020
streamlines CLI arguments and suggests defaults
