+++ src/ms3/dezrann.py
@@ -101,7 +101,7 @@ Here is an example of Dezrann file structure:
 import argparse
 import json
 import os
-from typing import Dict, List, TypedDict, Union, Tuple
+from typing import Dict, List, TypedDict, Union, Tuple, Optional
 
 from fractions import Fraction
 import pandas as pd
@@ -115,10 +115,11 @@ def safe_frac(s: str) -> Union[Fraction, str]:
         return s
 
 class DezrannLabel(TypedDict):
-    type: str #= "Harmony" # Default value ?
+    """Represents one label in a .dez file."""
+    type: str
     start: float
     duration: float
-    line: str #= "top.3" #Literal?
+    #line: str # Determined by the meta-layout
     tag: str
     layers: List[str]
 
@@ -128,13 +129,18 @@ class DezrannDict(TypedDict):
     meta: Dict
 
 class DcmlLabel(TypedDict):
+    """Represents one label from a TSV annotation file"""
     quarterbeats: float
     duration: float
     label: str
+    harmony: str
+    key: str
+    phrase: str
+    cadence: str
 
 
-def transform_df(labels: pd.DataFrame, 
-                measures: pd.DataFrame, 
+def transform_df(labels: pd.DataFrame,
+                measures: pd.DataFrame,
                 label_column: str = 'label') -> List[DcmlLabel]:
     """
 
@@ -166,20 +172,19 @@ def make_dezrann_label(
     else:
         layers = list(origin)
     return DezrannLabel(
-        type="Harmony",
+        type="Harmony", #TODO: adapt type to current label 
         start=quarterbeats,
         duration=duration,
-        line="top.3",
         tag=label,
         layers=layers
     )
 
 def convert_dcml_list_to_dezrann_list(values_dict: List[DcmlLabel],
-                                      cadences: bool,
-                                      harmony_line: int,
-                                      keys_line: int,
-                                      phrases_line: int,
-                                      raw_line: int,
+                                      cadences: bool = False,
+                                      harmony_line: Optional[str] = None,
+                                      keys_line: Optional[str] = None,
+                                      phrases_line: Optional[str] = None,
+                                      raw_line: Optional[str] = None,
                                       origin: Union[str, Tuple[str]] = "DCML") -> DezrannDict:
     label_list = []
     for e in values_dict:
@@ -192,18 +197,28 @@ def convert_dcml_list_to_dezrann_list(values_dict: List[DcmlLabel],
             )
         )
     layout = []
-    if raw_line > 0:
-        layout.append({"filter": {"type": "Harmony"}, "style": {"line": line}})
-    #if harmony_line > 0:
-    #    ...
-    #if keys_line > 0:
-    #    ...
+    if cadences:
+        layout.append({"filter": {"type": "Cadence"}, "style": {"line": "all"}})
+    if harmony_line:
+        layout.append({"filter": {"type": "Harmony"}, "style": {"line": harmony_line}})
+    if keys_line:
+        layout.append({"filter": {"type": "Localkey"}, "style": {"line": keys_line}})
+    if phrases_line:
+        layout.append({"filter": {"type": "Phrase"}, "style": {"line": phrases_line}})
+    if raw_line:
+        layout.append({"filter": {"type": "Harmony"}, "style": {"line": raw_line}})
+
     return DezrannDict(labels=label_list, meta={"layout": layout})
     
 
 def generate_dez(path_measures: str,
                  path_labels: str,
                  output_path: str = "labels.dez",
+                 cadences: bool = False,
+                 harmonies: Optional[str] = None,
+                 keys: Optional[str] = None,
[…]
