### 2023-02-24 8129db4
refines commandline interface with docstrings and better argument treatment; enables deactivating the default layers by passing 0
