+++ src/ms3/dezrann.py
@@ -308,6 +308,9 @@ def main(input_dir: str,
     harmony_measure_matches = []
     for tsv_name in input_files:
         measures_file_path = os.path.join(measures_dir, tsv_name)
+        if not os.path.isfile(measures_file_path):
+            # could be a directory
+            continue
         if os.path.isfile(measures_file_path):
             harmonies_file_path = os.path.join(input_dir, tsv_name)
             harmony_measure_matches.append((harmonies_file_path, measures_file_path))
@@ -345,13 +348,16 @@ LINE_VALUES = {
 }
 
 def transform_line_argument(line: Optional[Union[int, str]]) -> Optional[str]:
+    """Takes a number bet"""
     if line is None:
         return
     try:
         line = int(line)
-        assert line in [1,2,3,4,5,6, -1, -2, -3]
+        assert line in [1,2,3,4,5,6, 0 -1, -2, -3]
     except (TypeError, ValueError, AssertionError):
-        raise ValueError(f"{line} is not a valid argument, shoube within 1-6.")
+        raise ValueError(f"{line} is not a valid argument, should be within [0, 6].")
+    if line == 0:
+        return None
     if line < 0:
         line = abs(line) + 3
     return LINE_VALUES[line]
@@ -367,7 +373,8 @@ def resolve_dir(d):
     return os.path.abspath(d)
     
 
-def process_arguments(args) -> dict:
+def process_arguments(args: argparse.Namespace) -> dict:
+    """Transforms the user's input arguments into keyword arguments for :func:`main` or raises a ValueError."""
     input_dir = resolve_dir(args.dir)
     assert os.path.isdir(input_dir), f"{args.dir} is not an existing directory."
     if args.measures is None:
@@ -395,7 +402,13 @@ def process_arguments(args) -> dict:
         arg_val = getattr(args, arg)
         if arg_val is None:
             continue
-        kwargs[arg] = transform_line_argument(arg_val)
+        line_arg = transform_line_argument(arg_val)
+        if line_arg is None:
+            continue
+        kwargs[arg] = line_arg
+    if len(set(kwargs.values())) < len(kwargs.values()):
+        selected_args = {arg: f"'{getattr(args, arg)}' => {kwargs[arg]}" for arg in line_args if arg in kwargs}
+        raise ValueError(f"You selected the same annotation layer more than once: {selected_args}.")
     if args.cadences:
         kwargs['cadences'] = True
     print(kwargs)
@@ -405,25 +418,26 @@ def process_arguments(args) -> dict:
 def run():
     parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter,
                                      description='''\
-        -----------------------------
-        | DCML => Dezrann converter |
-        -----------------------------
+-----------------------------
+| DCML => Dezrann converter |
+-----------------------------
 
-        This script converts DCML harmony annotations into the .dez JSON format used by the dezrann.net app. It is 
-        standalone and does not require ms3 to be installed. Its only requirement is pandas.
+This script converts DCML harmony annotations into the .dez JSON format used by the dezrann.net app. It is 
+standalone and does not require ms3 to be installed. Its only requirement is pandas.
 
-        Apart from that, the script requires that you have previously extracted both harmonies and measures from the 
-        annotated scores or that you are converting a DCML corpus (https://github.com/DCMLab/dcml_corpora), 
-        where both facets are provided by default. In order to (re-) extract the labels, use the command:
+Apart from that, the script requires that you have previously extracted both harmonies and measures from the 
+annotated scores or that you are converting a DCML corpus (https://github.com/DCMLab/dcml_corpora), 
[…]
