### 2023-02-24 f4310fe
this version converts raw labels, independent of the given line arguments
