+++ src/ms3/dezrann.py
@@ -140,7 +140,7 @@ class DcmlLabel(TypedDict):
 
 
 def transform_df(labels: pd.DataFrame,
-                measures: pd.DataFrame,
+                measures: Optional[pd.DataFrame],
                 label_column: str = 'label') -> List[DcmlLabel]:
     """
 
@@ -148,10 +148,18 @@ def transform_df(labels: pd.DataFrame,
     ----------
     labels:
         Dataframe as found in the 'harmonies'  folder of a DCML corpus. Needs to have columns with
-        the correct dtypes {'mc': int, 'mc_onset': fractions.Fraction} and no missing values.
+        the correct dtypes {'mc': int,
+                            'mc_onset': fractions.Fraction,
+                            'duration_qb': float,
+                            'quarterbeats': fraction.Fraction,
+                            'label': str,
+                            'chord': str,
+                            'cadence': str,
+                            'phraseend': str}
+        and no missing values.
     measures:
-        Dataframe as found in the 'measures' folder of a DCML corpus. Requires the columns
-        {'mc': int, 'quarterbeats_all_endings': fractions.Fraction}
+        (optional) Dataframe as found in the 'measures' folder of a DCML corpus for computing quarterbeats for pieces with
+        voltas. Requires the columns {'mc': int, 'quarterbeats_all_endings': fractions.Fraction} (ms3 >= 1.0.0).
     label_column: str, optional
         The column that is to be used as label string. Defaults to 'label'.
 
@@ -159,10 +167,16 @@ def transform_df(labels: pd.DataFrame,
     -------
         List of dictionaries where each represents one row of the input labels.
     """
-    offset_dict = measures.set_index("mc")["quarterbeats_all_endings"]
-    quarterbeats = labels['mc'].map(offset_dict)
-    quarterbeats = quarterbeats.astype('float') + (labels.mc_onset * 4.0)
-    transformed_df = pd.concat([quarterbeats.rename('quarterbeats'), labels.duration_qb.rename('duration'), labels[label_column].rename('label')], axis=1)
+
+    if measures is None or "quarterbeats_all_endings" not in measures.columns:
+        assert "quarterbeats" in labels.columns, f"Labels are lacking 'quarterbeats': {labels.columns}"
+        quarterbeats = labels["quarterbeats"]
+    else:
+        offset_dict = measures.set_index("mc")["quarterbeats_all_endings"]
+        quarterbeats = labels['mc'].map(offset_dict)
+        quarterbeats = quarterbeats.astype('float') + (labels.mc_onset * 4.0)
+        quarterbeats.rename('quarterbeats', inplace=True)
+    transformed_df = pd.concat([quarterbeats, labels.duration_qb.rename('duration'), labels[label_column].rename('label')], axis=1)
     return transformed_df.to_dict(orient='records')
     
 def make_dezrann_label(
@@ -232,7 +246,7 @@ def generate_dez(path_measures: str,
     """
     harmonies_df = pd.read_csv(
         path_labels, sep='\t',
-        usecols=['mc', 'mc_onset', 'duration_qb', 'label'], #'chord'
+        usecols=['mc', 'mc_onset', 'duration_qb', 'quarterbeats', 'label', 'chord', 'cadence', 'phraseend'],
         converters={'mc_onset': safe_frac}
     )
     try:
@@ -241,10 +255,13 @@ def generate_dez(path_measures: str,
             usecols=['mc', 'quarterbeats_all_endings'],
             converters={'quarterbeats_all_endings': safe_frac}
         )
-    except ValueError as e:
-        raise ValueError(f"{path_measures} could not be loaded as a measure map because of the following error:\n'{e}'")
-
[…]
