### 2023-02-24 16b7cc3
adds to transform_df() the algorithm that copies the preceding label onto beat 1 of each alternative ending (volta); also prepares the function for creating labels pertaining to the various layers
