+++ src/ms3/dezrann.py
@@ -133,15 +133,39 @@ class DcmlLabel(TypedDict):
     quarterbeats: float
     duration: float
     label: str
-    harmony: str
-    key: str
-    phrase: str
-    cadence: str
+
+
+def get_volta_groups(mc2volta: pd.Series) -> List[List[int]]:
+    """Takes a Series where the index has measure counts and values are NA for 'normal' measures and 1, 2... for
+    measures belonging to a first, second... ending. Returns for each group a list of MCs each of which pertains
+    to the first measure of an alternative ending. For example, two alternative two-bar endings in MC [15, 16][17, 18]
+    would figure as [15, 17] in the result list.
+    """
+    volta_groups = []
+    filled_volta_col = mc2volta.fillna(-1).astype(int)
+    volta_segmentation = (filled_volta_col != filled_volta_col.shift()).fillna(True).cumsum()
+    current_groups_first_mcs = []
+    for i, segment in filled_volta_col.groupby(volta_segmentation):
+        volta_number = segment.iloc[0]
+        if volta_number == -1:
+            # current group ends, if there is one
+            if i == 1:
+                continue
+            elif len(current_groups_first_mcs) == 0:
+                raise RuntimeError(f"Mistake in the algorithm when processing column {filled_volta_col.volta}")
+            else:
+                volta_groups.append(current_groups_first_mcs)
+                current_groups_first_mcs = []
+        else:
+            first_mc = segment.index[0]
+            current_groups_first_mcs.append(first_mc)
+    return volta_groups
 
 
 def transform_df(labels: pd.DataFrame,
-                measures: Optional[pd.DataFrame],
-                label_column: str = 'label') -> List[DcmlLabel]:
+                 measures: pd.DataFrame,
+                 label_column: str = 'label',
+                 ) -> List[DcmlLabel]:
     """
 
     Parameters
@@ -154,31 +178,68 @@ def transform_df(labels: pd.DataFrame,
                             'quarterbeats': fraction.Fraction,
                             'label': str,
                             'chord': str,
+                            'localkey': str,
                             'cadence': str,
                             'phraseend': str}
         and no missing values.
     measures:
         (optional) Dataframe as found in the 'measures' folder of a DCML corpus for computing quarterbeats for pieces with
         voltas. Requires the columns {'mc': int, 'quarterbeats_all_endings': fractions.Fraction} (ms3 >= 1.0.0).
-    label_column: str, optional
+    label_column: {'label', 'chord', 'cadence', 'phraseend'}
         The column that is to be used as label string. Defaults to 'label'.
 
     Returns
     -------
         List of dictionaries where each represents one row of the input labels.
     """
-
-    if measures is None or "quarterbeats_all_endings" not in measures.columns:
-        assert "quarterbeats" in labels.columns, f"Labels are lacking 'quarterbeats': {labels.columns}"
+    score_has_voltas = "quarterbeats_all_endings" in measures.columns
+    if not score_has_voltas:
+        assert "quarterbeats" in labels.columns, f"Labels are lacking 'quarterbeats' column: {labels.columns}"
         quarterbeats = labels["quarterbeats"]
+        last_mc = measures.iloc[-1]
+        end_of_score = last_mc.quarterbeats + last_mc.act_dur * 4.0
     else:
-        offset_dict = measures.set_index("mc")["quarterbeats_all_endings"]
+        # the column 'quarterbeats_all_endings' is present, meaning the piece has first and second endings and the
+        # quarterbeats, which normally leave out first endings, need to be recomputed
[…]
