### 2023-02-24 a965a1e
integrates calls to transform_df with the current logic within generate_dez()
