### 2023-02-24 544efa1
fix label_type naming typo
