### 2023-02-24 c12c4cf
remove fixed manual post-processing steps (NaN durations and handling start=0.)
