### 2023-02-24 a925564
add safe conversion of quarterbeats
