+++ docs/manual/index.md
@@ -1,6 +1,5 @@
 ---
 jupytext:
-  formats: ipynb,md:myst
   text_representation:
     extension: .md
     format_name: myst
@@ -14,84 +13,99 @@ kernelspec:
 
 # Manual
 
-![](ms3_architecture.png)
-
-This page is a detailed guide for using ms3 for different tasks. It
-supposes you are working in an interactive Python interpreter such as
-IPython, Jupyter, Google Colab, or just the console.
+This page is a detailed guide for using ms3 for different tasks. The code examples
+suppose you are working in an interactive Python interpreter such as
+IPython, Jupyter, Google Colab, or simply in the Python console. The manual itself
+is written as a [MyST markdown Notebook](https://myst-nb.readthedocs.io/) which can be run 
+in Jupyter if the [Jupytext](https://jupytext.readthedocs.io) extension is installed. 
 
 +++
 
 ## Good to know
 
-(corpss)=
-### Corpss structure 
+The principal *raison d'être* of `ms3` is for extracting different types of information, 
+"facets", from MuseScore files and to store them in the form of tabular `.tsv` files, 
+[Tab-separated values](https://en.wikipedia.org/wiki/Tab-separated_values), for further processing.
+Over the years, however, it has gained additional functionality related to corpus curation.
+Therefore, the package operates on a couple of principles that have co-evolved together
+with the [DCML corpus initiative](https://www.epfl.ch/labs/dcml/projects/corpus-project/).
 
-ms3 extracts various aspects (e.g. notes, chord labels, dynamics) from a
-single MuseScore file and stores them in separate TSV files. In order to
-jointly evaluate this information (e.g. to combine harmony labels with
-the corresponding notes in the score), ms3 associates files by their
-names. Therefore it is important to not rename files at a later point
-and it is recommended to stick to one organizational principle for all
-corpora in use. The two main principles are suffixes and subfolders.
+### Corpus structure 
 
-[DCML corpora](https://github.com/DCMLab/dcml_corpora) use the same
-subfolder structure: MuseScore files are stored in the `MS3` folder of a
-corpus and all other aspects are stored with identical filenames (but
-different extension) in sibling folders. This structure results
-naturally when using default parameters such as `ms3 extract -N -M -X`:
+`ms3` follows a one-folder-per-feature approach (rather than a one-folder-per-file approach).
 
-``` console
-corpus1
-├── harmonies
-├── measures
-├── MS3
-└── notes
-corpus2
-├── harmonies
-├── measures
-├── MS3
-└── notes
+#### The default
+
+The library's main command, `ms3 extract`, creates this folder-based structure by default.
+Consider this fresh `corpus_before` containing only the folder `MS3` with
+7 uncompressed MuseScore files in `.mscx` format:
+
+```{code-cell}
+:tags: [remove-input]
+
+#%cd manual
+!tree --fromfile corpus_before
 ```
 
-When loading corpora, ms3 looks for the standard folder names (and
-suffixes) to indentify individual corpora and assign them keys
-automatically (e.g. in the example above, \'corpus1\' and \'corpus2\').
-Using the default names will therefore facilitate the use of the library
-considerably.
+The default command that is used on all [DCML corpora](https://github.com/DCMLab/dcml_corpora),
+`ms3 extract -M -N -X -D -a` extracts three TSV files (measures, notes and "eXpanded" harmonies) 
+from each score and places them in a separate folder, plus an additional `metadata.tsv` file:
+
+```{code-cell}
+:tags: [remove-input]
+
+!tree --fromfile corpus_after
+```
+
+`ms3` operates on the fundamental principle that files that belong together need to
[…]
