### 2023-02-25 cd6546d
abuses old_tests for using the files in the manual; that includes updating K281-3.mscx with the v2.0 version that includes cadences, and replacing the contents of the file Did03M-Son_regina-1762-Sarti.mscx with something different including form labels
