+++ old_tests/truth/README.md
@@ -1,15 +1,0 @@
-
-
-# Overview
-|         file_name          |measures|labels|standard|annotators|reviewers|
-|----------------------------|-------:|-----:|--------|----------|---------|
-|05_symph_fant               |      32|     0|        |          |         |
-|76CASM34A33UM               |      94|   173|        |          |         |
-|BWV_0815                    |     234|     0|        |          |         |
-|D973deutscher01             |      18|     0|        |          |         |
-|Did03M-Son_regina-1762-Sarti|     107|   193|        |          |         |
-|K281-3                      |     162|   375|        |          |         |
-|stabat_03_coloured          |      26|    85|        |          |         |
-|repeats0                    |       4|     0|        |          |         |
-|repeats1                    |       5|     0|        |          |         |
-|repeats2                    |       3|     0|        |          |         |
