+++ old_tests/README.md
@@ -1,0 +1,15 @@
+
+
+# Overview
+|         file_name          |measures|labels|annotators |
+|----------------------------|-------:|-----:|-----------|
+|05_symph_fant               |      32|     0|           |
+|76CASM34A33UM               |      94|     0|           |
+|BWV_0815                    |     234|     0|           |
+|D973deutscher01             |      18|     0|           |
+|Did03M-Son_regina-1762-Sarti|      63|   373|Ana Llorens|
+|K281-3                      |     162|   384|           |
+|stabat_03_coloured          |      26|    83|           |
+
+
+*Overview table updated using [ms3](https://johentsch.github.io/ms3/) 1.2.4.*
