### 2023-02-25 5b3762c
further update of the manual notebook
