+++ docs/manual/index.md
@@ -17,7 +17,7 @@ This page is a detailed guide for using ms3 for different tasks. The code exampl
 suppose you are working in an interactive Python interpreter such as
 IPython, Jupyter, Google Colab, or simply in the Python console. The manual itself
 is written as a [MyST markdown Notebook](https://myst-nb.readthedocs.io/) which can be run 
-in Jupyter if the [Jupytext](https://jupytext.readthedocs.io) extension is installed. 
+in Jupyter if the [Jupytext](https://jupytext.readthedocs.io) extension is installed.
 
 +++
 
@@ -81,18 +81,19 @@ to add suffixes to the file names. For example, using
 ```
 
 Based on these two principles, default folder and suffix names, 
-`ms3` is able to recognize which facets the files represent and to relate them to each other.
+`ms3` is able to recognize which facet of which piece the files 
+represent and to relate them to each other.
 
 +++
 
-(key_and_id)=
+(keys_and_ids)=
 ### Keys and IDs
 
 ```{note} ms3 version 1.0.0 and successors widely replace the mechanisms related to 
 the parameter `key`. Newer versions, instead, use IDs such as
 `corpus_name` and `(corpus_name, piece_name)`. If you come across a method
 where the first parameter is called `key`, you are likely dealing with an
-older version, or with a [ms3.Score](ms3.score.Score) object.
+older version, or with a [ms3.Score](Score) object.
 ```
 
 IDs are tuples that are used to identify corpora, pieces and files:
@@ -106,6 +107,65 @@ IDs are tuples that are used to identify corpora, pieces and files:
   * `ms3.Piece[i] -> File`
   * `ms3.Corpus[(fname, i)] -> File`
   * `ms3.Parse[(corpus_name, fname, i)] -> File`
+  
+#### The importance of the `fname` ID
+  
+The piece IDs `fname` relate to the file names in this way: `fname[suffix].ext`.
+In order to correctly match files together that belong together, without doing
+complicated string matching, `ms3` relies on a list of fnames that it will
+expect to be present in a column called `fname` in a file called `metadata.tsv`
+(see below). Generally, this should be the first column in these files, 
+used as index.
+
+```{hint} 
+If you find yourself stuck with `ms3` producing no output,
+it is likely because no `metadata.tsv` is present. In this case, use the option
+`-a` to parse everything regardless, and `-D` to create a `metadata.tsv`.
+```
+
++++
+
+(metadata_tsv)=
+### The important role of metadata
+
+As mentioned above, `ms3` relies on the `fname` as ID of a piece and uses
+it to identify the various files belonging to it although they may come 
+with additional suffixes (e.g. `_reviewed`) and be scattered all over the
+corpus. Importantly, it uses and expects a file called `metadata.tsv` that
+lists all piece IDs of the corpus in a column called `fname`. Scores and
+other files whose names do not begin with any of strings in that column 
+are excluded. Files, on the other hand, that begin with any of the strings
[…]
