### 2023-02-25 2b5c033
allows for extracting only metadata and nothing else
