+++ docs/manual/index.md
@@ -174,12 +174,12 @@ the variable `c`, typing `c.all` will activate the view that shows everything.
 
 ms3 recognizes and disambiguates different types of labels, depending on
 how they are encoded in MuseScore, see
-`harmony_layer <harmony_layer>`{.interpreted-text role="ref"}.
+{ref}`harmony_layer <harmony_layer>`.
 
 Independent of the type, ms3 will also try to infer whether a label
 conforms to the DCML syntax and/or other regular expressions registered
 via `ms3.Score.new_type`{.interpreted-text role="meth"}. The column
-`regex_match <regex_match>`{.interpreted-text role="ref"} contains for
+{ref}`regex_match <regex_match>` contains for
 each label the name of the first regEx that matched. information will
 appear with a subtype, e.g. `0 (dcml)`.
 
@@ -245,7 +245,7 @@ memory.
 ### Stacks-of-fifths intervals 
 
 In order to express note names (tonal pitch classes,
-`tpc <tpc>`{.interpreted-text role="ref"}), and scale degrees, ms3 uses
+{ref}`tpc <tpc>`), and scale degrees, ms3 uses
 stacks of fifths (the only way to express these as a single integer).
 For note names, `0` corresponds to C, for scale degrees to the local
 tonic.
@@ -285,7 +285,7 @@ ms3 checks for incorrect MNs and warns you if the score needs
 correction. It will also ask you to make all voltas the same length. If
 this is not possible for editorial reasons (although often the length of
 the second volta is arbitrary), ignore the warning and check in the
-`measures <measures>`{.interpreted-text role="ref"} table if the MN are
+{ref}`measures <measures>` table if the MN are
 correct for your purposes.
 
 +++
@@ -304,11 +304,11 @@ tables are:
 All score information, except the metadata, is contained in the
 following two tables:
 
--   `measures <measures>`{.interpreted-text role="ref"}
--   `notes <notes>`{.interpreted-text role="ref"}
--   `rests <rests>`{.interpreted-text role="ref"}
--   `notes_and_rests <notes_and_rests>`{.interpreted-text role="ref"}
--   `chords <chords>`{.interpreted-text role="ref"}: **Not to be
+-   {ref}`measures <measures>`
+-   {ref}`notes <notes>`
+-   {ref}`rests <rests>`
+-   {ref}`notes_and_rests <notes_and_rests>`
+-   {ref}`chords <chords>`: **Not to be
     confounded with labels or chord annotations**, a chord is a
     notational unit in which all included notes are part of the same
     notational layer and have the same onset and duration. Every chord
@@ -316,10 +316,10 @@ following two tables:
     used to convey score information that is not attached to a
     particular note, such as lyrics, staff text, dynamics and other
     markup.
--   `labels <labels>`{.interpreted-text role="ref"}
--   `expanded <expanded>`{.interpreted-text role="ref"}
--   `cadences <cadences>`{.interpreted-text role="ref"}
--   `events <events>`{.interpreted-text role="ref"}
+-   {ref}`labels <labels>`
+-   {ref}`expanded <expanded>`
+-   {ref}`cadences <cadences>`
+-   {ref}`events <events>`
 
 For each of the available tables you will see an example and you can
 click on the columns to learn about their meanings.
@@ -327,7 +327,7 @@ click on the columns to learn about their meanings.
 ### Measures
 
 DataFrame representing the measures in the MuseScore file (which can be
[…]
