### 2023-02-25 8885656
fixes cross-references in the new manual helped by manual_old.md created with rst2myst
