### 2023-02-25 8bab916
turns rst "bignums" into definition lists (myst extension "deflist") and fixes heading anchors
