+++ docs/old_src/manual_old.md
@@ -1,2225 +1,0 @@
----
-substitutions:
-  Ottava:15mb: '{ref}`Ottava:15mb <ottava>`'
-  Ottava:8va: '{ref}`Ottava:8va <ottava>`'
-  Ottava:8vb: '{ref}`Ottava:8vb <ottava>`'
-  TextLine: '{ref}`TextLine <textline>`'
-  act_dur: '{ref}`act_dur <act_dur>`'
-  added_tones: '{ref}`added_tones <chord_tones>`'
-  alt_label: '{ref}`alt_label <alt_label>`'
-  articulation: '{ref}`articulation <articulation>`'
-  barline: '{ref}`barline <barline>`'
-  bass_note: '{ref}`bass_note <bass_note>`'
-  breaks: '{ref}`breaks <breaks>`'
-  cadence: '{ref}`cadence <cadence>`'
-  changes: '{ref}`changes <changes>`'
-  chord: '{ref}`chord <chord>`'
-  chord_id: '{ref}`chord_id <chord_id>`'
-  chord_tones: '{ref}`chord_tones <chord_tones>`'
-  chord_type: '{ref}`chord_type <chord_type>`'
-  crescendo_hairpin: '{ref}`crescendo_hairpin <hairpins>`'
-  crescendo_line: '{ref}`crescendo_line <cresc_lines>`'
-  decrescendo_hairpin: '{ref}`decrescendo_hairpin <hairpins>`'
-  diminuendo_line: '{ref}`diminuendo_line <cresc_lines>`'
-  dont_count: '{ref}`dont_count <dont_count>`'
-  duration: '{ref}`duration <duration>`'
-  duration_qb: '{ref}`duration_qb <duration_qb>`'
-  dynamics: '{ref}`dynamics <dynamics>`'
-  figbass: '{ref}`figbass <figbass>`'
-  form: '{ref}`form <form>`'
-  globalkey: '{ref}`globalkey <globalkey>`'
-  globalkey_is_minor: '{ref}`globalkey_is_minor <globalkey_is_minor>`'
-  gracenote: '{ref}`gracenote <gracenote>`'
-  harmony_layer: '{ref}`harmony_layer <harmony_layer>`'
-  keysig: '{ref}`keysig <keysig>`'
-  label: '{ref}`label <label>`'
-  label_type: '{ref}`label_type <label_type>`'
-  localkey: '{ref}`localkey <localkey>`'
-  localkey_is_minor: '{ref}`localkey_is_minor <localkey_is_minor>`'
-  lyrics:1: '{ref}`lyrics:1 <lyrics_1>`'
-  mc: '{ref}`mc <mc>`'
-  mc_offset: '{ref}`mc_offset <mc_offset>`'
-  mc_onset: '{ref}`mc_onset <mc_onset>`'
-  midi: '{ref}`midi <midi>`'
-  mn: '{ref}`mn <mn>`'
-  mn_onset: '{ref}`mn_onset <mn_onset>`'
-  next: '{ref}`next <next>`'
-  nominal_duration: '{ref}`nominal_duration <nominal_duration>`'
-  numbering_offset: '{ref}`numbering_offset <numbering_offset>`'
-  numeral: '{ref}`numeral <numeral>`'
-  offset_x: '{ref}`offset_x <offset>`'
-  offset_y: '{ref}`offset_y <offset>`'
-  pedal: '{ref}`pedal <pedal>`'
-  phraseend: '{ref}`phraseend <phraseend>`'
-  qpm: '{ref}`qpm <qpm>`'
-  quarterbeats: '{ref}`quarterbeats <quarterbeats>`'
-  quarterbeats_all_endings: '{ref}`quarterbeats_all_endings <quarterbeats_all_endings>`'
[…]
