+++ docs/manual/index.md
@@ -42,7 +42,7 @@ The library's main command, `ms3 extract`, creates this folder-based structure b
 Consider this fresh `corpus_before` containing only the folder `MS3` with
 7 uncompressed MuseScore files in `.mscx` format:
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 #%cd manual
@@ -53,7 +53,7 @@ The default command that is used on all [DCML corpora](https://github.com/DCMLab
 `ms3 extract -M -N -X -D -a` extracts three TSV files (measures, notes and "eXpanded" harmonies) 
 from each score and places them in a separate folder, plus an additional `metadata.tsv` file:
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 !tree --fromfile corpus_after
@@ -76,7 +76,7 @@ Users who need their outputs in the same directory, say `out`, can specify the f
 to add suffixes to the file names. For example, using 
 `ms3 extract -M out -N out -X out -s` you would get:
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 !tree --fromfile out
@@ -176,12 +176,12 @@ the variable `c`, typing `c.all` will activate the view that shows everything.
 
 ms3 recognizes and disambiguates different types of labels, depending on
 how they are encoded in MuseScore, see
-{ref}`harmony-layer`.
+{ref}`harmony_layer`.
 
 Independent of the type, ms3 will also try to infer whether a label
 conforms to the DCML syntax and/or other regular expressions registered
 via {py:meth}`~.Score.new_type`. The column
-{ref}`regex_match <regex_match>` contains for
+{ref}`regex_match` contains for
 each label the name of the first regEx that matched. information will
 appear with a subtype, e.g. `0 (dcml)`.
 
@@ -268,6 +268,8 @@ tonic.
 
 +++
 
+(voltas)=
+
 ### Voltas
 
 \"Prima/Seconda volta\" is the Italian designation for \"First/Second
@@ -324,6 +326,8 @@ following two tables:
 For each of the available tables you will see an example and you can
 click on the columns to learn about their meanings.
 
+(measures)=
+
 ### Measures
 
 DataFrame representing the measures in the MuseScore file (which can be
@@ -337,7 +341,7 @@ together with their respective features. Required for unfolding repeats.
 >>> p.get_facet('measures')      # from a Parse object
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 import ms3
@@ -345,6 +349,8 @@ c = ms3.Corpus("../../old_tests", level='c')
 c.parse_scores()
 c.measures()
 ```
+
+(notes)=
 
 ### Notes
 
@@ -357,11 +363,13 @@ DataFrame representing the notes in the MuseScore file.
 >>> p.get_facet('notes')      # from a Parse object
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 c.notes()
 ```
+
+(rests)=
 
 ### Rests
 
@@ -374,13 +382,14 @@ DataFrame representing the rests in the MuseScore file.
 >>> p.get_facet('rests')      # from a Parse object
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 :tags: [remove-input]
 
 c.rests()
 ```
 
 (notes_and_rests)=
+
 ### Notes and Rests 
 
 DataFrame combining {ref}`notes` and
@@ -393,11 +402,13 @@ DataFrame combining {ref}`notes` and
 >>> p.get_facet('notes_and_rests')    # from a Parse object
 ```
 
[…]

[…]
