+++ docs/build/html/_sources/build/jupyter_execute/build/jupyter_execute/build/jupyter_execute/build/jupyter_execute/quick.ipynb.txt
@@ -1,1398 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "4f4a63c6",
-   "metadata": {},
-   "source": [
-    "# Quick `ms3` reference\n",
-    "\n",
-    "## To run this notebook\n",
-    "\n",
-    "* install ms3 (`pip install ms3`)\n",
-    "* set the `DATA_PATH` to where you want the folder `dcml_corpora` to be created that contains the data"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "id": "6a20160e",
-   "metadata": {
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "DATA_PATH = '~'"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2826d091",
-   "metadata": {},
-   "source": [
-    "## Setup"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "id": "9d608b25",
-   "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "dcml_corpora @ commit 9dcde40cba36d31b900ff12852cc557b8cca8221\n"
-     ]
-    }
-   ],
-   "source": [
-    "import os\n",
-    "import ms3\n",
-    "from git import Repo\n",
-    "\n",
-    "corpora_path = os.path.join(os.path.expanduser(DATA_PATH), 'dcml_corpora')\n",
-    "if os.path.isdir(corpora_path):\n",
-    "    repo = Repo(corpora_path)\n",
-    "else:\n",
-    "    repo = Repo.clone_from(url='https://github.com/DCMLab/dcml_corpora.git', \n",
-    "                to_path=corpora_path, \n",
-    "                multi_options=['--recurse-submodules', '--shallow-submodules'])\n",
-    "print(f\"dcml_corpora @ commit {repo.commit().hexsha}\")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "afdfa31b",
-   "metadata": {},
-   "source": [
-    "## Parsing multiple scores at once\n",
-    "\n",
-    "### The Corpus object\n",
-    "\n",
-    "Scores often come grouped into a corpus, so when we want to parse multiple scores, we create a [Corpus](Corpus) object and pass it the directory containing the scores. `ms3` will scan the directory and discover all scores and TSV files that can be potentially parsed:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "id": "f6fe838c",
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/plain": [
-       "[\u001b[1mdefault\u001b[0;0m|all]\n",
-       "Corpus 'tchaikovsky_seasons'\n",
-       "----------------------------\n",
-       "Location: /home/hentsche/dcml_corpora/tchaikovsky_seasons\n",
-       "View: This view is called 'default'. It \n",
-       "\t- excludes fnames that are not contained in the metadata,\n",
-       "\t- filters out file extensions requiring conversion (such as .xml), and\n",
-       "\t- excludes review files and folders.\n",
-       "\n",
-       "All 12 pieces are listed in 'metadata.tsv':\n",
-       "\n",
-       "          scores measures    notes expanded   events   chords\n",
-       "        detected detected detected detected detected detected\n",
-       "op37a01        1        1        1        1        1        1\n",
-       "op37a02        1        1        1        1        1        1\n",
-       "op37a03        1        1        1        1        1        1\n",
-       "op37a04        1        1        1        1        1        1\n",
-       "op37a05        1        1        1        1        1        1\n",
[…]
