+++ docs/build/html/_sources/old_src/quick_old.rst.txt
@@ -1,0 +1,457 @@
+===============
+Quick Reference
+===============
+
+.. figure:: ms3_architecture.png
+    :alt: The archicture of ms3
+
+There are two ways of interacting with ms3, namely using the commands that it installs to your system (such as
+``ms3 extract``), or using its main objects, :py:class:`~ms3.score.Score` and :py:class:`~ms3.parse.Parse`, for example
+in an interactive Python interpreter such as IPython, Jupyter, Google Colab, or just the console.
+
+
+
+Parsing a single score: Working with the Score object
+=====================================================
+
+The example score is called ``stabat.mscx`` and can be downloaded from
+`here <https://raw.githubusercontent.com/johentsch/ms3/master/docs/stabat.mscx>`__.
+
+Creating a Score object
+-----------------------
+
+Simply pass the path to the MuseScore 3 file that you want to parse:
+
+.. code-block:: python
+
+    >>> from ms3 import Score
+    >>> s = Score('~/ms3/docs/stabat.mscx')
+    >>> s
+        Uncompressed MuseScore file
+        ---------------------------
+
+        ~/ms3/docs/stabat.mscx
+
+        Attached annotations
+        --------------------
+
+        48 labels:
+        staff  voice  label_type  color_name
+        3      2      0 (dcml)    default       48
+
+As you see, upon inspecting the new object, the 48 detected labels are summarized. For the explanation of the various
+label types, :ref:`see here <label_types>`.
+
+A score object can also be created from a file that first needs to be converted (convertible format are, to date,
+CAP, CAPX, MID, MIDI, MUSICXML, MXL), as long as you have MuseScore 3 installed on your computer:
+
+.. code-block:: python
+
+    >>> s = Score('stabat.mxl')
+    ValueError: To open a mxl file, use 'ms3 convert' command or pass parameter 'ms' to Score to temporally convert.
+
+Ooops, we also need to tell ms3 where it can find the MuseScore executable. If you are lucky, you can let ms3 go check
+the default installation path for your system by trying ``Score('stabat.xml', ms='auto')``. Otherwise pass the path
+to the executable. Linux users might want to use the `AppImage <https://musescore.org/en/download#AppImage-(recommended)>`__
+which can be stored anywhere on the system.
+
+
+Storing the labels
+------------------
+
+The annotations contained in a score are stored in an :py:class:`~ms3.annotations.Annotations` object and can be accessed
+and stored as a tab-separated file (TSV) like this:
+
+.. code-block:: python
+
+    >>> s.annotations
+    48 labels:
+    staff  voice  label_type  color_name
+    3      2      0 (dcml)    default       48
+
+    >>> s.annotations.store_tsv('stabat_chords.tsv')
+    True
+
+.. _detaching:
+
+Removing annotation labels
+--------------------------
+
+The annotations will be stored with a keyword that you choose. It needs to be different from ``'annotations'``.
+
+.. caution::
+
+  The parameter ``key`` will be deprecated from version 0.6.0 onwards. See :ref:`keys_and_ids`.
+
+.. code-block:: python
+
+    >>> s.detach_labels(key='chords')
+    >>> s
+    Uncompressed MuseScore file (CHANGED!!!)
+    ----------------------------!!!!!!!!!!!!
+
+    ~/ms3/docs/stabat.mscx
+
+    No annotations attached.
+
+    Detached annotations
+    --------------------
+
+    chords -> 48 labels:
[…]
