+++ docs/build/html/_sources/README.rst.txt
@@ -1,6 +1,0 @@
-Building the docs
-=================
-
-#. In the folder ``docs``, run ``pip install -r requirements.txt``.
-#. Then you can build the current docs using ``make html``.
-#. See the output by opening ``docs/build/html/index.html``.
