+++ src/ms3/piece.py
@@ -83,13 +83,13 @@ class Piece(LoggedClass):
         """ Checks if parsed TSV files have been detected for all selected facets under the active or indicated view.
 
         Args:
-            view_name: Name of the view to check.
-            selected_facets: If passed, needs to be a subset of the facets selected by the view, otherwise the
-                result will be False. If no ``selected_facets`` are passed, check for those selected by the
-                active or indicated view.
+          view_name: Name of the view to check.
+          selected_facets: If passed, needs to be a subset of the facets selected by the view, otherwise the
+              result will be False. If no ``selected_facets`` are passed, check for those selected by the
+              active or indicated view.
 
         Returns:
-            True if for each selected facet at least one file has been registered.
+          True if for each selected facet at least one file has been registered.
         """
         view = self.get_view(view_name)
         view_facets = view.selected_facets
@@ -149,8 +149,8 @@ class Piece(LoggedClass):
         """
 
         Args:
-            choose:
-            as_dict: Set to True to change the return type from :obj:`pandas.Series` to :obj:`dict`.
+          choose:
+          as_dict: Set to True to change the return type from :obj:`pandas.Series` to :obj:`dict`.
 
         Returns:
 
@@ -481,20 +481,20 @@ class Piece(LoggedClass):
         attached to the Score in red, just like any deleted label.
 
         Args:
-            key: Key of the detached labels you want to compare to the ones in the score.
-            new_color, old_color:
-                The colors by which new and old labels are differentiated. Identical labels remain unchanged. Colors can be
-                CSS colors or MuseScore colors (see :py:attr:`utils.MS3_COLORS`).
-            detached_is_newer:
-                Pass True if the detached labels are to be added with ``new_color`` whereas the attached changed labels
-                will turn ``old_color``, as opposed to the default.
-            add_to_rna:
-                By default, new labels are attached to the Roman Numeral layer.
-                Pass False to attach them to the chord layer instead.
+          key: Key of the detached labels you want to compare to the ones in the score.
+          new_color, old_color:
+              The colors by which new and old labels are differentiated. Identical labels remain unchanged. Colors can be
+              CSS colors or MuseScore colors (see :py:attr:`utils.MS3_COLORS`).
+          detached_is_newer:
+              Pass True if the detached labels are to be added with ``new_color`` whereas the attached changed labels
+              will turn ``old_color``, as opposed to the default.
+          add_to_rna:
+              By default, new labels are attached to the Roman Numeral layer.
+              Pass False to attach them to the chord layer instead.
 
         Returns:
-            Number of scores in which labels have changed.
-            Number of scores in which no label has chnged.
+          Number of scores in which labels have changed.
+          Number of scores in which no label has chnged.
         """
         changed, unchanged = 0, 0
         for file, score in self.get_parsed_scores(view_name=view_name):
@@ -528,13 +528,13 @@ class Piece(LoggedClass):
         """ Count how many files per facet have been detected.
 
         Args:
-            include_empty:
-                By default, facets without files are not included in the dict. Pass True to include zero counts.
-            view_name:
-            prefix: Pass True if you want the facets prefixed with 'detected_'.
+          include_empty:
+              By default, facets without files are not included in the dict. Pass True to include zero counts.
+          view_name:
+          prefix: Pass True if you want the facets prefixed with 'detected_'.
 
         Returns:
-            {facet -> count of detected files}
+          {facet -> count of detected files}
         """
         result = {}
         for facet, files in self.iter_facet2files(view_name=view_name, include_empty=include_empty):
@@ -627,16 +627,16 @@ class Piece(LoggedClass):
 
 
         Args:
-            facets:
-            view_name:
-            force:
[…]
