### 2023-02-27 6fe059a
enables `ms3 transform -D` without parsing any other TSV
