### 2023-02-27 #5195 @editorialbot
[REVIEW]: ms3: A parser for MuseScore files, serving as data factory for annotated music corpora
**Submitting author:** @johentsch (<a href="http://orcid.org/0000-0002-1986-9545">Johannes Hentschel</a>)
**Repository:** https://github.com/johentsch/ms3
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.10
