### 2023-03-03 #5195 @allorens
@johentsch a few issues, mainly re. documentation:
1. Whereas in the paper the target audience is made clear, it is not in the documentation itself.
2. No automated tests to verify functionality are described.
3. Couldn't find guidelines to contribute, report or seek support.
4. In the paper, no comparison with other packages is included. Even if ms3's functionalities are different from those of music21, it should be stated.
