### 2023-03-03 8df1305
makes utils.no_collections_no_booleans() more robust against empty columns and suppresses pandas' annoying FutureWarnings
