### 2023-03-10 3dca29c
converts CONST argument '' to None for ms3 compare -c
