+++ src/ms3/utils.py
@@ -3604,7 +3604,7 @@ def write_metadata(metadata_df: pd.DataFrame,
 @function_logger
 def enforce_fname_index_for_metadata(metadata_df: pd.DataFrame, append=False) -> pd.DataFrame:
     """Returns a copy of the DataFrame that has an index level called 'fname'."""
-    possible_column_names = ('fname', 'fnames', 'name', 'names')
+    possible_column_names = ('fname', 'fnames', 'filename', 'name', 'names',)
     if any(name in metadata_df.index.names for name in possible_column_names):
         return metadata_df
     try:
