### 2023-03-17 b22791c
makes the stub of the new Instrumentation class available via `Score.mscx.parsed.instrumentation`
