### 2023-03-17 f65bd0b
makes the stub of the new Instrumentation class available via `Score.mscx.parsed.instrumentation`
