### 2023-03-17 6a4dc74
makes the stub of the new Instrumentation class available via `Score.mscx.parsed.instrumentation`
