### 2023-03-17 f724a97
Applies appropriate variables names
