### 2023-03-17 88cbb5f
Uses field_to_change as an argument not as a field  and fixes bugs
