### 2023-03-17 44db290
Uses field_to_change as an argument not as a field  and fixes bugs
