### 2023-03-17 f0de53f
Fills staff2part in the method map_staff2part only once
