### 2023-03-17 e56c9bf
Fills staff2part in the method map_staff2part only once
