### 2023-03-17 87a2a7e
Fills staff2part in the method map_staff2part only once
