+++ src/ms3/bs4_parser.py
@@ -2014,15 +2014,21 @@ class Instrumentation(LoggedClass):
         self.soup = soup
         self.fields_names = ['instrument', 'longName', 'shortName', 'trackName', 'instrumentId']
         self.parts = {f"part_{i}": part for i, part in enumerate(self.soup.find_all('Part'), 1)}
-        self.staff2part = {}
+
+    def map_staff2part(self) -> dict[list, str]:
+        """Returns the dict in the format {[2, 3]: 'part_1'} for staves 2 and 3 of part 1"""
+        staff2part = {}
+        for key_part, part in self.parts.items():
+            staves = [int(staff['id']) for staff in part.find_all('Staff')]
+            staff2part.update(dict.fromkeys(staves, key_part))
+        return staff2part
 
     @property
-    def text_tags(self) -> dict[str, bs4.Tag]:
+    def text_tags(self) -> dict[int, dict[str, bs4.Tag]]:
         tag_dict = {}
         for key_part, part in self.parts.items():
             instrument_info = part.Instrument
             staves = [int(staff['id']) for staff in part.find_all('Staff')]
-            self.staff2part.update(dict.fromkeys(staves, key_part))
             cur_dict = {"instrumentId": instrument_info["id"]}
             for name in self.fields_names:
                 if name != "instrumentId":
