### 2023-03-17 a4cfcb2
Fils text_tags only once (not a property but the field in init)
