### 2023-03-17 6652f3c
Fils text_tags only once (not a property but the field in init)
