### 2023-03-17 47226ae
Fils text_tags only once (not a property but the field in init)
