### 2023-03-17 a80d026
adds docs on how to contribute and get help
