### 2023-03-18 #5195 @johentsch
Thank you, @allorens . The referenced PR addresses these three points:

> * Whereas in the paper the target audience is made clear, it is not in the documentation itself.
> * No automated tests to verify functionality are described.
> * Couldn't find guidelines to contribute, report or seek support.

The fourth one will be following shortly.
