### 2023-03-18 johentsch/ms3#53 @allorens
Documentation - target audience
Now the _raison d'être_ of `ms3` is much clearly stated in the documentation, but not so the target audience: Are they musicologists, computer scientists, both?
See [review comment](https://github.com/openjournals/joss-reviews/issues/5195#issuecomment-1453345717)
