### 2023-03-18 johentsch/ms3#54 @allorens
Documentation - automated tests
No automated tests to verify functionality are described. See [review comment](https://github.com/openjournals/joss-reviews/issues/5195#issuecomment-1453345717).
