### 2023-03-18 johentsch/ms3#55 @allorens
Documentation - contributions, report and support
No guidelines to contribute, report or seek support are to be found. See [review comment](https://github.com/openjournals/joss-reviews/issues/5195#issuecomment-1453345717).
