### 2023-03-18 johentsch/ms3#56 @allorens
Documentation - comparison with other libraries
In the paper, no comparison with other packages is included. Even if `ms3`'s functionalities are different from those of `music21`, it should be stated.

See [review comment](https://github.com/openjournals/joss-reviews/issues/5195#issuecomment-1453345717).
