+++ paper/paper.bib
@@ -1,3 +1,27 @@
+@inproceedings{Cancino-Chacon2022_PartituraPythonPackage,
+  title = {Partitura: {{A}} Python Package for Symbolic Music Processing},
+  booktitle = {Proceedings of the Music Encoding Conference ({{MEC2022}})},
+  author = {{Cancino-Chac{\'o}n}, Carlos Eduardo and Peter, Silvan David and Karystinaios, Emmanouil and Foscarin, Francesco and Grachten, Maarten and Widmer, Gerhard},
+  year = {2022},
+  address = {{Halifax, Canada}}
+}
+
+@inproceedings{Cuthbert2010_Music21ToolkitComputerAided,
+  title = {Music21: {{A Toolkit}} for {{Computer-Aided Musicology}} and {{Symbolic Music Data}}},
+  shorttitle = {{{ISMIR}} 2010},
+  booktitle = {11th {{International Society}} for {{Music Information Retrieval Conference}}},
+  author = {Cuthbert, Michael Scott},
+  editor = {{International Conference on Music Information Retrieval} and Downie, J. Stephen and Veltkamp, Remco C},
+  year = {2010},
+  pages = {637--642},
+  publisher = {{International Society for Music Information Retrieval}},
+  address = {{Utrecht, Netherlands}},
+  isbn = {978-90-393-5381-3},
+  langid = {english},
+  annotation = {OCLC: 810276091}
+}
+
+
 @article{Hentschel2021_AnnotatedMozartSonatas,
   title = {The {{Annotated Mozart Sonatas}}: {{Score}}, {{Harmony}}, and {{Cadence}}},
   author = {Hentschel, Johannes and Neuwirth, Markus and Rohrmeier, Martin},
@@ -7,8 +31,7 @@
   number = {1},
   pages = {67--80},
   doi = {10.5334/tismir.63},
-  abstract = {This article describes a new expert-labelled dataset featuring harmonic, phrase, and cadence analyses of all piano sonatas by W.A. Mozart. The dataset draws on the DCML standard for harmonic annotation and is being published adopting the FAIR principles of Open Science. The annotations have been verified using a data triangulation procedure which is presented as an alternative approach to handling annotator subjectivity. This procedure is suited for ensuring consistency, within the dataset and beyond, despite the high level of analytical detail afforded by the employed harmonic annotation syntax. The harmony labels also encode contextual information and are therefore suited for investigating music theoretical questions related to tonal harmony and the harmonic makeup of cadences in the classical style. Apart from providing basic statistical analyses characterizing the dataset, its music theoretical potential is illustrated by two preliminary experiments, one on the terminal harmonies of cadences and the other on the relation between performance durations and harmonic density. Furthermore, particular features can be selected to produce more coarse-grained training data, for example for chord detection algorithms that require less analytical detail. Facilitating the dataset's reusability, it comes with a Python script that allows researchers to easily access various representations of the data tailored to their particular needs.},
-  file = {/home/hentsche/Zotero/storage/WF5J7KZX/Hentschel et al. - 2021 - The Annotated Mozart Sonatas Score, Harmony, and .pdf}
[…]
