+++ paper/paper.bib
@@ -9,9 +9,8 @@
 @inproceedings{Cuthbert2010_Music21ToolkitComputerAided,
   title = {Music21: {{A Toolkit}} for {{Computer-Aided Musicology}} and {{Symbolic Music Data}}},
   shorttitle = {{{ISMIR}} 2010},
-  booktitle = {11th {{International Society}} for {{Music Information Retrieval Conference}}},
+  booktitle = {Proceedings of the 11th {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
   author = {Cuthbert, Michael Scott},
-  editor = {{International Conference on Music Information Retrieval} and Downie, J. Stephen and Veltkamp, Remco C},
   year = {2010},
   pages = {637--642},
   publisher = {{International Society for Music Information Retrieval}},
@@ -153,7 +152,7 @@
 @inproceedings{Pugin2014_VerovioLibraryEngraving,
   title = {Verovio: {{A Library}} for {{Engraving MEI Music Notation}} into {{SVG}}},
   shorttitle = {Verovio},
-  booktitle = {Pugin, {{Laurent}}; {{Zitellini}}, {{Rodolfo}}; {{Roland}}, {{Perry}} (2014). {{Verovio}}: {{A Library}} for {{Engraving MEI Music Notation}} into {{SVG}}. {{In}}: 15th {{International Society}} for {{Music Information Retrieval Conference}} ({{ISMIR}} 2014). {{Taipei}}. 27-31 {{Oktober}} 2014.},
+  booktitle = {Proceedings of the 15th {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
   author = {Pugin, Laurent and Zitellini, Rodolfo and Roland, Perry},
   year = {2014},
   pages = {107--112},
