### 2023-03-18 3e4bb9b
adds a paragraph on the target audience
