### 2023-03-22 8981754
adds method count_pieces() to Corpus and Parse
