### 2023-03-22 db5efeb
small bugfix for the setting View.fnames_with_incomplete_facets=False
