+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+Version 1.2.5
+=============
+
+* :obj:`~ms3.Corpus` and :obj:`~ms3.Piece` come with the new method ``count_pieces()``
+* ``ms3 transform -D`` to concatenate only metadata works
+* ``View.fnames_with_incomplete_facets = False`` enforces selected facets if some have been excluded
+
 Version 1.2.4
 =============
 
@@ -17,7 +24,7 @@ Version 1.2.3
 
 * Piece.get_facet() gets parameter 'force' which defaults to False (analogous to the other methods),
   in order to avoid unsolicited score parsing.
-* improves `ms3 transform`:
+* improves ``ms3 transform``:
   * parse only facets to be concatenated (rather than all TSV files)
   * do not accidentally output metadata if not requested
 * prevents including 'volta_mcs' in metadata of pieces that don't have voltas
