### 2023-04-03 fa7ee1f
bugfix concerning View.update_config(folder_paths)
