### 2023-04-19 d8f661a
makes ms3 empty work under the new CLI
