### 2023-04-19 5d2f5b8
add structure and docstring for measure_map converter module
