### 2023-04-19 850e7a2
add conversion function
