### 2023-04-19 7cec5fa
add save function and test case
