### 2023-04-25 1e0da13
adds classmethod File.from_corpus_path()
