### 2023-04-25 efb3b0e
adds File.replace_extension()
