### 2023-04-25 23f7325
enables casting utils.scale_degree2name() to Series, array, list and tuple
