### 2023-04-25 c7d1a1c
more precise type hint for Corpus.get_files()
