+++ src/ms3/parse.py
@@ -1,5 +1,5 @@
 import re
-from typing import Literal, Collection, Generator, Tuple, Union, Dict, Optional, List, Iterator
+from typing import Literal, Collection, Generator, Tuple, Union, Dict, Optional, List, Iterator, overload
 
 import sys, os
 from collections import Counter, defaultdict
@@ -702,7 +702,26 @@ class Parse(LoggedClass):
                                            include_empty=include_empty,
                                            concatenate=concatenate,
                                            )
-
+    @overload
+    def get_files(self, facets: FacetArguments = None,
+                  view_name: Optional[str] = None,
+                  parsed: bool = True,
+                  unparsed: bool = True,
+                  choose: Literal['all', 'auto', 'ask'] = 'all',
+                  flat: bool = False,
+                  include_empty=False,
+                  ) -> Dict[CorpusFnameTuple, FileDict]:
+        ...
+    @overload
+    def get_files(self, facets: FacetArguments = None,
+                  view_name: Optional[str] = None,
+                  parsed: bool = True,
+                  unparsed: bool = True,
+                  choose: Literal['all', 'auto', 'ask'] = 'all',
+                  flat: bool = True,
+                  include_empty=False,
+                  ) -> Dict[CorpusFnameTuple, FileList]:
+        ...
     def get_files(self, facets: FacetArguments = None,
                   view_name: Optional[str] = None,
                   parsed: bool = True,
