+++ src/ms3/operations.py
@@ -103,8 +103,7 @@ def check(parse_obj: Parse,
           ignore_labels: bool = False,
           ignore_scores: bool = False,
           assertion: bool = False,
-          parallel: bool = True,
-          warnings_file: Optional[str] = None) -> bool:
+          parallel: bool = True) -> bool:
     assert ignore_labels + ignore_scores < 2, "Activate either ignore_labels or ignore_scores, not both."
     all_warnings = []
     check_logger = get_logger("ms3.check", level=parse_obj.logger.getEffectiveLevel())
@@ -127,10 +126,6 @@ def check(parse_obj: Parse,
         elif len(warnings) > 0:
             all_warnings.extend(warnings)
             check_logger.warning("Warnings detected while checking DCML labels (see above).")
-    if warnings_file is not None and len(all_warnings) > 0:
-        with open(warnings_file, 'a', encoding='utf-8') as f:
-            f.writelines(w + '\n' for w in all_warnings)
-        parse_obj.logger.info(f"Added captured warnings to {warnings_file}")
     if assertion:
         assert len(all_warnings) == 0, "Encountered warnings, check failed."
     if len(all_warnings) == 0:
