### 2023-04-26 b5bcbc9
moves make_coloring_reports_and_warnings from cli to operations
