+++ src/ms3/operations.py
@@ -1,8 +1,9 @@
+import os
 from typing import Literal, Optional, Tuple, Dict, List, Union
 
 from ms3 import Parse, Corpus
 from ms3._typing import AnnotationsFacet
-from ms3.utils import capture_parse_logs, LATEST_MUSESCORE_VERSION, pretty_dict, check_argument_against_literal_type
+from ms3.utils import capture_parse_logs, LATEST_MUSESCORE_VERSION, pretty_dict, check_argument_against_literal_type, compute_path_from_file, write_tsv, fifths2sd, scale_degree2name
 from ms3.logger import get_logger, temporarily_suppress_warnings, function_logger
 
 def insert_labels_into_score(ms3_object: Union[Parse, Corpus],
@@ -240,3 +241,50 @@ def update(parse_obj: Parse,
         corpus.logger.info(f"Labels updated in {len(scores_with_updated_labels)}")
         file_paths = corpus.store_parsed_scores(overwrite=overwrite, only_changed=True)
         return file_paths
+
+
+def make_coloring_reports_and_warnings(parse_obj: Parse,
+                                       out_dir: Optional[str] = None,
+                                       threshold: float = 0.6) -> bool:
+    """Performs the note coloring, stores the reports as TSV files in the reviewed folder, and logs warnings about
+    those chord label segments where the ratio of out-of-label chords is greater than the given threshold.
+
+
+    Args:
+        parse_obj:
+            Parse object with parsed scores containing labels. Coloring will be performed in the XML structure in the
+            memory and scores have to be written to disk to see the result.
+        out_dir: By default, reports are written to <CORPUS_PATH>/reviewed unless another path is specified here.
+        threshold: Above which ratio of out-of-label tones a warning is to be issued.
+
+    Returns:
+        False if at least one label went beyond the threshold, True otherwise.
+    """
+    review_reports = parse_obj.color_non_chord_tones()
+    test_passes = True
+    for (corpus_name, fname), file_df_pairs in review_reports.items():
+        piece_logger = get_logger(parse_obj[corpus_name].logger_names[fname])
+        is_first = True
+        for file, report in file_df_pairs:
+            report_path = compute_path_from_file(file, root_dir=out_dir, folder='reviewed')
+            report_file = os.path.join(report_path, file.fname + '_reviewed.tsv')
+            if not is_first and os.path.isfile(report_file):
+                get_logger('ms3.review').warning(f"This coloring report has been overwritten because several scores have the same fname:\n{report_file}")
+            write_tsv(report, report_file)
+            is_first = False
+            warning_selection = (report.count_ratio > threshold) & report.chord_tones.notna()
+            if warning_selection.sum() > 0:
+                test_passes = False
+            else:
+                continue
+            for t in report[warning_selection].itertuples():
+                if len(t.added_tones) > 0:
+                    added = f" plus the added {(fifths2sd(t.added_tones, t.localkey_is_minor))} [{scale_degree2name(t.added_tones, t.localkey, t.globalkey)}]"
+                else:
+                    added = ""
+                msg = f"""The label '{t.label}' in m. {t.mn}, onset {t.mn_onset} (MC {t.mc}, onset {t.mc_onset}) seems not to correspond well to the score (which does not necessarily mean it is wrong).
+In the context of {t.globalkey}.{t.localkey}, it expresses the scale degrees {(fifths2sd(t.chord_tones, t.localkey_is_minor))} [{scale_degree2name(t.chord_tones, t.localkey, t.globalkey)}]{added}.
+The corresponding score segment has {t.n_untouched} within-label and {t.n_colored} out-of-label note onsets, a ratio of {t.count_ratio} > {threshold} (the current, arbitrary, threshold).
[…]
