### 2023-04-26 f61a6d4
bug fix for writing Prelims containing HTML
