### 2023-04-28 c748cdb
Updated Instrumentation class to test_accessing_source_instrument_names tests
