### 2023-04-28 a8ede1a
Updated Instrumentation class to test_accessing_source_instrument_names tests
