### 2023-04-28 10aa6fa
Updated Instrumentation class to test_accessing_source_instrument_names tests
