+++ paper/paper.md
@@ -66,12 +66,12 @@ functionalities accessible for productive every-day workflows.
 
 `ms3` has already been used for creating several datasets, namely version 2 of the Annotated Beethoven Corpus
 [@Neuwirth2018_AnnotatedBeethovenCorpus], the Annotated Mozart Sonatas [@Hentschel2021_AnnotatedMozartSonatas],
-and an annotated corpus of 19th century piano music [@Hentschelunderreview_AnnotatedCorpusTonal]. It has been successful
+and an annotated corpus of 19th century piano music [@Hentschelinpress_AnnotatedCorpusTonal]. It has been successful
 in formatting training and validation data for a chord inference algorithm and for inserting its analytical outputs
 into the respective scores [@McLeod:21; @McLeod].
 Moreover, the library is at the heart of a semi-automated annotation workflow running on GitHub
 [@Hentschel2021_SemiautomatedWorkflowParadigm] and a dependency of the music corpus analysis library
-DiMCAT [@Hentschel2022_DiMCATDIgitalMusicology].
+DiMCAT [@Hentschelsubmittedforreview_IntroducingDiMCATUtilize].
 
 # Acknowledgements
 
