+++ paper/paper.bib
@@ -47,20 +47,19 @@
   copyright = {Creative Commons Attribution 4.0 International, Open Access}
 }
 
-@article{Hentschel2022_DiMCATDIgitalMusicology,
-  title = {{{DiMCAT}}: {{A DIgital Musicology Corpus Analysis Toolkit}}},
-  author = {Hentschel, Johannes and McLeod, Andrew and Rohrmeier, Martin},
-  year = {2022},
-  url = {https://github.com/DCMLab/dimcat},
-  journal = {GitHub},
-  journal = {GitHub repository},
+@article{Hentschelsubmittedforreview_IntroducingDiMCATUtilize,
+  title = {Introducing {{DiMCAT}} to Utilize the Dataframe for Processing and Analyzing Notated Music on a Very Large Scale},
+  author = {Hentschel, Johannes and McLeod, Andrew and Rammos, Yannis and Rohrmeier, Martin},
+  year = {submitted for review}
 }
 
 
-@article{Hentschelunderreview_AnnotatedCorpusTonal,
-  title = {An {{Annotated Corpus}} of {{Tonal Piano Music}} from the {{Long}} 19th {{Century}}},
-  author = {Hentschel, Johannes and Rammos, Yannis and Neuwirth, Markus and Rohrmeier, Martin},
-  year = {under review}
+
+@article{Hentschelinpress_AnnotatedCorpusTonal,
+  title = {An Annotated Corpus of Tonal Piano Music from the Long 19th Century},
+  author = {Hentschel, Johannes and Rammos, Yannis and Neuwirth, Markus and Moss, Fabian Claude and Rohrmeier, Martin},
+  year = {in press},
+  journal = {Empirical Musicology Review}
 }
 
 
