### 2023-05-01 #5195 @johentsch
> **Version:** 1.1.2

Don't know if it makes any difference, @faroit, that in the meantime the library is at version 1.2.7, or do I need to update the bot's issue body above?
