### 2023-05-01 19ddc4e
more expressive output for test_instrumentation_after_instrument_change()
