### 2023-05-01 906527f
turns Instrumentation.INSTRUMENT_DEFAULTS into a class variable
