### 2023-05-01 4f0fc2a
turns Instrumentation.INSTRUMENT_DEFAULTS into a class variable
