### 2023-05-01 d2c04d4
turns Instrumentation.INSTRUMENT_DEFAULTS into a class variable
