### 2023-05-01 3c0c6d1
more error handling
