### 2023-05-01 b8b0578
renames transform_df() => dcml_labels2dicts()
