+++ src/ms3/dezrann.py
@@ -101,12 +101,23 @@ Here is an example of Dezrann file structure:
 import argparse
 import json
 import os
-from typing import Dict, List, TypedDict, Union, Tuple, Optional
+from typing import Dict, List, TypedDict, Union, Tuple, Optional, TypeAlias, Literal
 
 from fractions import Fraction
 import pandas as pd
 
+LINE_VALUES = {
+    1: "top.1",
+    2: "top.2",
+    3: "top.3",
+    4: "bot.1",
+    5: "bot.2",
+    6: "bot.3"
+}
+"""The six annotation layers of the Dezrann app, three above ('top') and three below ('bot') the score."""
 
+DezrannLayer: TypeAlias = Literal["top.1", "top.2", "top.3", "bot.1", "bot.2", "bot.3"]
+"""More expressive than simply annotating with 'str'."""
 
 def safe_frac(s: str) -> Union[Fraction, str]:
     try:
@@ -276,10 +287,10 @@ def convert_dcml_list_to_dezrann_list(values_dict: List[DcmlLabel],
 
 def make_layout(
                cadences: bool = False,
-               harmonies: Optional[str] = None,
-               keys: Optional[str] = None,
-               phrases: Optional[str] = None,
-               raw: Optional[str] = None):
+               harmonies: Optional[DezrannLayer] = None,
+               keys: Optional[DezrannLayer] = None,
+               phrases: Optional[DezrannLayer] = None,
+               raw: Optional[DezrannLayer] = None):
     """
     Compile the line positions for target labels into Dezrann layout parameter.
     """
@@ -299,22 +310,25 @@ def make_layout(
     
 def generate_dez(path_measures: str,
                  path_labels: str,
-                 output_path: str = "labels.dez",
+                 output_path: str,
                  cadences: bool = False,
-                 harmonies: Optional[str] = None,
-                 keys: Optional[str] = None,
-                 phrases: Optional[str] = None,
-                 raw: Optional[str] = None,
-                 origin: Union[str, Tuple[str]] = "DCML"):
-    """
-    path_measures : :obj:`str`
-        Path to a TSV file as output by format_data().
-    path_labels : :obj:`str`
-        Path to a TSV file as output by format_data().
-    output_labels : :obj:`str`
-        Path to a TSV file as output by format_data().
-    origin : :obj:`tuple`
-        Tuple of source(s) from which the labels originate. Defaults to "DCML".
+                 harmonies: Optional[DezrannLayer] = None,
+                 keys: Optional[DezrannLayer] = None,
+                 phrases: Optional[DezrannLayer] = None,
+                 raw: Optional[DezrannLayer] = None,
+                 origin: Union[str, Tuple[str]] = "DCML") -> None:
+    """ Create a .dez file from a path to a measures TSV file and a path to a labels/expanded TSV file.
+
+    Args:
+        path_measures: Path to a DCML measures TSV file.
+        path_labels: Path to a DCML labels/expanded TSV file.
+        output_path: Path to the .dez file to be (over)written.
+        cadences: If True (default), labels in the 'cadence' column (if present) are included as vertical lines.
+        harmonies: Specify a DezrannLayer to include the labels from the 'chord' column.
+        keys: Specify a DezrannLayer to include the labels from the 'localkey' column.
+        phrases: Specify a DezrannLayer to include the labels from the 'phraseend' column.
+        raw: Specify a DezrannLayer to include the labels from the 'label' column.
+        origin: Value to show in Dezrann's "Layer" field. Defaults to 'DCML'.
     """
     try:
         harmonies_df = pd.read_csv(
@@ -338,7 +352,7 @@ def generate_dez(path_measures: str,
[…]
