### 2023-05-01 3fafa71
factors out generate_dez_from_dfs()
