### 2023-05-01 683908b
generate_dez() and generate_dez_from_dfs() return boolean and also accept numerical arguments for Dezrann layers
