+++ src/ms3/dezrann.py
@@ -116,8 +116,26 @@ LINE_VALUES = {
 }
 """The six annotation layers of the Dezrann app, three above ('top') and three below ('bot') the score."""
 
-DezrannLayer: TypeAlias = Literal["top.1", "top.2", "top.3", "bot.1", "bot.2", "bot.3"]
+DEZ_LINE_ARGS = (0, 1, 2, 3, 4, 5, 6, -1, -2, -3)
+
+DezrannLayer: TypeAlias = Literal["top.1", "top.2", "top.3", "bot.1", "bot.2", "bot.3", 1, 2, 3, 4, 5, 6, -1, -2, -3]
 """More expressive than simply annotating with 'str'."""
+
+def transform_line_argument(line: Optional[Union[int, str]]) -> Optional[str]:
+    """Takes a number between -3 and 6 and turns it into one of the possible Dezrann line values.
+    0 is interpreted as None. -1, -2, -3 correspond to 4, 5, 6"""
+    if line is None:
+        return
+    try:
+        line = int(line)
+        assert line in DEZ_LINE_ARGS
+    except (TypeError, ValueError, AssertionError):
+        raise ValueError(f"{line} is not a valid argument, should be within [-3, 6].")
+    if line == 0:
+        return None
+    if line < 0:
+        line = abs(line) + 3
+    return LINE_VALUES[line]
 
 def safe_frac(s: str) -> Union[Fraction, str]:
     try:
@@ -223,7 +241,7 @@ def dcml_labels2dicts(labels: pd.DataFrame,
     label_and_qb = pd.concat([labels[label_column].rename('label'), quarterbeats.astype(float)], axis=1)
     n_before = len(labels.index)
     if label_column == 'phraseend':
-        label_and_qb = label_and_qb[label_and_qb.label == '{']
+        label_and_qb = label_and_qb[label_and_qb.label.fillna('').str.contains('{')]
     if label_column == 'localkey':
         label_and_qb = label_and_qb[label_and_qb.label != label_and_qb.label.shift().fillna(True)]
     else: # {'chord', 'cadence', 'label'}
@@ -318,7 +336,7 @@ def generate_dez_from_dfs(measures_df: pd.DataFrame,
                           keys: Optional[DezrannLayer] = None,
                           phrases: Optional[DezrannLayer] = None,
                           raw: Optional[DezrannLayer] = None,
-                          origin: Union[str, Tuple[str]] = "DCML") -> None:
+                          origin: Union[str, Tuple[str]] = "DCML") -> bool:
     """ Create a .dez file from a measures and a labels/expanded dataframe.
 
     Args:
@@ -331,7 +349,16 @@ def generate_dez_from_dfs(measures_df: pd.DataFrame,
         phrases: Specify a DezrannLayer to include the labels from the 'phraseend' column.
         raw: Specify a DezrannLayer to include the labels from the 'label' column.
         origin: Value to show in Dezrann's "Layer" field. Defaults to 'DCML'.
+
+    Returns:
+        True if a .dez file was written.
     """
+    annotation_layer_arguments = {arg: transform_line_argument(arg_val) for arg, arg_val in zip(("harmonies", "keys", "phrases", "raw"), (harmonies, keys, phrases, raw))}
+    parameters = {arg: arg_val is not None for arg, arg_val in annotation_layer_arguments.items()}
+    parameters['cadences'] = cadences
+    if not any(parameters.values()):
+        print(f"Nothing to do because no features have been selected.")
+        return False
     dezrann_labels = []
     if cadences and 'cadence' in harmonies_df.columns:
         dcml_labels = dcml_labels2dicts(labels=harmonies_df, measures=measures_df, label_column='cadence')
[…]
