### 2023-05-01 10e1ef0
extends test_dcml2dez() so that the number of written .dez labels correspond to the original annotation file
