### 2023-05-01 487242a
adds additional unittest
