### 2023-05-02 026dd73
Added set_instrument method to Instrumentation class. Tests passed
