### 2023-05-02 c08fc6c
Added set_instrument method to Instrumentation class. Tests passed
