### 2023-05-02 e1204b0
Added set_instrument method to Instrumentation class. Tests passed
