+++ src/ms3/bs4_parser.py
@@ -2006,7 +2006,6 @@ and {loc_after} before the subsequent {nxt_name}.""")
 #######################################################################
 ####################### END OF CLASS DEFINITION #######################
 #######################################################################
-
 class Parts(LoggedClass):
     def __init__(self, soup: bs4.BeautifulSoup, **logger_cfg):
         super().__init__('Parts', logger_cfg)
@@ -2048,6 +2047,7 @@ class Instrumentation(LoggedClass):
         super().__init__('Instrumentation', logger_cfg)
         self.part_tracknames = [elem['part_trackName'] for elem in self.INSTRUMENT_DEFAULTS.values()]
         self.soup = soup
+        # 'instrument',
         self.fields_names = ['longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName']
         self.parts = Parts(soup)
         self.text_tags = self.text_tags_fn()  # store references to XML tags
@@ -2062,15 +2062,6 @@ class Instrumentation(LoggedClass):
             cur_dict = {"instrumentId": instrument_info["id"]}
             for name in self.fields_names:
                 if name == "part_trackName":
-                    part_tracknames = part.find_all("trackName")
-                    tags = list(filter(lambda k: k.get_text() in self.part_tracknames,
-                                       part_tracknames))
-                    if len(tags) > 0:
-                        tag = tags[0]
-                    else:
-                        tag = part_tracknames[0]
-
-                    part.trackName.string = tag.get_text()
                     tag = part.trackName
                 else:
                     tag = instrument_info.find(name)
@@ -2093,34 +2084,15 @@ class Instrumentation(LoggedClass):
                 result[key][key_instr_data] = value
         return result
 
-    def get_instrument_name(self, staff_name, return_full=False):
+    def get_instrument_name(self, staff_name):
         fields_data = self.fields
         if staff_name not in self.parts.staff2part.keys() or staff_name not in fields_data:
             raise KeyError(f"No data for staff '{staff_name}'")
         else:
-            return fields_data[staff_name] if return_full else fields_data[staff_name]['trackName']
+            return fields_data[staff_name]['trackName']
 
     def set_instrument(self, staff, trackname):
-        if staff not in self.parts.map_staff2part.keys():
-            raise KeyError(f"Don't recognize key '{staff}'")
-        existing_value = self.get_instrument_name(staff)
-        new_value = str(trackname)
-        if existing_value is not None and existing_value == new_value:
-            self.logger.debug(f"The {staff} was already '{existing_value}' and doesn't need changing.")
-            return
-        new_values = self.INSTRUMENT_DEFAULTS[trackname]
-        for field_to_change in self.fields_names:
-            value = new_values[field_to_change]
-            if self.text_tags[staff][field_to_change] is not None:
-                self.text_tags[staff][field_to_change].string = value
-            else:
-                self.text_tags[staff][field_to_change] = value
-                new_tag = self.soup.new_tag(field_to_change)
-                if value is not None:
-                    new_tag.string = value
-                else:
-                    self.logger.debug(f"The value is None.")
-                self.parts.parts_data[self.parts.map_staff2part[staff]].Instrument.append(new_tag)
+        pass
 
 
 
