+++ new_tests/test_instrumentation.py
@@ -3,7 +3,6 @@
 
 """
 Unittests for changing the instrumentation within MuseScore files.
-
 Instrumentation is encoded in <Part> tags. Each <Part> includes exactly one <Instrument> tag and
 one or several <Staff> tags which are assigned the same instrument. The relevant tags can be seen
 in the TypedDict PartInfo.
@@ -188,7 +187,7 @@ TEST_CASES = {
     "Brahms Op. 99iv.mscx": {
         (1, 'Piano'): {1: 'Piano', 2: 'Piano', 3: 'Piano'},
         (2, 'Piano'): {1: 'Violoncello', 2: 'Piano', 3: 'Piano'},
-        (3, 'Violoncello'): {1: 'Violoncello', 2: 'Violoncello', 3: 'Violoncello'},
+        (3, 'Violoncello'): {1: 'Violoncello', 2: 'Piano', 3: 'Violoncello'},
     },
 
 }
@@ -212,7 +211,7 @@ def test_instrumentation_after_instrument_change(source_path):
         parts = get_instrumentation(soup)
         test_results = {}
         for part in parts:
-            result = tested_object.get_instrument_name(part['staves'][0], return_full=True) #part_info_without_staves(part)
+            result = part_info_without_staves(part)
             for staff_name in part['staves']:
                 if staff_name not in expectation:
                     del(expectation[staff_name])
