### 2023-05-02 4a4766d
Renamed variables and functions to probably something meaningful
