### 2023-05-02 ac0d9f9
Renamed variables and functions to probably something meaningful
