### 2023-05-02 f02c927
has operations.insert_labels_into_score() include exactly the facet in question (not fuzzily)
