### 2023-05-04 johentsch/ms3#64 @TGabor
Documentation - JOSS Paper
Here are some small comments regarding the [JOSS paper submission](https://github.com/openjournals/joss-reviews/issues/5195):

* l.25-l.27: please clarify or add an example where the lossy conversion through musicXML poses a problem in the proposed workflow. In the paper, it is not clear why this could not be solved with MusicXML. I would guess that especially when you need to edit the MuseScore file in ms3, an intermediate step through MusicXML would make certain operations impossible or destroy some parts of the score layout.

 * l.29-30: DataFrames are not necessarily feature matrices. The feature matrices extracted by ms3 are represented as DataFrames.

 * l.31: "... to enable version control" sounds a bit like TSV is required for version control. It definitely simplifies it and it is also much easier to see differences between the annotations across different commits. Could you make this point a bit clearer?

 * l. 35 "write back" -> "save"

 * L.35: To me the term "data translocation" is not clear. Do you mean conversion?
