### 2023-05-04 johentsch/ms3#65 @TGabor
Documentation - JOSS Paper Reference
There is a reference labeled as "to be submitted" in the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5195)

> (McLeod, A., & Rohrmeier, M. (To be submitted). An integrated system for harmonic analysis including chord tone alterations. Transactions of the International Society for Music Information Retrieval.

As it is unclear when this will be the case and the submission also doesn't have a DOI, I'd suggest to remove it.
