### 2023-05-05 2a4c049
adapts Brahms test cases
