### 2023-05-07 5a0a0fc
Allow staff_name to have type str or int
