### 2023-05-07 ef3b064
Allow staff_name to have type str or int
