### 2023-05-08 a96b8cd
Handle invalid trackname
