### 2023-05-08 8e83142
Define lowercase INSTRUMENT_DEFAULTS keys
