### 2023-05-08 e534660
Define lowercase INSTRUMENT_DEFAULTS keys
