### 2023-05-08 534a095
Define lowercase INSTRUMENT_DEFAULTS keys
