### 2023-05-09 e8c74a8
Added enlarge_instrument_defaults_keys method to allow user set the instrument by trackName or part_trackName, also adapted tests to check it
