### 2023-05-09 6c521ec
more robust treatment of input values
