### 2023-05-09 4927fc8
more robust treatment of input values
