+++ paper/paper.bib
@@ -148,6 +148,16 @@
   keywords = {beethoven,Beethoven,corpus research,Corpus research,digital musicology,Digital Musicology,ground truth,Ground truth,harmony,music,Music,music; digital musicology; corpus research; ground,symbolic music data,Symbolic Music Data}
 }
 
+@phdthesis{Petersohn2021_DataframeSystemsTheory,
+  title = {Dataframe {{Systems}}: {{Theory}}, {{Architecture}}, and {{Implementation}}},
+  author = {Petersohn, Devin},
+  year = {2021},
+  address = {{Berkeley}},
+  langid = {english},
+  school = {University of California},
+}
+
+
 @inproceedings{Pugin2014_VerovioLibraryEngraving,
   title = {Verovio: {{A Library}} for {{Engraving MEI Music Notation}} into {{SVG}}},
   shorttitle = {Verovio},
