+++ paper/paper.md
@@ -53,9 +53,8 @@ Despite being one of the most wide-spread score encoding formats, current score 
 [e.g., @Cancino-Chacon2022_PartituraPythonPackage; @Cuthbert2010_Music21ToolkitComputerAided; @Pugin2014_VerovioLibraryEngraving], 
 do not handle it without first performing a lossy conversion to the musicXML format.
 The Python library `ms3` fills this gap. It loads the XML tree of a MuseScore file into working memory, 
-computes the temporal positions of all encoded elements, and transforms those requested by the user
-into [DataFrames](https://pandas.pydata.org/pandas-docs/stable/user_guide/dsintro.html#dataframe), i.e., feature
-matrices. The DataFrames can be used by other Python programs and scripts, or written to Tab-Separated Values (TSV) to
+computes the temporal positions of all encoded elements, and transforms those requested by the user into DataFrames [@Petersohn2021_DataframeSystemsTheory]. 
+The DataFrames can be used by other Python programs and scripts, or written to Tab-Separated Values (TSV) to
 enable version control and processing with other software. The most typical aspects that users extract from a score are
 tables containing notes, measures (bars), metadata, and text labels, in particular those representing analytical annotations.
 Moreover, `ms3` allows the user to remove and insert analytical labels from and into scores and to write back the modified
