### 2023-05-10 09fa4ca
turns inline footnote into normal footnote
