### 2023-05-10 f6c46fc
Avoid anti pattern with INSTRUMENTS DEFAULTS
