### 2023-05-10 c306a64
Avoid anti pattern with INSTRUMENTS DEFAULTS
