+++ paper/paper.bib
@@ -104,21 +104,23 @@
   langid = {english}
 }
 
-@inproceedings{McLeod:21,
-  title={A Modular System for the Harmonic Analysis of Musical Scores using a large vocabulary},
-  author={McLeod, Andrew and Rohrmeier, Martin},
-  booktitle= {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
-  year={2021},
-  pages={435--442},
-  doi = {10.5281/zenodo.5624432}
+@inproceedings{Mcleod2021_ModularSystemHarmonic,
+  title = {A Modular System for the Harmonic Analysis of Musical Scores Using a Large Vocabulary},
+  booktitle = {Proceedings of the 22nd {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
+  author = {Mcleod, Andrew and Rohrmeier, Martin A.},
+  year = {2021},
+  month = nov,
+  pages = {435--442},
+  publisher = {{ISMIR}},
+  address = {{Online}},
+  doi = {10.5281/zenodo.5624433},
+  urldate = {2023-04-21},
+  abstract = {The harmonic analysis of a musical composition is a fundamental step towards understanding its structure. Central to this analysis is the labeling of segments of a piece with chord symbols and local key information. In this work, we propose a modular system for performing such a harmonic analysis, incorporating spelled pitches (i.e., not treating enharmonically equivalent pitches as identical) and using a very large vocabulary of 1540 chords (each with a root, type, and inversion) and 70 keys (with a tonic and mode), leading to a full harmonic characterization similar to Roman numeral analysis. Our system's modular design allows each of its components to model an aspect of harmony at an appropriate level of granularity, and also aids in both flexibility and interpretability. We show that our system improves upon a state-of-the-art model for the task, both on a previously available corpus consisting mostly of pieces from the Classical and Romantic eras of Western music, as well as on a much larger corpus spanning a wider range from the 16th through the 20th centuries.},
+  file = {/home/hentsche/Zotero/storage/LLYHNHZE/Mcleod and Rohrmeier - 2021 - A Modular System for the Harmonic Analysis of Musi.pdf}
 }
 
-@article{McLeod,
-  author  = {McLeod, Andrew and Rohrmeier, Martin},
-  title   = {An Integrated System for Harmonic Analysis Including Chord Tone Alterations},
-  journal = {Transactions of the International Society for Music Information Retrieval},
-  year    = {To be submitted},
-}
+
+
 
 
 
