### 2023-05-11 652f8d4
fixes the problem that make_coloring_reports_and_warnings() failed even if all warnings were to be ignored
