### 2023-05-11 4db0ffe
correct Git paths even on Windows
