+++ setup.cfg
@@ -30,7 +30,16 @@ package_dir =
 # DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
 setup_requires = pyscaffold>=3.2a0,<3.3a0
 # Add here dependencies of your project (semicolon/line-separated), e.g.
-install_requires = beautifulsoup4; lxml; pandas; pathos; webcolors; tqdm; pytablewriter; GitPython; typing_extensions
+install_requires =
+  beautifulsoup4
+  lxml
+  pandas==1.5.3
+  pathos
+  webcolors
+  tqdm
+  pytablewriter
+  GitPython
+  typing_extensions
 # The usage of test_requires is discouraged, see `Dependency Management` docs
 # tests_require = pytest; pytest-cov
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
