### 2023-05-16 cbcff56
pulls INSTRUMENT_DEFAULTS and enlarge_instrument_default_keys() out of Instrumentation class making sure that the latter is not called more than once
