### 2023-05-16 9b27f01
renames enlarge_instrument_defaults_keys() => get_enlarged_default_dict() and has it include the default instrumentations (rather than detour via translation to default keys). Normalizes the user input and enables using other default values (instrument abbreviations, MIDI channel name) as input values, too. Adapts test cases accordingly.
