### 2023-05-16 bb79661
integrates test_instrumentation.py with the test suite
