### 2023-05-16 47cde98
coloring report count columns consistently as integers
