### 2023-05-16 johentsch/ms3#64 @TGabor
@johentsch I disagree that “lossy conversion” suffices as an exclusion criterion for research data.
For example, if you’re only interested to derive some musical features independent of the layout, then it should be sufficient to rely on a lossy conversion that only affects the layout.
I see that it will be especially important to have a lossless conversion once you want to write modified annotations back to the original MuseScore file.
But this is not clear in the motivation in l.25
