### 2023-05-16 johentsch/ms3#64 @johentsch
The _raison d'être_ for a MuseScore parser is that no other software parses MuseScore format. The musicXML discussion is misguided.
