### 2023-05-16 johentsch/ms3#66 @TGabor
What is the difference between old_tests and new_tests?
When reviewing the tests for https://github.com/openjournals/joss-reviews/issues/5195, I was wondering what the reason was to create a new_tests folder. Are the old_tests deprecated or still relevant?
