### 2023-05-16 johentsch/ms3#66 @johentsch
The old tests are deprecated.
