### 2023-05-16 05745bc
adds score.mscx.parsed.get_instrumentation()
