+++ src/ms3/bs4_parser.py
@@ -2030,6 +2030,11 @@ INSTRUMENT_DEFAULTS = {
                         'part_trackName': 'Harpsichord',
                         'shortName': 'Hch.',
                         'trackName': 'Harpsichord'},
+        'organ': {'instrumentId': 'keyboard.organ',
+                  'longName': 'Organ',
+                  'part_trackName': 'Organ',
+                  'shortName': 'Org.',
+                  'trackName': 'Organ'},
         'piano': {'instrumentId': 'keyboard.piano',
                   'longName': 'Piano',
                   'part_trackName': 'Piano',
