### 2023-05-17 d43077b
Add full INSTRUMENT_DEFAULTS file and adapted tests
