### 2023-05-17 efaf06d
Add full INSTRUMENT_DEFAULTS file and adapted tests
