+++ new_tests/test_instrumentation.py
@@ -181,9 +181,9 @@ TEST_CASES = {
     },
     "Brahms Op. 99iv.mscx": {
         (1, 'keyboard.piano'): {1: 'piano'},
-        (2, 'pno'): {1: 'violoncello', 2: 'piano', 3: 'piano'},
-        (3, 'cello'): {1: 'violoncello', 2: 'violoncello', 3: 'violoncello'},
-        (3, 'violoncello'): {1: 'violoncello', 2: 'violoncello', 3: 'violoncello'},
+        (2, 'pno'): {2: 'piano', 3: 'piano'},
+        # (3, 'cello'): {3: 'violoncello'},
+        (3, 'violoncello'): {3: 'violoncello'},
     },
 
 }
