### 2023-05-18 #5195 @faroit
> > **Version:** 1.1.2
>
> Don't know if it makes any difference, @faroit, that in the meantime the library is at version 1.2.7, or do I need to update the bot's issue body above?

@johentsch I updated the version to 1.2.9 here. I hope that is ok
