### 2023-05-18 johentsch/ms3#64 @faroit
@johentsch hooking into this discussion here:

in the paper you state:

> For example, the Free and Open Source Software [MuseScore](https://musescore.org/) provides a full-featured yet
intuitive interface for engraving music, but its native XML format does not explicitly encode the temporal positions of
events such as notes and rests. Hence the need for a parser that extracts the implicit information and stores it in an
interoperable format.

> Despite being one of the most wide-spread score encoding formats, current score parsers
[e.g., @Cancino-Chacon2022_PartituraPythonPackage; @Cuthbert2010_Music21ToolkitComputerAided; @Pugin2014_VerovioLibraryEngraving],
do not handle it without first performing a lossy conversion to the musicXML format.
The Python library `ms3` fills this gap.

I agree that having only `lossy conversion` parsers, is indeed sufficient as a motivation to write a new one. However, for readers out-of-domain (like myself), I am with @TGabor, that a short, easy to understand, example where that loss would matter and what ms3 enables specifically that wasn't possible with existing parsers.
