### 2023-05-23 e8c11a8
adds a footnote with an example for lossy conversion to musicXML, addressing #64
