### 2023-05-23 828a266
cleaning up old_tests: updating test_results, removing files generated during tests
