+++ old_tests/test_parse.py
@@ -21,8 +21,8 @@ def test_json_parse():
     os.chdir(os.path.dirname(os.path.realpath(__file__)))
     directory = 'test_results'
     goal = 72
-    p = Parse(directory=directory, file_paths=paths, key='old_tests', logger_cfg=dict(level='d'))
-    assert len(p.files['old_tests']) >= goal, f"Failed to parse list of paths that included several JSON files containing paths: Loaded only {len(p.files['old_tests'])} instead of {goal} or more."
+    p = Parse(directory=directory, file_paths=paths, level='d')
+    assert p.n_detected >= goal, f"Failed to parse list of paths that included several JSON files containing paths: Loaded only {len(p.files['old_tests'])} instead of {goal} or more."
 
 
 
@@ -30,7 +30,7 @@ def test_json_parse():
 def parsed_mscx():
     test_dir = os.path.dirname(os.path.realpath(__file__))
     ms3_dir = os.path.dirname(test_dir)
-    p = Parse(ms3_dir, file_re='mscx$', folder_re='MS3', logger_cfg=dict(level='d'))
+    p = Parse(ms3_dir, file_re='mscx$', folder_re='MS3', level='d')
     p.parse()
     return p
 
@@ -43,7 +43,7 @@ class TestParse:
     def test_extract(self, parsed_mscx):
         target = self.test_results
         path_dict = parsed_mscx.store_extracted_facets(notes_folder=target, notes_suffix='_notes', measures_folder=target, measures_suffix="_measures", labels_folder=target,
-                                                       labels_suffix='_labels')
+                                                       labels_suffix='_labels', metadata_suffix='')
         # for path, what in path_dict.items():
         #     original_path = os.path.join(test_folder, what, )
 
