### 2023-05-23 7cb6aef
updates defaults so that inserting labels from detached and on-disk Annotations objects yields the same scores
