### 2023-05-23 d60ae13
propagate labels_cfg down to the Piece and the Score level; improve the three change_labels_cfg() methods
