+++ src/ms3/piece.py
@@ -12,7 +12,8 @@ from .transformations import dfs2quarterbeats
 from .utils import File, infer_tsv_type, automatically_choose_from_disambiguated_files, ask_user_to_choose_from_disambiguated_files, \
     files2disambiguation_dict, get_musescore, load_tsv, metadata2series, pretty_dict, resolve_facets_param, \
     available_views2str, argument_and_literal_type2list, check_argument_against_literal_type, make_file_path, write_tsv, assert_dfs_equal, \
-    parse_tsv_file_at_git_revision, disambiguate_files, replace_index_by_intervals, AUTOMATIC_COLUMNS, MUSESCORE_HEADER_FIELDS, MUSESCORE_METADATA_FIELDS, LEGACY_COLUMNS
+    parse_tsv_file_at_git_revision, disambiguate_files, replace_index_by_intervals, AUTOMATIC_COLUMNS, MUSESCORE_HEADER_FIELDS, MUSESCORE_METADATA_FIELDS, LEGACY_COLUMNS, \
+    update_labels_cfg
 from .score import Score
 from .logger import LoggedClass
 from .view import View, DefaultView
@@ -23,7 +24,7 @@ class Piece(LoggedClass):
 
     _deprecated_elements = ["get_dataframe"]
 
-    def __init__(self, fname: str, view: View = None, logger_cfg={}, ms=None):
+    def __init__(self, fname: str, view: View = None, labels_cfg={}, ms=None, **logger_cfg):
         super().__init__(subclass='Piece', logger_cfg=logger_cfg)
         self.name = fname
         available_types = ('scores',) + Score.dataframe_types
@@ -63,20 +64,21 @@ class Piece(LoggedClass):
         """If the :class:`~.corpus.Corpus` has :attr:`~.corpus.Corpus.metadata_tsv`, this field will contain the
         {column: value} pairs of the row pertaining to this piece.
         """
-        # self.score_obj: Score = None
-        # self.score_metadata: pd.Series = None
-        # """:obj:`pandas.Series`
-        # Metadata from :attr:`.Score.metadata`, transformed by :func:`.utils.metadata2series`.
-        # """
-        # self.measures: pd.DataFrame = None
-        # self.notes: pd.DataFrame = None
-        # self.rests: pd.DataFrame = None
-        # self.notes_and_rests: pd.DataFrame = None
-        # self.labels: pd.DataFrame = None
-        # self.expanded: pd.DataFrame = None
-        # self.events: pd.DataFrame = None
-        # self.chords: pd.DataFrame = None
-        # self.form_labels: pd.DataFrame = None
+
+        self.labels_cfg = {
+            'staff': None,
+            'voice': None,
+            'harmony_layer': None,
+            'positioning': False,
+            'decode': True,
+            'column_name': 'label',
+            'color_format': None,
+        }
+        """:obj:`dict`
+        Configuration dictionary to determine the output format of :py:attr:`~.score.Score.labels` and
+        :py:attr:`~.score.Score.expanded` tables. The dictonary is passed to :py:attr:`~.score.Score` upon parsing.
+        """
+        self.labels_cfg.update(update_labels_cfg(labels_cfg, logger=self.logger))
 
     def all_facets_present(self, view_name: Optional[str] = None,
                            selected_facets: Optional[Facets] = None) -> bool:
@@ -430,43 +432,26 @@ class Piece(LoggedClass):
         if inferred_type in ('labels', 'expanded'):
             self.ix2annotations = Annotations(df = parsed_tsv)
 
-    # def make_annotation_objects(self, view_name: Optional[str] = None, label_col='label'):
[…]
