+++ old_tests/test_score.py
@@ -6,7 +6,8 @@ import os
 import tempfile
 
 import ms3
-from ms3.utils import assert_all_lines_equal, assert_dfs_equal, decode_harmonies, load_tsv
+from ms3.utils import assert_all_lines_equal, assert_dfs_equal, decode_harmonies, load_tsv, write_tsv, no_collections_no_booleans
+
 
 @pytest.fixture(
     params=['Did03M-Son_regina-1762-Sarti.mscx', 'D973deutscher01.mscx', '05_symph_fant.mscx', 'BWV_0815.mscx', 'K281-3.mscx', '76CASM34A33UM.mscx', 'stabat_03_coloured.mscx'],
@@ -29,6 +30,7 @@ class TestBasic:
 class TestScore:
 
     test_folder = os.path.dirname(os.path.realpath(__file__))
+    test_results = os.path.join(test_folder, 'test_results')
 
     def test_parse_and_write_back(self, score_object):
         original_mscx = score_object.full_paths['mscx']
@@ -50,8 +52,8 @@ class TestScore:
     def test_store_and_load_labels(self, score_object):
         if score_object.mscx.has_annotations:
             fname = score_object.fnames['mscx'] + '_labels.tsv'
-            score_object.store_annotations(tsv_path=fname)
-            score_object.load_annotations(fname, key='tsv')
+            labels_path = os.path.join(self.test_results, fname)
+            score_object.load_annotations(labels_path, key='tsv')
             score_object.detach_labels('labels')
             score_object.attach_labels('tsv')
             try:
@@ -67,41 +69,49 @@ class TestScore:
 
     def test_expanded_labels(self, score_object):
         if score_object.mscx.has_annotations:
-            fname = score_object.fnames['mscx']
-            fpath = os.path.join(score_object.paths['mscx'], '..')
-            new_path = os.path.join(fpath, fname + '_labels.tsv')
-            old_path = os.path.join(fpath, 'harmonies', fname + '.tsv')
-            new_expanded = load_tsv(new_path)
-            old_expanded = load_tsv(old_path)
-            assert_dfs_equal(old_expanded, decode_harmonies(new_expanded, logger='old_test'))
+            fname = score_object.fnames['mscx'] + '_labels.tsv'
+            old_path = os.path.join(self.test_results, fname)
+            old_labels = decode_harmonies(load_tsv(old_path))
+            try:
+                extracted_labels = no_collections_no_booleans(score_object.mscx.labels())
+                with tempfile.NamedTemporaryFile(mode='r+', suffix='.tsv', dir=self.test_folder, encoding='utf-8', delete=False) as tmp_file:
+                    extracted_labels.to_csv(tmp_file, sep='\t', index=False)
+                    new_path = tmp_file.name
+                new_labels = load_tsv(new_path)
+                assert len(new_labels) > 0
+                assert_dfs_equal(old_labels, new_labels)
+            finally:
+                os.remove(tmp_file.name)
 
     def test_parse_to_measurelist(self, score_object):
-        fname = score_object.fnames['mscx']
-        fpath = os.path.join(score_object.paths['mscx'], '..')
-        old_path = os.path.join(fpath, 'measures', fname + '.tsv')
-        old_measurelist = load_tsv(old_path, index_col=None)
-        new_measurelist = score_object.mscx.measures()
-        # Exclude 'repeat' column because the old parser used startRepeat, endRepeat and newSection
-        # Exclude 'offset' and 'next' because the new parser does them more correctly
-        excl = ['repeats', 'offset', 'next']
-        new_measurelist.next = new_measurelist.next.map(lambda l: ', '.join(str(s) for s in l))
[…]
