+++ old_tests/parse_original.sh
@@ -1,1 +1,0 @@
-python ~/corpora/scripts/parallel_parse.py .
