+++ docs/examples/parse_single_score.py
@@ -1,5 +1,5 @@
 from ms3 import Score
 import os
-stabat_path = os.path.join("..", "docs", "stabat.mscx")
+stabat_path = os.path.abspath(os.path.join("..",  "..", "docs", "stabat.mscx"))
 s = Score(stabat_path)
 print(s)
