### 2023-06-05 f399a3c
set logging level on top-level logger, not using logging.basicConfig(). solves #68
