### 2023-06-05 a9deb9f
detecting the type of TSV files based on path and filenames defaults to 'labels' (not 'unknown')
