### 2023-06-05 a35c326
moves conftest.py up to tests folder
