+++ tests/test_local_files/conftest.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    Dummy conftest.py for ms3.
-
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
-"""
-
-# import pytest
