### 2023-06-06 johentsch/ms3#64 @johentsch
Thanks for your feedback, @TGabor and @faroit. I hadn't pinged you after my changes because I took #66 to reorganize the tests. It's done locally and they'll come out with v1.3.0. Today or tomorrow. Could we wait until then, @faroit?
