### 2023-06-13 4b1b249
adds 'conti' to utils.FORM_TOKEN_ABBREVIATIONS
