### 2023-06-13 14c9376
supresses warnings about multiple Fingering_text
