### 2023-06-13 johentsch/ms3#66 @johentsch
Cleaned up the unittests, old tests have been integrated again with v1.2.10
