### 2023-06-16 ab9a3c5
omit version from .warnings files
