### 2023-06-17 ca7566f
replaces all lru_cache wrappers with cache (since Python 3.9 this is considered the more readable equivalent to lru_cache(maxsize=None))
