### 2023-06-17 782b1ac
makes sure 'scalar' is always of type Fraction (not int); removes the alias 'frac' for fractions.Fraction
