+++ src/ms3/bs4_parser.py
@@ -73,7 +73,7 @@
 
 import re, sys, warnings
 from copy import copy
-from fractions import Fraction as frac
+from fractions import Fraction
 from collections import defaultdict, ChainMap # for merging dictionaries
 from itertools import zip_longest
 from typing import Literal, Optional, List, Tuple, Dict, overload, Union, Collection, Hashable
@@ -111,20 +111,20 @@ class _MSCX_bs4(LoggedClass):
 
     """
 
-    durations = {"measure": frac(1),
-                 "breve": frac(2),  # in theory, of course, they could have length 1.5
-                 "long": frac(4),   # and 3 as well and other values yet
-                 "whole": frac(1),
-                 "half": frac(1 / 2),
-                 "quarter": frac(1 / 4),
-                 "eighth": frac(1 / 8),
-                 "16th": frac(1 / 16),
-                 "32nd": frac(1 / 32),
-                 "64th": frac(1 / 64),
-                 "128th": frac(1 / 128),
-                 "256th": frac(1 / 256),
-                 "512th": frac(1 / 512),
-                 "1024th": frac(1 / 1024)}
+    durations = {"measure": Fraction(1),
+                 "breve": Fraction(2),  # in theory, of course, they could have length 1.5
+                 "long": Fraction(4),   # and 3 as well and other values yet
+                 "whole": Fraction(1),
+                 "half": Fraction(1, 2),
+                 "quarter": Fraction(1, 4),
+                 "eighth": Fraction(1, 8),
+                 "16th": Fraction(1, 16),
+                 "32nd": Fraction(1, 32),
+                 "64th": Fraction(1, 64),
+                 "128th": Fraction(1, 128),
+                 "256th": Fraction(1, 256),
+                 "512th": Fraction(1, 512),
+                 "1024th": Fraction(1, 1024)}
 
     def __init__(self, mscx_src, read_only=False, logger_cfg={}):
         """
@@ -251,9 +251,9 @@ class _MSCX_bs4(LoggedClass):
                 for voice_id, voice_node in enumerate(voice_nodes, start=1):
                     if not self.read_only:
                         self.tags[mc][staff_id][voice_id] = defaultdict(list)
-                    current_position = frac(0)
-                    duration_multiplier = 1
-                    multiplier_stack = [1]
+                    current_position = Fraction(0)
+                    duration_multiplier = Fraction(1)
+                    multiplier_stack = [Fraction(1)]
                     tremolo_type = None
                     tremolo_component = 0
                     # iterate through children of <voice> which constitute the note level of one notational layer
@@ -265,7 +265,7 @@ class _MSCX_bs4(LoggedClass):
                             'staff': staff_id,
                             'voice': voice_id,
                             'mc_onset': current_position,
-                            'duration': frac(0)}
+                            'duration': Fraction(0)}
 
                         if event_name == 'Chord':
                             event['chord_id'] = chord_id
@@ -317,10 +317,10 @@ class _MSCX_bs4(LoggedClass):
                         elif event_name == 'Rest':
                             event['duration'], dot_multiplier = bs4_rest_duration(event_node, duration_multiplier)
                         elif event_name == 'location':  # <location> tags move the position counter
-                            event['duration'] = frac(event_node.fractions.string)
+                            event['duration'] = Fraction(event_node.fractions.string)
                         elif event_name == 'Tuplet':
                             multiplier_stack.append(duration_multiplier)
[…]
