### 2023-06-18 7aeefc4
turns utils.py into package and adds make_frictionless_schema() to utils.frictionless
