+++ setup.cfg
@@ -32,6 +32,7 @@ setup_requires = pyscaffold>=3.2a0,<3.3a0
 # Add here dependencies of your project (semicolon/line-separated), e.g.
 install_requires =
   beautifulsoup4
+  frictionless
   lxml
   pandas==1.5.3
   pathos
