### 2023-06-18 bfcefdc
replaces CSVW metadata files with frictionless resource descriptors; Corpus.store_extracted_facets(validate=False) to disable
