+++ src/ms3/corpus.py
@@ -9,23 +9,19 @@ from collections import Counter, defaultdict
 
 import pandas as pd
 import numpy as np
-from git import Repo, InvalidGitRepositoryError
-from gitdb.exc import BadName
 
-from .annotations import Annotations
-from .logger import LoggedClass, get_logger, get_log_capture_handler, temporarily_suppress_warnings, function_logger, normalize_logger_name
+from ms3.utils.frictionless import store_descriptor_and_validate
+
+from .logger import LoggedClass, get_logger, get_log_capture_handler, normalize_logger_name
 from .piece import Piece
 from .score import Score, compare_two_score_objects
-from ._typing import FileDict, FileList, ParsedFile, FileParsedTuple, ScoreFacets, FileDataframeTupleMaybe, FacetArguments, \
+from ._typing import FileDict, FileList, ParsedFile, FileParsedTuple, ScoreFacets, FacetArguments, \
     Facet, ScoreFacet, FileScoreTuple, FileDataframeTuple, AnnotationsFacet
-from .utils import File, column_order, get_musescore, get_path_component, group_id_tuples, iter_selection, join_tsvs, \
-    load_tsv, make_continuous_offset_series, \
-    make_id_tuples, make_playthrough_info, METADATA_COLUMN_ORDER, path2type, \
-    pretty_dict, resolve_dir, \
-    update_labels_cfg, write_metadata, write_tsv, available_views2str, prepare_metadata_for_writing, \
+from .utils import File, column_order, get_musescore, join_tsvs, load_tsv, METADATA_COLUMN_ORDER, path2type, \
+    pretty_dict, resolve_dir, update_labels_cfg, write_metadata, write_tsv, available_views2str, prepare_metadata_for_writing, \
     files2disambiguation_dict, ask_user_to_choose, resolve_paths_argument, make_file_path, resolve_facets_param, check_argument_against_literal_type, LATEST_MUSESCORE_VERSION, \
     convert, string2identifier, write_markdown, parse_ignored_warnings_file, parse_tsv_file_at_git_revision, disambiguate_files, enforce_fname_index_for_metadata, \
-    store_csvw_jsonld, scan_directory
+    scan_directory
 from .view import DefaultView, View, create_view_from_parameters
 
 
@@ -2596,7 +2592,9 @@ class Corpus(LoggedClass):
                                simulate: bool = False,
                                unfold: bool = False,
                                interval_index: bool = False,
-                               silence_label_warnings: bool = False) -> List[str]:
+                               silence_label_warnings: bool = False,
+                               validate: bool = True,
+                               ) -> List[str]:
         """  Store facets extracted from parsed scores as TSV files.
 
         Args:
@@ -2618,6 +2616,10 @@ class Corpus(LoggedClass):
               playthrough, including correct positioning of first and second endings.
           interval_index:
           silence_label_warnings:
+          validate:
+            If True (default), the file is written together with a frictionless resource descriptor JSON file
+            whose column schema is used to validate the stored TSV file.
+
 
         Returns:
 
@@ -2641,12 +2643,6 @@ class Corpus(LoggedClass):
 
         # if the view is default (no additional filters have been set), write one CSVW metadata file per facet
         view = self.get_view(view_name)
-        store_tsv_metadata = view.is_default(relax_for_cli=True)
-        if store_tsv_metadata:
-            self.logger.debug(f"Found that the view '{view.name}' has default settings. Will create csv-metadata.json file(s).")
-            column_combinations = defaultdict(set)
-            facet2files = defaultdict(list)
-            facet2path = dict()
         if target > 0:
             for piece_name, piece in self.iter_pieces(view_name=view_name):
[…]
