+++ src/ms3/utils/frictionless.py
@@ -1,15 +1,21 @@
+import json
+import os
+import re
 from functools import cache
+from pprint import pformat
+from typing import Sequence
 
 import frictionless as fl
 import pandas as pd
 
-from ms3.utils import TSV_COLUMN_TITLES, TSV_COLUMN_DESCRIPTIONS, TSV_DTYPES, TSV_COLUMN_CONVERTERS, safe_frac, safe_int, str2inttuple, int2bool
+from ms3._typing import ScoreFacet
+from ms3.utils import TSV_COLUMN_TITLES, TSV_COLUMN_DESCRIPTIONS, TSV_DTYPES, TSV_COLUMN_CONVERTERS, function_logger, safe_frac, safe_int, str2inttuple, int2bool, File
 
 FIELDS_WITHOUT_MISSING_VALUES = (
     'mc',
     'mc_playthrough',
 )
-FRACTION_REGEX = r"?\d+(?:\/\d+)?"
+FRACTION_REGEX = r"\d+(?:\/\d+)?" # r"-?\d+(?:\/\d+)?" for including negative fractions
 INT_ARRAY_REGEX = r"^[([]?(?:-?\d+\s*,?\s*)*[])]?$"
 
 
@@ -63,14 +69,106 @@ def column_name2frictionless_field(column_name) -> dict:
     return field
 
 
-def make_frictionless_schema(df: pd.DataFrame,
-                             as_descriptor: bool = True) -> dict | fl.Schema:
+def make_frictionless_descriptor(column_names: Sequence[str]) -> dict:
     fields = []
-    for column_name in df.columns:
+    for column_name in column_names:
         field = column_name2frictionless_field(column_name)
         fields.append(field)
     descriptor = dict(fields=fields)
-    if as_descriptor:
+    return descriptor
+
+FRICTIONLESS_REGEX = r"^([-a-z0-9._/])+$"
+FRICTIONLESS_INVERSE = r"[^-a-z0-9._/]"
+schemata = {}
+
+def make_resource_name(name: str, replace_char="_") -> str:
+    name = name.lower()
+    if not re.match(FRICTIONLESS_REGEX, name):
+        name = re.sub(FRICTIONLESS_INVERSE, replace_char, name)
+    return name
+
+def assemble_resource_descriptor(name, path, schema):
+    return {
+        "name": make_resource_name(name),
+        "type": "table",
+        "path": path,
+        "scheme": "file",
+        "format": "tsv",
+        "mediatype": "text/tsv",
+        "encoding": "utf-8",
+        "dialect": {
+            "csv": {
+                "delimiter": "\t"
+            }
+        },
+        "schema": schema
+    }
+
+def replace_extension(filepath: str, new_extension: str) -> str:
+    if new_extension[0] != '.':
+        new_extension = '.' + new_extension
+    return os.path.splitext(filepath)[0] + new_extension
+
+def make_json_path(file: File) -> str:
+    return os.path.join(file.directory, f"{file.fname}.resource.json")
+
+SCHEMAS_DIR = os.path.join(os.path.dirname(__file__), 'schemas')
+
+def get_schema_or_url(facet: str, column_names: Sequence[str],):
+    column_initials = ''.join(str(col)[0] for col in column_names)
+    schema_name = f"{facet}_{column_initials}"
+    schema_filename = f"{schema_name}.schema.json"
+    schema_path = os.path.join(SCHEMAS_DIR, schema_filename)
+    if os.path.exists(schema_path):
+        schema_url = f"https://raw.githubusercontent.com/johentsch/ms3/main/src/ms3/utils/schemas/{schema_filename}"
+        # check if URL exists
+        try:
+            fl.Schema(schema_url)
+            return schema_url
[…]
