### 2023-06-18 cf34467
factors out utils.write_to_warnings_file()
