+++ src/ms3/utils/__init__.py
@@ -5404,3 +5404,24 @@ def check_phrase_annotations(df: pd.DataFrame,
         logger.warning(f"Phrase beginning and endings don't match:\n{compare.to_string(index=False)}", extra={"message_id": (16,)})
         return False
     return True
+
+@function_logger
+def write_to_warnings_file(
+        warnings: List[str],
+        file: File,
+        root_dir: Optional[str] = None,
+        remove_if_empty: bool = False,
+):
+    warnings_path = compute_path_from_file(file, root_dir=root_dir, folder='reviewed')
+    warnings_file = os.path.join(warnings_path, file.fname + file.suffix + '.warnings')
+    if len(warnings) > 0:
+        header = f"Warnings encountered during the last execution of ms3 review"
+        header = f"{header}\n{'=' * len(header)}\n\n"
+        os.makedirs(warnings_path, exist_ok=True)
+        with open(warnings_file, 'w', encoding='utf-8') as f:
+            f.write(header)
+            f.write('\n'.join(warnings))
+        logger.info(f"Written warnings to {warnings_file}.")
+    elif remove_if_empty and os.path.isfile(warnings_file):
+        logger.info(f"Problems seem to be solved, removing {warnings_file}")
+        os.remove(warnings_file)
