+++ src/ms3/cli.py
@@ -184,7 +184,7 @@ def extract_cmd(args, parse_obj: Optional[Parse] = None):
             unfold=args.unfold,
             interval_index=args.interval_index,
             silence_label_warnings=silence_label_warnings,
-            iter_corpus=args.iter_corpus,
+            corpuswise=args.corpuswise,
             **suffixes)
 
 def metadata(args, parse_obj: Optional[Parse] = None):
@@ -547,7 +547,7 @@ def get_arg_parser():
                                 help="Unfold the repeats for all stored DataFrames.")
     extract_args.add_argument('--interval_index', action='store_true',
                                 help="Prepend a column with [start, end) intervals to the TSV files.")
-    extract_args.add_argument('--iter_corpus', action='store_true', help="Parse one corpus after the other rather than all at once.")
+    extract_args.add_argument('--corpuswise', action='store_true', help="Parse one corpus after the other rather than all at once.")
 
     select_facet_args = argparse.ArgumentParser(add_help=False)
     select_facet_args.add_argument('--ask', action='store_true',
