### 2023-06-18 6048c16
utils.frictionless.store_descriptor_and_validate() returns report and Corpus.store_extracted_facets() uses the adapted utils.write_to_warnings_file() to write it to an .errors file, analogous to the .warnings files created by ms3 review
