### 2023-06-18 fc0fab9
do not over-correct log message ID components to integers, otherwise the IGNORED_WARNINGS mechanism will not work
