### 2023-06-18 9a2ca24
adds test to test_instrumentation.py that includes accessing the Instrumentation through a Score object and uses a new commit of the test repo where one score is missing <longName> and <shortName>
