+++ tests/conftest.py
@@ -9,8 +9,8 @@ from ms3.utils import scan_directory, capture_parse_logs, ignored_warnings2dict
 
 
 CORPUS_DIR = "~"        # Directory holding your clone of DCMLab/unittest_metacorpus
-TEST_COMMIT = "0fadf30" # commit of DCMLab/unittest_metacorpus for which the tests should pass
-MS3_DIR = os.path.abspath(os.path.join(os.path.realpath(__file__), '..', '..'))
+TEST_COMMIT = "c34b94a" # commit of DCMLab/unittest_metacorpus for which the tests should pass
+MS3_DIR = os.path.normpath(os.path.join(os.path.realpath(__file__), '..', '..'))
 DOCS_DIR = os.path.join(MS3_DIR, 'docs')
 DOCS_EXAMPLES_DIR = os.path.join(DOCS_DIR, 'examples')
 
@@ -101,7 +101,6 @@ def parse_obj(directory, request) -> Parse:
             p.add_dir(os.path.join(directory, 'outputs'))
             for path in files:
                 p.add_files(path, corpus_name='sweelinck_keyboard')
-
     return p
 
 @pytest.fixture(
