+++ src/ms3/utils/constants.py
@@ -1,0 +1,284 @@
+import re
+from collections import namedtuple
+
+import webcolors
+
+MS3_VERSION = '1.2.10'
+LATEST_MUSESCORE_VERSION = '3.6.2'
+COMPUTED_METADATA_COLUMNS = ['TimeSig', 'KeySig', 'last_mc', 'last_mn', 'length_qb', 'last_mc_unfolded', 'last_mn_unfolded', 'length_qb_unfolded',
+                         'volta_mcs', 'all_notes_qb', 'n_onsets', 'n_onset_positions',
+                         'guitar_chord_count', 'form_label_count', 'label_count', 'annotated_key', ]
+"""Automatically computed columns"""
+
+DCML_METADATA_COLUMNS = ['harmony_version', 'annotators', 'reviewers', 'score_integrity', 'composed_start', 'composed_end', 'composed_source']
+"""Arbitrary column names used in the DCML corpus initiative"""
+
+
+MUSESCORE_METADATA_FIELDS = ['composer', 'workTitle', 'movementNumber', 'movementTitle',
+                         'workNumber', 'poet', 'lyricist', 'arranger', 'copyright', 'creationDate',
+                         'mscVersion', 'platform', 'source', 'translator',]
+"""Default fields available in the File -> Score Properties... menu."""
+
+VERSION_COLUMNS = ['musescore', 'ms3_version',]
+"""Software versions"""
+
+INSTRUMENT_RELATED_COLUMNS = ['has_drumset', 'ambitus']
+
+MUSESCORE_HEADER_FIELDS = ['title_text', 'subtitle_text', 'lyricist_text', 'composer_text', 'part_name_text',]
+"""Default text fields in MuseScore"""
+
+OTHER_COLUMNS = ['subdirectory', 'rel_path',]
+
+LEGACY_COLUMNS = ['fnames', 'rel_paths', 'md5',]
+
+AUTOMATIC_COLUMNS = COMPUTED_METADATA_COLUMNS + VERSION_COLUMNS + INSTRUMENT_RELATED_COLUMNS + OTHER_COLUMNS
+"""This combination of column names is excluded when updating metadata fields in MuseScore files via ms3 metadata."""
+
+METADATA_COLUMN_ORDER = ['fname'] + COMPUTED_METADATA_COLUMNS + DCML_METADATA_COLUMNS + MUSESCORE_METADATA_FIELDS + MUSESCORE_HEADER_FIELDS + VERSION_COLUMNS + OTHER_COLUMNS + INSTRUMENT_RELATED_COLUMNS
+"""The default order in which columns of metadata.tsv files are to be sorted."""
+
+SCORE_EXTENSIONS = ('.mscx', '.mscz', '.cap', '.capx', '.midi', '.mid', '.musicxml', '.mxl', '.xml')
+
+STANDARD_COLUMN_ORDER = [
+    'mc', 'mc_playthrough', 'mn', 'mn_playthrough', 'quarterbeats', 'mc_onset', 'mn_onset', 'beat',
+    'event', 'timesig', 'staff', 'voice', 'duration', 'tied',
+    'gracenote', 'nominal_duration', 'scalar', 'tpc', 'midi', 'volta', 'chord_id']
+
+STANDARD_NAMES = ['notes_and_rests', 'rests', 'notes', 'measures', 'events', 'labels', 'chords', 'expanded',
+                  'harmonies', 'cadences', 'form_labels', 'MS3', 'scores']
+""":obj:`list`
+Indicators for corpora: If a folder contains any file or folder beginning or ending on any of these names, it is 
[…]
