### 2023-06-25 69c3e15
adds schema data for column piece, corpus, i; performs parsing string-encoded tuples using ast.literal_eval
