+++ src/ms3/utils/__init__.py
@@ -5,6 +5,7 @@ import json
 import logging
 import sys, platform, shutil, subprocess
 import warnings
+from ast import literal_eval
 from collections import defaultdict, Counter
 from contextlib import contextmanager
 from dataclasses import dataclass, replace
@@ -1634,180 +1635,193 @@ TSV_COLUMN_CONVERTERS = {
     'quarterbeats_all_endings': safe_frac,
     'onset': safe_frac,
     'duration': safe_frac,
-    'scalar': safe_frac, }
+    'scalar': safe_frac,
+    'volta_mcs': literal_eval,
+}
 
-TSV_COLUMN_TITLES = {'absolute_base': 'Int64',
- 'absolute_root': 'Absolute Root',
- 'act_dur': 'Actual Length',
- 'added_tones': 'Added Tones',
- 'alt_label': 'Alternative Label',
- 'barline': 'Barline',
- 'base': 'Base',
- 'bass_note': 'Bass Note',
- 'breaks': 'Breaks',
- 'cadence': 'Cadence',
- 'cadences_id': 'Cadence ID',
- 'changes': 'Changes',
- 'chord': 'Chord',
- 'chord_id': 'Chord ID',
- 'chord_tones': 'Chord Tones',
- 'chord_type': 'Chord Type',
- 'color_a': 'Color Alpha',
- 'color_b': 'Color Blue',
- 'color_g': 'Color Green',
- 'color_html': 'Color HTML',
- 'color_name': 'Color Name',
- 'color_r': 'Color Red',
- 'composed_end': 'Latest Composition Year',
- 'composed_start': 'Earliest Composition Year',
- 'dont_count': 'Exclude from measure count',
- 'duration': 'Duration',
- 'duration_qb': 'Duration in ♩',
- 'expanded_id': 'Label ID',
- 'figbass': 'Chord Inversion',
- 'form': 'Chord Category',
- 'globalkey': 'Global Key',
- 'globalkey_is_minor': 'Global Key is Minor',
- 'gracenote': 'Grace Note',
- 'harmonies_id': 'Label ID',
- 'harmony_layer': 'Harmony Encoding Layer',
- 'keysig': 'Key Signature',
- 'label': 'Label',
- 'label_type': 'Label Type',
- 'leftParen': 'Left Parenthesis',
- 'localkey': 'Local Key',
- 'localkey_is_minor': 'Local Key is Minor',
- 'marker': 'Marker',
- 'mc': 'Measure Count',
- 'mc_offset': 'Offset of Encoded Measure',
- 'mc_onset': 'Offset within Encoded Measure',
- 'mc_playthrough': 'Measure Count Unfolded',
- 'midi': 'MIDI pitch',
- 'mn': 'Measure Number',
- 'mn_onset': 'Offset within Logical Measure',
- 'movementNumber': 'Movement Number',
- 'name': 'Name',
- 'nashville': 'Nashville',
- 'next': 'Next Measure Counts',
- 'nominal_duration': 'Nominal Duration',
- 'notes_id': 'Notes ID',
- 'numbering_offset': 'Numbering Offset',
- 'numeral': 'Roman Numeral',
- 'octave': 'Octave',
- 'offset:x': 'Horizontal Offset',
- 'offset:y': 'Vertical Offset',
- 'offset_x': 'Horizontal Offset',
- 'offset_y': 'Vertical Offset',
- #'onset':
- 'pedal': 'Pedal Point',
- 'phraseend': 'Phrase Annotation',
- #'playthrough':
- 'quarterbeats': 'Offset from Beginning',
[…]
