### 2023-06-25 1fa0042
schemas unequivocally named according to a truncated hash of their column names; columns parsed via ast.literal_eval are interpreted as type='array'; make_frictionless_schema_descriptor accepts primary_key argument;
