+++ src/ms3/utils/frictionless.py
@@ -1,9 +1,12 @@
+import hashlib
 import json
 import os
 import re
+from ast import literal_eval
+from base64 import urlsafe_b64encode
 from functools import cache
-from pprint import pformat
-from typing import Sequence
+from pprint import pprint
+from typing import Optional, Tuple, Iterable
 
 import frictionless as fl
 import pandas as pd
@@ -37,8 +40,6 @@ def column_name2frictionless_field(column_name) -> dict:
         field['title'] = title
     if description:
         field['description'] = description
-    if string_converter is not None:
-        pass
     if string_converter:
         if string_converter == safe_frac:
             field['type'] = 'string'
@@ -51,6 +52,8 @@ def column_name2frictionless_field(column_name) -> dict:
             constraints["pattern"] = INT_ARRAY_REGEX
         elif string_converter == int2bool:
             field['type'] = 'boolean'
+        elif string_converter == literal_eval:
+            field['type'] = 'array'
         else:
             NotImplementedError(f"Unfamiliar with string converter {string_converter}")
     elif pandas_dtype:
@@ -61,20 +64,29 @@ def column_name2frictionless_field(column_name) -> dict:
         elif pandas_dtype in (str, 'string'):
             field['type'] = 'string'
         else:
-            NotImplementedError(f"Don't know how to handle pandas dtype {pandas_dtype}")
+            raise NotImplementedError(f"Don't know how to handle pandas dtype {pandas_dtype}")
     else:
-        NotImplementedError(f"Don't know how to handle column {column_name}")
+        raise NotImplementedError(f"Don't know how to handle column {column_name}")
     if len(constraints) > 0:
         field['constraints'] = constraints
     return field
 
 
-def make_frictionless_descriptor(column_names: Sequence[str]) -> dict:
+def make_frictionless_schema_descriptor(column_names: Tuple[str, ...],
+                                        primary_key: Optional[Tuple[str, ...]] = None,
+                                        **custom_data
+                                        ) -> dict:
     fields = []
     for column_name in column_names:
         field = column_name2frictionless_field(column_name)
+        if not "type" in field:
+            raise ValueError(f"column_name2frictionless_field({column_name!r}) = {field} (missing 'type'!)")
         fields.append(field)
     descriptor = dict(fields=fields)
+    if primary_key:
+        descriptor['primaryKey'] = primary_key
+    if len(custom_data) > 0:
+        descriptor.update(custom_data)
     return descriptor
 
 FRICTIONLESS_REGEX = r"^([-a-z0-9._/])+$"
@@ -87,14 +99,26 @@ def make_resource_name(name: str, replace_char="_") -> str:
         name = re.sub(FRICTIONLESS_INVERSE, replace_char, name)
     return name
 
-def assemble_resource_descriptor(name, path, schema):
-    return {
+def assemble_resource_descriptor(name: str,
+                                 path: str,
+                                 schema: str,
+                                 innerpath: Optional[str] = None,
+                                 ):
+    is_zipped = path.endswith(".zip")
+    if is_zipped:
+        assert innerpath is not None, "Must specify innerpath for zip files."
+    descriptor = {
         "name": make_resource_name(name),
         "type": "table",
         "path": path,
         "scheme": "file",
         "format": "tsv",
         "mediatype": "text/tsv",
+    }
+    if is_zipped:
+        descriptor["compression"] = "zip"
+        descriptor["innerpath"] = innerpath
+    descriptor.update({
         "encoding": "utf-8",
         "dialect": {
             "csv": {
[…]
