### 2023-06-26 9eccb19
more faithful extraction of logger name
