### 2023-06-26 69be1fa
more faithful extraction of logger name
