### 2023-06-26 af0b611
converts --threshold argument to float
