+++ paper/paper.bib
@@ -47,10 +47,11 @@
   copyright = {Creative Commons Attribution 4.0 International, Open Access}
 }
 
-@article{Hentschelsubmittedforreview_IntroducingDiMCATUtilize,
-  title = {Introducing {{DiMCAT}} to Utilize the Dataframe for Processing and Analyzing Notated Music on a Very Large Scale},
+@inproceedings{Hentschelinpress_IntroducingDiMCATProcessing,
+  title = {Introducing {{DiMCAT}} for Processing and Analyzing Notated Music on a Very Large Scale},
+  booktitle = {Proceedings of the 24th {{International Society}} for {{Music Information Retrieval Conference}}, {{ISMIR}}},
   author = {Hentschel, Johannes and McLeod, Andrew and Rammos, Yannis and Rohrmeier, Martin},
-  year = {submitted for review}
+  year = {in press}
 }
 
 
