### 2023-06-29 johentsch/ms3#76 @faroit
[JOSS] reference rendering
👋 @johentsch here is a few more small issues about the references in the JOSS paper

1. the following paper (I guess ISMIR 23) should probably get a year even though it was just accepted. Otherwise try to change the rendering so that `inPress` isn't repeated twice.

2. I guess the company is called `NPC Imaging`. So please change the capitalization here.

3. the colon in "bigger tent". should be outside of the hyphens.

this is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/5195
