### 2023-06-29 johentsch/ms3#76 @johentsch
Thanks @faroit.

1. I had seen already and fixed together with your PR.
2. Done locally, they are `npc Imaging`, actually.
3. This is beyond my control, as I don't control where the period is rendered. It's not part of the [bibtex entry](https://github.com/johentsch/ms3/blob/joss_paper/paper/paper.bib#L220). What do you advise me to do?
