### 2023-06-29 johentsch/ms3#76 @johentsch
Seems more like glitch to me. As we could see from 1. the typesetting is not quite as accomplished as in LaTeX... I'll halt until your update on how to proceed.
