+++ paper/paper.bib
@@ -195,7 +195,7 @@
   title = {The Art of Music Engraving \& Processing},
   author = {Ross, Ted},
   year = {2001},
-  publisher = {{Npc Imaging}},
+  publisher = {{npc Imaging}},
   address = {{Santa Rosa, Calif.}},
   abstract = {Includes history and background of music reproduction, review of processes of music reproduction, preparing music for engraving, discussion of general notation, including charts of beam angles, discussion of specific notation, including spacing for key/time signatures, accidentals, and first notes and repeat bars},
   isbn = {978-0-9706231-1-9},
