### 2023-06-30 d1ba24b
pins pytablewriter to 1.0.0 and updates df2md, renaming it to dataframe2markdown
