+++ setup.cfg
@@ -37,7 +37,7 @@ install_requires =
   pathos
   webcolors
   tqdm
-  pytablewriter
+  pytablewriter==1.0.0
   GitPython
   typing_extensions
 # The usage of test_requires is discouraged, see `Dependency Management` docs
