### 2023-07-04 233fb3c
Tests passed with instrumentation fields: ["id", 'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName',                                        'ChannelName', 'ChannelValue']
