### 2023-07-06 c052372
Tests passed with instrumentation fields: ["id", 'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName',                                        'ChannelName', 'ChannelValue']
