+++ tests/test_metarepo_files/test_instrumentation.py
@@ -209,12 +209,13 @@ def test_instrumentation_after_instrument_change(source_path):
         test_results = {}
         for part in parts:
             print("PART", part)
-            actual_result = part_info_without_staves(part)
+            # actual_result = part_info_without_staves(part)
+            actual_results = tested_object.fields
             for staff_name in part['staves']:
                 if staff_name not in expectation:
                     continue
-                test_results[staff_name] = actual_result
-        print(f"ASSERT: {test_results} == {expectation}")
+                test_results[staff_name] = actual_results[staff_name]
+        print(f"ASSERT: {test_results} == \n {expectation}")
         if test_results != expectation:
             print(f"Setting {staff_to_modify} to {new_instrument!r} did not result in the expected instrumentation.")
             assert test_results == expectation
