### 2023-07-06 236744b
Adapted test_instrumentation_after_instrument_change to transposed default_instruments.csv
