### 2023-07-06 4aaa7d0
Fixed ChannelValue problem
