### 2023-07-06 d0094f7
this parses <Tempo> information at parse time rather than later as processing the events table; this adds three new columns to chords facet, namely "metronome_base", "metronome_number", "metronome_visible"
