+++ src/ms3/bs4_parser.py
@@ -35,6 +35,9 @@
 .. |mc| replace:: :ref:`mc <mc>`
 .. |mc_offset| replace:: :ref:`mc_offset <mc_offset>`
 .. |mc_onset| replace:: :ref:`mc_onset <mc_onset>`
+.. |metronome_base| replace:: :ref:`metronome_base <metronome_base>`
+.. |metronome_number| replace:: :ref:`metronome_number <metronome_number>`
+.. |metronome_visible| replace:: :ref:`metronome_visible <metronome_visible>`
 .. |midi| replace:: :ref:`midi <midi>`
 .. |mn| replace:: :ref:`mn <mn>`
 .. |mn_onset| replace:: :ref:`mn_onset <mn_onset>`
@@ -97,7 +100,16 @@ NOTE_SYMBOL_MAP = {
     'metNoteHalfUp': '𝅗𝅥',
     'metNoteQuarterUp': '𝅘𝅥',
     'metNote8thUp': '𝅘𝅥𝅮',
-    'metAugmentationDot': '.'
+    'metAugmentationDot': '.',
+    "": "𝅝",
+    "": "𝅗𝅥",
+    "": "𝅘𝅥",
+    "": "𝅘𝅥𝅮",
+    "": "𝅘𝅥𝅯",
+    "": "𝅘𝅥𝅰",
+    "": "𝅘𝅥𝅱",
+    "": "𝅘𝅥𝅲",
+    "": ".",
 }
 
 
@@ -351,10 +363,25 @@ class _MSCX_bs4(LoggedClass):
                                         event['thoroughbass_duration'] = duration
                             for text_tag in event_node.find_all('text'):
                                 parent_name = text_tag.parent.name
-                                text = text_tag2str(text_tag)
+                                text_including_html = text_tag2str(text_tag)
+                                text_excluding_html = text_tag2str_recursive(text_tag)
                                 if parent_name == 'Fingering':
                                     # fingerings occur within <Note> tags, if they are to be extracted, they should go into the notes table
                                     continue
+                                if parent_name == 'Tempo':
+                                    tempo_tag = text_tag.parent
+                                    quarters_per_second = float(tempo_tag.tempo.string)
+                                    event['qpm'] = round(quarters_per_second * 60)
+                                    event["tempo"] = text_excluding_html
+                                    metronome_match = re.match(r"^(.+)=(([0-9]+(?:\.[0-9]*)?))$", text_excluding_html)
+                                    if metronome_match:
+                                        base, value = metronome_match.group(1), metronome_match.group(2)
+                                        event["metronome_base"] = base
+                                        event["metronome_number"] = float(value)
+                                    try:
+                                        event["metronome_visible"] = int(tempo_tag.visible.string)
+                                    except AttributeError:
+                                        event["metronome_visible"] = 1
                                 elif parent_name == 'Lyrics':
                                     lyrics_tag = text_tag.parent
                                     no_tag = lyrics_tag.find('no')
@@ -368,20 +395,21 @@ class _MSCX_bs4(LoggedClass):
                                     if syllabic_tag is not None:
                                         match syllabic_tag.string:
                                             case 'begin':
-                                                text = text + '-'
+                                                text_including_html = text_including_html + '-'
                                             case 'middle':
-                                                text = '-' + text + '-'
+                                                text_including_html = '-' + text_including_html + '-'
                                             case 'end':
-                                                text = '-' + text
+                                                text_including_html = '-' + text_including_html
[…]
