### 2023-07-06 d23e451
more explicit logging
