### 2023-07-06 2911512
throw an warning if first measure does not contain a metronom marking
