### 2023-07-07 johentsch/ms3#76 @faroit
@johentsch sorry for taking so long... I checked for 3 and it turns out that this is actually a US grammar rule and not a bug 😱

https://www.grammar-monster.com/lessons/quotation_(speech)_marks_punctuation_in_or_out.htm

So this can be closed.
