### 2023-07-07 a2eda0d
adds docs and base path/url arguments to utils.get_schema_or_url()
