### 2023-07-08 4e9d99a
avoids error when creating frictionless field for 'volta_mcs' that now is parsed via `safe_tuple`
