### 2023-07-08 909361e
Replace output by debug messages
