### 2023-07-09 1466ee3
Replaced KeyError by Warning and used the most similar instrument name instead of the wrong name, added a test case with the wrong name, printed the affected staves within the part
