+++ tests/test_metarepo_files/test_instrumentation.py
@@ -173,7 +173,8 @@ def test_accessing_source_instrument_names(source_path):
 TEST_CASES = {
     "BWV_0815.mscx": {
         (1, 'harpsichord'): {1: 'harpsichord', 2: 'harpsichord'},
-        (2, 'hch'): {1: 'harpsichord', 2: 'harpsichord'}
+        (2, 'hch'): {1: 'harpsichord', 2: 'harpsichord'},
+        (2, 'hchhh'): {1: 'harpsichord', 2: 'harpsichord'}
     },
     "K281-3.mscx": {
         (1, 'piano'): {1: 'piano', 2: 'piano'},
