+++ src/ms3/logger.py
@@ -52,6 +52,9 @@ class MessageType(Enum):
     UNFOLDING_REPEATS_FAILED_WARNING = 26
     DCML_DEFAULT_CORRECTION_WARNING = 27
     WRONGLY_ENCODED_POSITION_WARNING = 28
+    FIRST_BAR_MISSING_TEMPO_MARK_WARNING = 29
+    CORRECTED_INSTRUMENT_TRACKNAME_WARNING = 30
+    INCONSISTENT_INSTRUMENT_CHANGE_WITHIN_PART = 31
 
 
 
