+++ schemas/notes/4tZe7EhYhd694A.schema.yaml
@@ -1,0 +1,125 @@
+fields:
+  - name: corpus
+    type: string
+    title: Corpus
+    description: Name of a group of pieces
+  - name: fname
+    type: string
+    title: Piece identifier
+    description: Name identifier (filename without suffixes) of a piece
+  - name: ix
+    type: string
+  - name: i
+    type: integer
+    title: Index
+    description: An integer serving as row ID
+  - name: mc
+    type: integer
+    title: Measure Count
+    description: Running count of encoded <Measure> tags which do not necessarily
+      express a full measure (e.g. in case of an anacrusis).
+    constraints:
+      required: true
+  - name: mn
+    type: string
+    title: Measure Number
+    description: Measure number as printed in the score, computed from mc, dont_count
+      and numbering_offset.
+  - name: quarterbeats
+    type: string
+    title: Offset from Beginning
+    description: Distance of an event from the piece's beginning. By default, only
+      second endings are taken into account to reflect the proportions of a simply
+      playthrough without repeats.
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: duration_qb
+    type: number
+    title: Duration in ♩
+    description: A float corresponding to duration * 4
+  - name: mc_onset
+    type: string
+    title: Offset within Encoded Measure
+    description: Distance of an event from the beginning of the <Measure> tag.
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: mn_onset
+    type: string
+    title: Offset within Logical Measure
+    description: Distance from the beginning of the logical measure. Relevant, for
+      example, to compute the metric position of an event.
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: timesig
+    type: string
+    title: Time Signature
+    description: Given as string, e.g. "4/4".
+  - name: staff
+    type: integer
+    title: Staff
+    description: Number of the staff where an event occurs, 1 designating the top
+      staff.
+  - name: voice
+    type: integer
+    title: Notational Layer
+    description: A number between 1-4 where 1 is MuseScore's default layer (blue),
+      2 the second layer in green with downward stems, etc.
+  - name: duration
+    type: string
+    title: Duration
+    description: As fraction of a whole note.
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: gracenote
+    type: string
+    title: Grace Note
+    description: Name given to a type of grace note in the MuseScore encoding, e.g.
+      "grace16"
+  - name: nominal_duration
+    type: string
+    title: Nominal Duration
+    description: The duration corresponding to a note/rest value without applying
+      any dots or n-tuplets, as fraction of a whole note. Multiplied with "scalar"
+      to yield the actual duration of the note/rest.
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: scalar
+    type: string
+    title: Scalar
+    description: Decimal value reflecting all dots and n-tuplets modifying the duration
+      of a note/rest. Yields the actual duration when multiplied with "nominal_duration".
+    constraints:
+      pattern: \d+(?:\/\d+)?
+  - name: tied
+    type: integer
+    title: Tied Note
+    description: 1 if a note is tied to the following one, -1 if it is being tied
+      to by the previous one, 0 if both.
+  - name: tpc
+    type: integer
+    title: Tonal Pitch Class
+    description: Specified on the line of fifths such that 0 = C, 1 = G, -1 = F, etc.
+  - name: midi
+    type: integer
+    title: MIDI pitch
+    description: Value between 0 and 127 where 60 = C4, 61 = C#4/Db4, etc.
+  - name: name
+    type: string
+    title: Name
+  - name: octave
+    type: integer
[…]
