+++ tests/test_metarepo_files/test_instrumentation.py
@@ -1,0 +1,361 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+Unittests for changing the instrumentation within MuseScore files.
+Instrumentation is encoded in <Part> tags. Each <Part> includes exactly one <Instrument> tag and
+one or several <Staff> tags which are assigned the same instrument. The relevant tags can be seen
+in the TypedDict PartInfo.
+"""
+import os
+from functools import lru_cache
+from typing import Dict, Tuple, Optional, List, TypedDict
+
+import bs4
+import pytest
+from git import Repo
+
+from ms3.bs4_parser import Instrumentation, INSTRUMENT_DEFAULTS
+from ms3 import Score
+from conftest import TEST_COMMIT
+
+UNITTEST_METACORPUS = "~/unittest_metacorpus"
+
+
+@lru_cache()
+def check_metarepo_commit(directory) -> str:
+    repo = Repo(directory)
+    commit = repo.commit('HEAD')
+    sha = commit.hexsha[:len(TEST_COMMIT)]
+    if sha != TEST_COMMIT:
+        print(f"Please checkout unittest_metarepo to {TEST_COMMIT}")
+        assert sha == TEST_COMMIT
+    return directory
+
+# region test utilities
+
+def get_source_target_paths() -> Dict[str, Tuple[str, str]]:
+    """Returns a {file_name -> (source_path, target_path)} dictionary."""
+    metarepo_path = check_metarepo_commit(os.path.expanduser(UNITTEST_METACORPUS))
+    source_files_folder = os.path.join(metarepo_path, "mixed_files")
+
+    target_files_folder = os.path.join(source_files_folder, "changed_instruments")
+    target_files = os.listdir(target_files_folder)
+    print(target_files)
+    file2source_target_path = {}
+    for path, subdirs, files in os.walk(source_files_folder):
+        current_folder = os.path.basename(path)
+        if current_folder.startswith('.') or current_folder == "changed_instruments":
+            continue
+        musescore_files = [f for f in files if f.endswith('.mscx')]
+        for file in musescore_files:
+            if file not in target_files:
+                continue
+            source_path = os.path.join(path, file)
+            target_path = os.path.join(target_files_folder, file)
+            file2source_target_path[file] = (source_path, target_path)
+    assert len(file2source_target_path) > 0, f"Didn't find any relevant files at {target_files_folder}"
+    return file2source_target_path
+
+def get_tag_string_or_none(parent_tag: bs4.Tag, tag_name: str) -> Optional[str]:
+    """Looks for a tag and, if found, returns its non-tag content as a string, otherwise none."""
+    found_tag = parent_tag.find(tag_name)
+    if found_tag is None:
+        return
+    if found_tag.string is None:
+        return
+    return found_tag.string.strip()
+
+class PartInfo(TypedDict):
+    """Relevant instrumentation info of one <Part> tag."""
+    part_trackName: str
+    staves: List[str]
+    longName: str
+    shortName: str
+    trackName: str
+    instrumentId: str
+
+def get_soup(source_path: str) -> bs4.BeautifulSoup:
+    """Parse MSCX file."""
+    with open(source_path, 'r', encoding='utf-8') as F:
+        soup = bs4.BeautifulSoup(F.read(), 'xml')
+    return soup
+
+def get_instrumentation(soup: bs4.BeautifulSoup) -> List[PartInfo]:
+    parts = []
+    for part_tag in soup.find_all('Part'):
+        part_info = PartInfo(
[…]
