+++ src/ms3/logger.py
@@ -53,7 +53,6 @@ class MessageType(Enum):
     DCML_DEFAULT_CORRECTION_WARNING = 27
     WRONGLY_ENCODED_POSITION_WARNING = 28
     FIRST_BAR_MISSING_TEMPO_MARK_WARNING = 29
-    FIRST_BAR_MISSING_TEMPO_MARK_WARNING = 29
     CORRECTED_INSTRUMENT_TRACKNAME_WARNING = 30
     INCONSISTENT_INSTRUMENT_CHANGE_WITHIN_PART = 31
 
