### 2023-07-11 e97e4e4
adapts the parsing of <text> tags such that writing any information to the `event` row dict uses the new safe_update_event (which was not the case for the recently added metronome information)
