+++ src/ms3/utils/__init__.py
@@ -22,7 +22,6 @@ from zipfile import ZipFile as Zip
 import git
 import pandas as pd
 import numpy as np
-import webcolors
 from gitdb.exc import BadName
 from numpy.typing import NDArray
 from pandas._typing import Dtype
@@ -34,9 +33,6 @@ from typing_extensions import Self
 
 from ms3.logger import update_cfg, LogCapturer, function_logger
 from ms3._typing import FileDict, Facet, ViewDict, FileDataframeTupleMaybe
-# from .constants import SCORE_EXTENSIONS, STANDARD_COLUMN_ORDER, rgba, MS3_HTML,\
-#     FORM_LEVEL_CAPTURE_REGEX, FORM_LEVEL_SPLIT_REGEX, FORM_TOKEN_ABBREVIATIONS, \
-#     STANDARD_NAMES_OR_GIT,
 from .constants import *
 
 
