### 2023-07-11 d4c7f1a
turns tests into package to allow for relative import
