+++ tests/test_metarepo_files/test_parse.py
@@ -42,10 +42,10 @@ def assert_all_loggers_level(level):
 
 class TestLogging():
 
-    def test_parallel_log_capture(self, directory):
+    def test_parallel_log_capture(self, small_directory):
         """Compare log messages emitted when parsing the same thing in parallel or iteratively."""
-        a = Parse(directory, level='d')
-        b = Parse(directory, level='d')
+        a = Parse(small_directory, level='d')
+        b = Parse(small_directory, level='d')
         with capture_parse_logs(a.logger) as captured_msgs:
             a.parse_scores(parallel=False)
             non_parallel_msgs = captured_msgs.content_list
@@ -55,21 +55,21 @@ class TestLogging():
         for msg in non_parallel_msgs:
             assert msg in parallel_msgs
 
-    def test_default(self, directory):
-        p = Parse(directory)
+    def test_default(self, small_directory):
+        p = Parse(small_directory)
         p.parse()
         _ = p.get_dataframes(expanded=True)
         assert_all_loggers_level(30)
 
-    def test_debug(self, directory):
-        p = Parse(directory, level='d')
+    def test_debug(self, small_directory):
+        p = Parse(small_directory, level='d')
         p.parse()
         _ = p.get_dataframes(expanded=True)
         assert_all_loggers_level(10)
 
 
-    def test_info(self, directory):
-        p = Parse(directory, level='i')
+    def test_info(self, small_directory):
+        p = Parse(small_directory, level='i')
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
@@ -80,8 +80,8 @@ class TestLogging():
             print('\n'.join(wrong_msgs))
             assert False
 
-    def test_warning(self, directory):
-        p = Parse(directory, level='w')
+    def test_warning(self, small_directory):
+        p = Parse(small_directory, level='w')
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
@@ -93,8 +93,8 @@ class TestLogging():
             assert False
 
 
-    def test_error(self, directory):
-        p = Parse(directory, level='e')
+    def test_error(self, small_directory):
+        p = Parse(small_directory, level='e')
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
@@ -106,8 +106,8 @@ class TestLogging():
             assert False
 
 
-    def test_critical(self, directory):
-        p = Parse(directory, level='c')
+    def test_critical(self, small_directory):
+        p = Parse(small_directory, level='c')
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
@@ -118,8 +118,8 @@ class TestLogging():
             assert False
 
 
-    def test_notset(self, directory):
-        p = Parse(directory, level=0)
+    def test_notset(self, small_directory):
+        p = Parse(small_directory, level=0)
         p.parse()
         _ = p.get_dataframes(expanded=True)
         assert_all_loggers_level(30)
