+++ src/ms3/view.py
@@ -26,14 +26,14 @@ class View(LoggedClass):
     categories = (
         'corpora',
         'folders',
-        'fnames',
+        'pieces',
         'files',
         'suffixes',
         'facets',
         'paths',
     )
     available_facets = ('scores',) + Score.dataframe_types + ('unknown',)
-    singular2category: Dict[str, Category] = dict(zip(('corpus', 'folder', 'fname', 'file', 'suffix', 'facet', 'path'),
+    singular2category: Dict[str, Category] = dict(zip(('corpus', 'folder', 'piece', 'file', 'suffix', 'facet', 'path'),
                                    categories))
     tsv_regex = re.compile(r"\.tsv$", re.IGNORECASE)
     convertible_regex = Score.make_extension_regex(native=False, convertible=True, tsv=False)
@@ -41,7 +41,7 @@ class View(LoggedClass):
 
     def __init__(self,
                  view_name: Optional[str] = 'all',
-                 only_metadata_fnames: bool = False,
+                 only_metadata_pieces: bool = False,
                  include_convertible: bool = True,
                  include_tsv: bool = True,
                  exclude_review: bool = False,
@@ -65,9 +65,9 @@ class View(LoggedClass):
         """{criterion -> {excluded_name -> n_excluded}} dict for keeping track of which file was discarded based on which criterion.
         """
         # booleans
-        self.fnames_in_metadata: bool = True
-        self.fnames_not_in_metadata: bool = not only_metadata_fnames
-        self.fnames_with_incomplete_facets = True
+        self.pieces_in_metadata: bool = True
+        self.pieces_not_in_metadata: bool = not only_metadata_pieces
+        self.pieces_with_incomplete_facets = True
         # properties
         self.name = view_name
         self.include_convertible = include_convertible
@@ -96,24 +96,24 @@ class View(LoggedClass):
         self._name = new_name
 
     @property
-    def only_metadata_fnames(self) -> bool:
-        return self.fnames_in_metadata and not self.fnames_not_in_metadata
+    def only_metadata_pieces(self) -> bool:
+        return self.pieces_in_metadata and not self.pieces_not_in_metadata
 
-    @only_metadata_fnames.setter
-    def only_metadata_fnames(self, value):
-        self.fnames_in_metadata = True
-        self.fnames_not_in_metadata = False
+    @only_metadata_pieces.setter
+    def only_metadata_pieces(self, value):
+        self.pieces_in_metadata = True
+        self.pieces_not_in_metadata = False
 
     def is_default(self,
                    relax_for_cli: bool = False) -> bool:
-        """Checks includes and excludes that may influence the selection of fnames. Returns True if the settings
-        do not filter out any fnames. Only if ``relax_for_cli`` is set to True, the filters :attr:`include_convertible`
+        """Checks includes and excludes that may influence the selection of pieces. Returns True if the settings
+        do not filter out any pieces. Only if ``relax_for_cli`` is set to True, the filters :attr:`include_convertible`
         and :attr:`exclude_review` are permitted, too."""
         # define the expected number of filter regexes per category (ignore 'corpora' and 'facets')
         default_excluding_lengths = {
             'suffixes': 0,
             'folders': 0,
-            'fnames': 0,
+            'pieces': 0,
             'files': 0,
             'paths': 0
         }
@@ -121,7 +121,7 @@ class View(LoggedClass):
             if self.exclude_review:
                 default_excluding_lengths.update({
                     'folders': 1,
-                    'fnames': 1,
+                    'pieces': 1,
                     'files': 1
                 })
             default_excluding_lengths['files'] += not self.include_convertible
@@ -131,18 +131,18 @@ class View(LoggedClass):
[…]
