+++ tests/test_metarepo_files/test_score.py
@@ -12,9 +12,9 @@ from ms3 import Score, assert_dfs_equal
 def assert_store_scores_identical(sc_obj, suffix):
     original_mscx = sc_obj.full_paths['mscx']
     original_path = sc_obj.paths['mscx']
-    original_fname = sc_obj.fnames['mscx']
-    fname = f"{original_fname}{suffix}.mscx"
-    tmp_persist = os.path.join(original_path, fname)
+    original_piece_name = sc_obj.piece_names['mscx']
+    piece_name = f"{original_piece_name}{suffix}.mscx"
+    tmp_persist = os.path.join(original_path, piece_name)
     try:
         tmp_file = tempfile.NamedTemporaryFile(mode='r', suffix='.mscx', dir=original_path, encoding='utf-8', delete=True)
         sc_obj.store_score(tmp_file.name)

[…]
