### 2023-07-12 ba66033
replaces the MS3_VERSION constant by a _version.py module that includes the current (potentially dirty) version, rewritten by ms3.__init__.py
