+++ src/ms3/__init__.py
@@ -4,18 +4,23 @@ All functionality of the library is available through creating a ``ms3.Score`` o
 ``ms3.Annotations`` object.
 """
 # -*- coding: utf-8 -*-
-from pkg_resources import get_distribution, DistributionNotFound
+from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
 import os, logging
 os.environ["NUMEXPR_MAX_THREADS"] = "64"
 
 try:
     # Change here if project is renamed and does not equal the package name
     dist_name = __name__
-    __version__ = get_distribution(dist_name).version
-except DistributionNotFound:
-    __version__ = 'unknown'
+    __version__ = version(dist_name)
+except PackageNotFoundError:  # pragma: no cover
+    __version__ = "unknown"
 finally:
-    del get_distribution, DistributionNotFound
+    del version, PackageNotFoundError
+
+version_file_path = os.path.join(os.path.dirname(__file__), "_version.py")
+# store version in the "once canonical place" (https://stackoverflow.com/a/7071358)
+with open(version_file_path, "w") as f:
+    f.write(f'__version__ = "{__version__}"')
 
 from .score import Score
 from .annotations import Annotations
@@ -23,8 +28,6 @@ from .parse import Parse
 from .piece import Piece
 from .corpus import Corpus
 from .utils import *
-from .utils.constants import *
-from .utils.frictionless import *
 from .transformations import *
 from .operations import *
 from .logger import config_logger
