+++ src/ms3/cli.py
@@ -12,8 +12,8 @@ from typing import Optional
 from ms3 import Parse, make_coloring_reports_and_warnings
 from ms3.operations import extract, check, compare, update, store_scores, insert_labels_into_score
 from ms3.utils import convert_folder, resolve_dir, write_tsv, capture_parse_logs, write_to_warnings_file
-from ms3.utils.constants import MS3_VERSION
 from ms3.logger import get_logger, inspect_loggers
+from ms3._version import __version__
 
 __author__ = "johentsch"
 __copyright__ = "École Polytechnique Fédérale de Lausanne"
@@ -576,7 +576,7 @@ def get_arg_parser():
 
 The library offers you the following commands. Add the flag -h to one of them to learn about its parameters. 
 ''')
-    parser.add_argument('--version', action='version', version=MS3_VERSION)
+    parser.add_argument('--version', action='version', version=__version__)
     subparsers = parser.add_subparsers(help='The action that you want to perform.', dest='action')
 
 
