### 2023-07-12 99837b7
factors out operations.transform() from the renamed cli.transform_cli()
