### 2023-07-12 5862abc
enables 'ms3 transform --resources --safe' and adds the 'filename' parameter to operations.transform(), naming the outputs "<filename><suffix>.<facet>.tsv" (previously "concatenated_<facet><suffix>.tsv")
