+++ src/ms3/utils/frictionless_helpers.py
@@ -1,0 +1,476 @@
+import hashlib
+import json
+import os
+import re
+from ast import literal_eval
+from base64 import urlsafe_b64encode
+from functools import cache
+from pprint import pprint
+from typing import Optional, Tuple, Iterable, Literal
+
+import frictionless as fl
+import pandas as pd
+import yaml
+
+from ms3._typing import ScoreFacet
+from .functions import TSV_COLUMN_TITLES, TSV_COLUMN_DESCRIPTIONS, TSV_DTYPES, TSV_COLUMN_CONVERTERS, function_logger, safe_frac, safe_int, str2inttuple, int2bool, File, \
+    safe_tuple, write_tsv
+from ms3.logger import function_logger
+
+FIELDS_WITHOUT_MISSING_VALUES = (
+    'mc',
+    'mc_playthrough',
+)
+FRACTION_REGEX = r"\d+(?:\/\d+)?" # r"-?\d+(?:\/\d+)?" for including negative fractions
+INT_ARRAY_REGEX = r"^[([]?(?:-?\d+\s*,?\s*)*[])]?$"
+
+
+@cache
+def column_name2frictionless_field(column_name) -> dict:
+    global FRACTION_REGEX, INT_ARRAY_REGEX
+    field = dict(
+        name = column_name,
+    )
+    if column_name in FIELDS_WITHOUT_MISSING_VALUES:
+        constraints = dict(required=True)
+    else:
+        constraints = dict()
+    title = TSV_COLUMN_TITLES.get(column_name)
+    description = TSV_COLUMN_DESCRIPTIONS.get(column_name)
+    pandas_dtype = TSV_DTYPES.get(column_name, str)
+    string_converter = TSV_COLUMN_CONVERTERS.get(column_name)
+    if title:
+        field['title'] = title
+    if description:
+        field['description'] = description
+    if string_converter:
+        if string_converter == safe_frac:
+            field['type'] = 'string'
+            constraints["pattern"] = FRACTION_REGEX
+        elif string_converter == safe_int:
+            field['type'] = 'integer'
+            field['bareNumber'] = False # allow other leading and trailing characters
+        elif string_converter == str2inttuple:
+            field['type'] = 'string'
+            constraints["pattern"] = INT_ARRAY_REGEX
+        elif string_converter == int2bool:
+            field['type'] = 'boolean'
+        elif string_converter in (literal_eval, safe_tuple):
+            field['type'] = 'array'
+        else:
+            NotImplementedError(f"Unfamiliar with string converter {string_converter}")
+    elif pandas_dtype:
+        if pandas_dtype in (int, 'Int64'):
+            field['type'] = 'integer'
+        elif pandas_dtype == float:
+            field['type'] = 'number'
+        elif pandas_dtype in (str, 'string'):
+            field['type'] = 'string'
+        else:
+            raise NotImplementedError(f"Don't know how to handle pandas dtype {pandas_dtype}")
+    else:
+        raise NotImplementedError(f"Don't know how to handle column {column_name}")
+    if len(constraints) > 0:
+        field['constraints'] = constraints
+    return field
+
+
+def make_frictionless_schema_descriptor(column_names: Tuple[str, ...],
+                                        primary_key: Optional[Tuple[str, ...]] = None,
+                                        **custom_data
+                                        ) -> dict:
+    fields = []
+    for column_name in column_names:
+        field = column_name2frictionless_field(column_name)
+        if not "type" in field:
+            raise ValueError(f"column_name2frictionless_field({column_name!r}) = {field} (missing 'type'!)")
+        fields.append(field)
[…]
