### 2023-07-13 1b287e9
Complete the new ms3 transform, which now stores datapackage.json descriptor by default, regardless of whether --umcompressed is added or not. --resources stores .resource.json descriptors instead and also works with --umcompressed and without. Since all facets are stored with default suffixes separated by a dot now, the --suffix argument was removed from facet extraction
