### 2023-07-14 a382868
any MultiIndex resulting from concatenation has the last level named "i" (not f"{facet}_i"). store
