### 2023-07-14 0543c5d
store_dataframe_resource() also stores and describes the index levels except when the index is a normal RangeIndex
