### 2023-07-14 213e718
adds ms3 transform schemas for all_subcorpora
