### 2023-07-14 f52f4f6
moves the creation/removal of .errors files to validate_descriptor_path() (previously called validate_resource_descriptor()) and has them appear next to the descriptor file; the boolean parameter 'validate' is replaced: stored descriptors are always validated but the behaviour upon validation errored can be controlled via "raise_exception" and "write_or_remove_errors_file"
