+++ src/ms3/utils/frictionless_helpers.py
@@ -5,7 +5,7 @@ import re
 from ast import literal_eval
 from base64 import urlsafe_b64encode
 from functools import cache
-from pprint import pprint
+from pprint import pprint, pformat
 from typing import Optional, Tuple, Iterable, Literal
 
 import frictionless as fl
@@ -14,7 +14,7 @@ import yaml
 
 from ms3._typing import ScoreFacet, TSVtypes, TSVtype
 from .functions import TSV_COLUMN_TITLES, TSV_COLUMN_DESCRIPTIONS, TSV_DTYPES, TSV_COLUMN_CONVERTERS, function_logger, safe_frac, safe_int, str2inttuple, int2bool, File, \
-    eval_string_to_nested_list, write_tsv, resolve_facets_param
+    eval_string_to_nested_list, write_tsv, resolve_facets_param, write_validation_errors_to_file
 from ms3.logger import function_logger
 
 FIELDS_WITHOUT_MISSING_VALUES = (
@@ -162,7 +162,10 @@ def get_truncated_hash(S: str | Iterable[str],
         S = [S]
     hasher = hash_func()
     for s in S:
-        hasher.update(s.encode('utf-8'))
+        try:
+            hasher.update(s.encode('utf-8'))
+        except AttributeError as e:
+            raise ValueError(f"Element {s!r} from {S} resulted in error {e!r}") from e
     return urlsafe_b64encode(hasher.digest()[:length]).decode('utf-8').rstrip('=')
 
 
@@ -241,7 +244,7 @@ def get_schema_or_url(facet: str,
 
 
 def get_schema(df: pd.DataFrame,
-               facet: ScoreFacet,
+               facet: str,
                include_index_levels: bool = False,
                ) -> dict | str:
     index_levels = df.index.names if include_index_levels else None
@@ -272,7 +275,7 @@ def store_as_json_or_yaml(
 def make_resource_descriptor(
         df: pd.DataFrame,
         piece_name: str,
-        facet: ScoreFacet,
+        facet: str,
         filepath: Optional[str] = None,
         innerpath: Optional[str] = None,
         include_index_levels: bool = False,
@@ -320,7 +323,7 @@ def make_and_store_resource_descriptor(
         df: pd.DataFrame,
         directory: str,
         piece_name: str,
-        facet: ScoreFacet,
+        facet: str,
         filepath: Optional[str] = None,
         innerpath: Optional[str] = None,
         descriptor_extension: Literal["json", "yaml"] = "json",
@@ -366,16 +369,35 @@ def make_and_store_resource_descriptor(
     store_as_json_or_yaml(descriptor, descriptor_path, logger=logger)
     return descriptor_path
 
-def validate_resource_descriptor(
+@function_logger
+def validate_descriptor_path(
         descriptor_path: str,
         raise_exception: bool = True,
+        write_or_remove_errors_file: bool = True,
 ) -> fl.Report:
     report = fl.validate(descriptor_path)
-    if not report.valid:
-        errors = [err.message for task in report.tasks for err in task.errors]
-        if raise_exception:
-            raise fl.FrictionlessException("\n".join(errors))
+    validation_tasks = []
+    if report.valid:
+        logger.debug(f"{descriptor_path} successfully validated.")
+    else:
+        validation_tasks, error_lists = zip(*((pformat(task), [err.message for err in task.errors]) for task in report.tasks))
+        all_errors = sum(error_lists, [])
+        errors_block = '\n'.join(all_errors)
+        logger.error(
+            f"Validation of {descriptor_path} failed with {len(all_errors)} validation errors:\n{errors_block}",
+            extra={"message_id": (32,)}
+        )
+    if write_or_remove_errors_file:
+        # the following call removes the .errors file if the validation was successful
[…]
