### 2023-07-15 6ec63c5
adds the convenience function operations.corpus2default_package() and harmonizes interfaces further
