+++ src/ms3/logger.py
@@ -55,7 +55,7 @@ class MessageType(Enum):
     FIRST_BAR_MISSING_TEMPO_MARK_WARNING = 29
     CORRECTED_INSTRUMENT_TRACKNAME_WARNING = 30
     INCONSISTENT_INSTRUMENT_CHANGE_WITHIN_PART = 31
-    TSV_RESOURCE_INVALID_ERROR = 32
+    FRICTIONLESS_VALIDATION_ERROR_WARNING = 32
 
 
 
