### 2023-07-15 e367aed
adds resource name to header of .errors files
