+++ src/ms3/operations.py
@@ -123,10 +123,24 @@ expanded_folder=expanded_folder,
     for corpus_name, corpus, in parse_obj.iter_independent_corpora():
         print(f"PARSING SCORES FOR CORPUS '{corpus_name}' {mode}...")
         corpus.parse_scores(parallel=parallel)
-        corpus.store_extracted_facets(root_dir=root_dir, notes_folder=notes_folder, rests_folder=rests_folder, notes_and_rests_folder=notes_and_rests_folder,
-                                      measures_folder=measures_folder, events_folder=events_folder, labels_folder=labels_folder, chords_folder=chords_folder,
-                                      expanded_folder=expanded_folder, cadences_folder=cadences_folder, form_labels_folder=form_labels_folder, metadata_suffix=metadata_suffix,
-                                      markdown=markdown, simulate=simulate, unfold=unfold, interval_index=interval_index, silence_label_warnings=silence_label_warnings, **suffixes)
+        corpus.store_extracted_facets(
+            root_dir=root_dir,
+            notes_folder=notes_folder,
+            rests_folder=rests_folder,
+            notes_and_rests_folder=notes_and_rests_folder,
+            measures_folder=measures_folder,
+            events_folder=events_folder,
+            labels_folder=labels_folder,
+            chords_folder=chords_folder,
+            expanded_folder=expanded_folder,
+            cadences_folder=cadences_folder,
+            form_labels_folder=form_labels_folder,
+            metadata_suffix=metadata_suffix,
+            markdown=markdown,
+            simulate=simulate,
+            unfold=unfold,
+            interval_index=interval_index,
+            )
 
 
 def check(parse_obj: Parse,
