+++ src/ms3/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.11"
+__version__ = "1.2.10.post0.dev75+gdec86fc.dirty"
