+++ src/ms3/bs4_parser.py
@@ -2064,13 +2064,34 @@ and {loc_after} before the subsequent {nxt_name}.""")
 ####################### END OF CLASS DEFINITION #######################
 #######################################################################
 class ParsedParts(LoggedClass):
+    """
+           Storing found parts object from a BeautifulSoup file
+
+           Args:
+               soup: bs4.BeautifulSoup,
+                   BeautifulSoup object to parse
+               **logger_cfg:obj:`dict`, optional
+                   The following options are available:
+                   'name': LOGGER_NAME -> by default the logger name is based on the parsed file(s)
+                   'level': {'W', 'D', 'I', 'E', 'C', 'WARNING', 'DEBUG', 'INFO', 'ERROR', 'CRITICAL'}
+                   'file': PATH_TO_LOGFILE to store all log messages under the given path.
+           """
     def __init__(self, soup: bs4.BeautifulSoup, **logger_cfg):
         super().__init__('ParsedParts', logger_cfg)
         self.parts_data: Dict[str, bs4.Tag] = {f"part_{i}": part for i, part in enumerate(soup.find_all('Part'), 1)}
 
     @property
     def staff2part(self) -> dict[list, str]:
-        """Returns the dict in the format {[2, 3]: 'part_1'} for staves 2 and 3 of part 1"""
+        """
+        Allows users to determine the corresponding part based on the staff number
+
+        Example:
+            Returns {[2, 3]: 'part_1'} for staves 2 and 3 of part 1
+
+        Returns:
+            dict[list, str]: the dictionary mapping parts to staves
+
+        """
         staff2part = {}
         for key_part, part in self.parts_data.items():
             staves = [f"staff_{staff['id']}" for staff in part.find_all('Staff')]
@@ -2080,16 +2101,25 @@ class ParsedParts(LoggedClass):
     def __repr__(self):
         return pformat(self.parts_data, sort_dicts=False)
 
+"""Instrument Defaults is a csv file that includes all possible instruments and their properties: 'id', 'longName',
+    'shortName', 'trackName', 'instrumentId', 'part_trackName', 'ChannelName', 'ChannelValue'"""
 INSTRUMENT_DEFAULTS = pd.read_csv(os.path.join(os.path.dirname(os.path.realpath(__file__)), "instrument_defaults.csv"), index_col=0).replace({np.nan: None})
 
 def get_enlarged_default_dict() -> Dict[str, dict]:
     """
-    Allows users to set an instrument by 'id', 'longName', 'shortName',
-       'trackName', 'instrumentId', 'part_trackName'
+    Allows users to point to an instrument not only with a 'trackName', but also with 'id', 'longName', 'shortName',
+       'instrumentId', 'part_trackName'
+
+    Returns:
+        Dict[str, dict]: dictionary mapping any of the possible fields ('id', 'longName', 'shortName', trackName',
+       'instrumentId', 'part_trackName') corresponding to an instrument into complete information about the instrument
+       ('id',  'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName', 'ChannelName', 'ChannelValue')
     """
     enlarged_dict = dict(INSTRUMENT_DEFAULTS.T)
+    # we drop "ChannelName", "ChannelValue" because they are not unique for the instrument, so they can't be keys
     for cur_key, cur_value in INSTRUMENT_DEFAULTS.T.drop(["ChannelName", "ChannelValue"]).to_dict().items():
         added_value = INSTRUMENT_DEFAULTS.T[cur_key]
+        # additional_key takes values from 'id', 'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName'
         for additional_key in cur_value.values():
[…]
