### 2023-07-17 #5195 @faroit
@johentsch can you please proceed with [the checklist above](https://github.com/openjournals/joss-reviews/issues/5195#issuecomment-1613802257) to finish up the review process?

Let me know if you issue a new release, i can then update the version here.
