### 2023-07-17 8f40b16
adds the mechanism from ba66033c that writes the (dirty) version into _version.py when initializing ms3
