### 2023-07-17 0c90771
replaces MS3_VERSION constant with the version from _version.py
