### 2023-07-17 bcfbf64
adapts unittests for ms3 2.0.0
