### 2023-07-18 2cd0abf
adds myst-nb requirement
