+++ docs/requirements.txt
@@ -3,3 +3,4 @@ pydata_sphinx_theme
 sphinx-argparse
 sphinxcontrib-programoutput
 jupyter_sphinx
+myst-nb
