+++ src/ms3/bs4_parser.py
@@ -2103,7 +2103,8 @@ class ParsedParts(LoggedClass):
 
 """Instrument Defaults is a csv file that includes all possible instruments and their properties: 'id', 'longName',
     'shortName', 'trackName', 'instrumentId', 'part_trackName', 'ChannelName', 'ChannelValue'"""
-INSTRUMENT_DEFAULTS = pd.read_csv(os.path.join(os.path.dirname(os.path.realpath(__file__)), "instrument_defaults.csv"), index_col=0).replace({np.nan: None})
+INSTRUMENT_DEFAULTS = pd.read_csv(os.path.join(os.path.dirname(os.path.realpath(__file__)),
+                                               "instrument_defaults.csv"), index_col=0).replace({np.nan: None})
 
 def get_enlarged_default_dict() -> Dict[str, dict]:
     """
@@ -2116,8 +2117,9 @@ def get_enlarged_default_dict() -> Dict[str, dict]:
        ('id',  'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName', 'ChannelName', 'ChannelValue')
     """
     enlarged_dict = dict(INSTRUMENT_DEFAULTS.T)
-    # we drop "ChannelName", "ChannelValue" because they are not unique for the instrument, so they can't be keys
-    for cur_key, cur_value in INSTRUMENT_DEFAULTS.T.drop(["ChannelName", "ChannelValue"]).to_dict().items():
+    # we drop "ChannelName", "ChannelValue" etc because they are not unique for the instrument, so they can't be keys
+    for cur_key, cur_value in INSTRUMENT_DEFAULTS.T.drop(["ChannelName", "ChannelValue", 'useDrumset', 'clef', 'group',
+                                                          'staff_type_name', 'defaultClef']).to_dict().items():
         added_value = INSTRUMENT_DEFAULTS.T[cur_key]
         # additional_key takes values from 'id', 'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName'
         for additional_key in cur_value.values():
@@ -2142,7 +2144,7 @@ class Instrumentation(LoggedClass):
         self.part_tracknames = INSTRUMENT_DEFAULTS['part_trackName']
         self.soup = soup
         self.instrumentation_fields = ["id", 'longName', 'shortName', 'trackName', 'instrumentId', 'part_trackName',
-                                       'ChannelName', 'ChannelValue']
+                                       'ChannelName', 'ChannelValue', 'useDrumset', 'clef', 'group', 'staff_type_name', 'defaultClef']
         self.parsed_parts = ParsedParts(soup)
         self.soup_references_data = self.soup_references()  # store references to XML tags
 
@@ -2161,10 +2163,16 @@ class Instrumentation(LoggedClass):
         tag_dict = {}
         for key_part, part in self.parsed_parts.parts_data.items():
             instrument_tag = part.Instrument
-            staves: Dict[str, bs4.Tag] = [f"staff_{(staff['id'])}" for staff in part.find_all('Staff')]
+            staves_list = part.find_all('Staff')
+            staves: Dict[str, bs4.Tag] = [f"staff_{(staff['id'])}" for staff in staves_list]
+            staves_dict = {}
+            for key_staff, data_staff in zip(staves, staves_list):
+                staff_type = data_staff.StaffType
+                staves_dict[key_staff] = {'group': staff_type['group'], 'staff_type_name': staff_type.find("name"), "defaultClef": data_staff.find("defaultClef")}
             channel_info = part.Channel
             channel_name = None if "name" not in channel_info.attrs.keys() else channel_info["name"]
             cur_dict = {"id": instrument_tag["id"], "ChannelName": channel_name, "ChannelValue": channel_info.program}
+            cur_dict.update(staves_dict[staves[0]])
[…]
