### 2023-08-02 e4dcca1
when checking for metronome mark, look in the first two MCs, allowing pieces with incipit or pickup to have theirs only in MC2
