### 2023-08-03 3a65c88
Adapts versions I, II etc but can't work with Cello because it is not included in default._instruments.csv, solves #80
