+++ src/ms3/bs4_parser.py
@@ -2229,6 +2229,46 @@ class Instrumentation(LoggedClass):
         else:
             return fields_data[staff_name]['trackName']
 
+    def add_versions(self, new_values, version):
+        """
+        Adds version in the format I, II etc
+        Args:
+            new_values: the distionary of fields to update
+            version: the string containing version
+
+        Returns:
+            the dictionary with updated names with versions
+        """
+        for version_key in ['trackName', 'longName', 'shortName']:
+            version_value = new_values[version_key]
+            if version_value is not None:
+                new_values[version_key] = f"{version_value} {version.upper()}"
+        return new_values
+
+    def modify_drumset_tags(self, staff_type, value, changed_part, field_to_change):
+        """
+        Sets tags specific for Drumset instruments
+        Args:
+            staff_type: the tags containing info of the field
+            value: new value of the field
+            changed_part: the index of part to update
+            field_to_change: the name of field to update
+
+        """
+        for elem in staff_type:
+            tag = elem.find(field_to_change)
+            if value is not None:
+                if tag is not None:
+                    tag.string = value
+                else:
+                    new_tag = self.soup.new_tag(field_to_change)
+                    new_tag.string = value
+                    elem.append(new_tag)
+                    self.logger.debug(f"Added new {new_tag} with value {value!r} to part {changed_part}")
+            else:
+                if tag is not None:
+                    tag.extract()
+
     def set_instrument(self, staff_id: Union[str, int], trackname):
         """
         Modifies the instrument and all its corresponding information in the soup source file
@@ -2255,86 +2295,77 @@ class Instrumentation(LoggedClass):
 
         # preprocessing and verification of correctness of trackname
         trackname_norm = trackname.lower().strip('.')
+        # add splitting by I or II etc and then adapt other names to it
+        split_trackname = trackname_norm.split()
+        version = split_trackname[-1] if set(split_trackname[-1]) == {"i"} else None
+        if version is not None:
+            trackname_norm = " ".join(split_trackname[:-1])
         if trackname_norm not in self.key2default_instrumentation:
             trackname_norm = difflib.get_close_matches(trackname_norm, list(self.key2default_instrumentation.keys()), n=1)[0]
-            self.logger.warning(f"Don't recognize trackName '{trackname}'. Did you mean {trackname_norm}?", extra=dict(message_id=(30,)))
+            trackname_old = self.fields[staff_id]['trackName'].lower().strip('.')
+            self.logger.warning(f"Don't recognize trackName '{trackname}'. Did you mean {trackname_norm}? I use the fields of old trackName {trackname_old}", extra=dict(message_id=(30,)))
+            if trackname_old not in self.key2default_instrumentation:
+                trackname_old = self.fields[staff_id]['part_trackName'].lower().strip('.')
+            new_values = self.key2default_instrumentation[trackname_old]
         else:
             new_values = self.key2default_instrumentation[trackname_norm]
+            self.updated.update({staff_id: new_values['id']})
+        if version is not None:
+            new_values = self.add_versions(new_values, version)
 
-            # checking that the current changes will not affect other staves
-            staves_within_part = np.array([staff_key for staff_key, part_value in self.parsed_parts.staff2part.items() if
-                                  part_value == changed_part and staff_key != staff_id])  # which staves share this part
-            if len(staves_within_part) > 0:
-                damaged_upd_staves = \
-                list(filter(None, [staff_key if new_values["id"] != self.updated[staff_key] else None for staff_key in set(staves_within_part) & self.updated.keys()]))
-                if len(damaged_upd_staves) > 0:
-                    damaged_dict = {elem: self.updated[elem] for elem in damaged_upd_staves}
-                    damaged_dict[staff_id] = new_values["id"]
[…]
