+++ tests/test_metarepo_files/test_instrumentation.py
@@ -181,7 +181,6 @@ TEST_CASES = {
         (2, 'Pno'): {1: 'piano'}
     },
     "Brahms Op. 99iv.mscx": {
-        (1, 'keyboard.piano'): {1: 'piano'},
         (2, 'pno'): {2: 'piano', 3: 'piano'},
         # (3, 'cello'): {3: 'violoncello'},
         (3, 'violoncello'): {3: 'violoncello'},
