+++ src/ms3/logger.py
@@ -56,7 +56,6 @@ class MessageType(Enum):
     CORRECTED_INSTRUMENT_TRACKNAME_WARNING = 30
     INCONSISTENT_INSTRUMENT_CHANGE_WITHIN_PART = 31
     FRICTIONLESS_VALIDATION_ERROR_WARNING = 32
-    INCONSISTENT_INSTRUMENT_CHANGE_WITHIN_PART_TO_UPDATE = 33
 
 
 
