### 2023-08-03 a0f79de
Fixed version adaptation (changed suffixes, changed typos and descriptions, removed .upper())
