+++ setup.cfg
@@ -4,7 +4,7 @@
 
 [metadata]
 name = ms3
-description = A parser for annotated MuseScore 3 files.
+description = A parser for MuseScore files, serving as data factory for annotated music corpora.
 author = Johannes Hentschel
 author_email = johannes.hentschel@epfl.ch
 license = gpl3
@@ -18,8 +18,13 @@ platforms = any
 # Add here all kinds of additional classifiers as defined under
 # https://pypi.python.org/pypi?%3Aaction=list_classifiers
 classifiers =
-    Development Status :: 4 - Beta
-    Programming Language :: Python
+    Development Status :: 5 - Production/Stable
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Natural Language :: English
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3.10
+    Topic :: Multimedia :: Sound/Audio :: Editors
+    Topic :: Scientific/Engineering
 
 [options]
 zip_safe = False
