### 2023-08-09 21eb728
use .reindex() instead of iteratively adding empty columns (which leads to data fragmentation)
