### 2023-08-09 e53d404
replace '/' (resulting from empty <continueAt> tags) with pd.NA instead of None
