### 2023-08-09 adc29cf
enables installing ms3[dev], currently an alias for ms3[testing]
