+++ setup.cfg
@@ -64,6 +64,8 @@ exclude =
 testing =
     pytest
     pytest-cov
+dev =
+    ms3[testing]
 
 [options.entry_points]
 # Add here console scripts like:
