### 2023-08-09 72b6508
adds pre-commit hooks config
