+++ src/ms3/bs4_measures.py
@@ -1,525 +1,21 @@
 from collections import defaultdict
+from fractions import Fraction as frac
 from typing import Dict, List, Optional, Tuple
 
-import pandas as pd
 import numpy as np
+import pandas as pd
 from numpy.typing import NDArray
-from fractions import Fraction as frac
 
-from .logger import get_logger, function_logger, LoggedClass
+from .logger import LoggedClass, function_logger
 
-class MeasureList(LoggedClass):
-    """ Turns a _MSCX_bs4._measures DataFrame into a measure list and performs a couple of consistency checks on the score.
-
-    Attributes
-    ----------
-    df : :obj:`pandas.DataFrame`
-        The input DataFrame from _MSCX_bs4.raw_measures
-    sections : :obj:`bool`, default True
-        By default, section breaks allow for several anacrusis measures within the piece (relevant for `mc_offset` column)
-        and make it possible to omit a repeat sign in the following bar (relevant for `next` column).
-        Set to False if you want to ignore section breaks.
-    secure : :obj:`bool`, default False
-        By default, measure information from lower staves is considered to contain only redundant information.
-        Set to True if you want to be warned about additional measure information from lower staves that is not taken into account.
-    reset_index : :obj:`bool`, default True
-        By default, the original index of `df` is replaced. Pass False to keep original index values.
-
-    cols : :obj:`dict`
-        Dictionary of the relevant columns in `df` as present after the parse.
-    ml : :obj:`pandas.DataFrame`
-        The measure list in the making; the final result.
-    volta_structure : :obj:`dict`
-        Keys are first MCs of volta groups, values are dictionaries of {volta_no: [mc1, mc2 ...]}
-
-    """
-
-    def __init__(self, df, sections=True, secure=True, reset_index=True, columns={}, logger_cfg={}):
-        """
-
-        Parameters
-        ----------
-        df
-        sections : :obj:`bool`, optional
-            By default, pieces containing section breaks (where counting MNs restarts) receive two more columns in the measures
-            list, namely ``section`` and ``ambiguous_mn`` to grant access to MNs as shown in MuseScore. Pass False to not
-            add such columns.
-        secure
-        reset_index
-        logger_cfg : :obj:`dict`, optional
-            The following options are available:
-            'name': LOGGER_NAME -> by default the logger name is based on the parsed file(s)
-            'level': {'W', 'D', 'I', 'E', 'C', 'WARNING', 'DEBUG', 'INFO', 'ERROR', 'CRITICAL'}
-            'file': PATH_TO_LOGFILE to store all log messages under the given path.
-        """
-        super().__init__(subclass='MeasureList', logger_cfg=logger_cfg)
-        assert len(df) > 0, "Score contains no measures."
-        self.df = df
-        self.ml = pd.DataFrame()
-        self.sections = sections
-        self.secure = secure
-        self.reset_index = reset_index
-        self.volta_structure = {}
-        self.cols = {'barline': 'voice/BarLine/subtype',
-                     'breaks': 'LayoutBreak/subtype',
-                     'dont_count': 'irregular',
-                     'endRepeat': 'endRepeat',
-                     'jump_bwd': 'Jump/jumpTo',
-                     'jump_fwd': 'Jump/continueAt',
-                     'keysig_col': 'voice/KeySig/accidental',
-                     'len_col': 'Measure:len',
-                     'markers': 'Marker/label',
-                     'mc': 'mc',
-                     'numbering_offset': 'noOffset',
[…]
