### 2023-08-09 #5195 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
