### 2023-08-10 9cb6b9e
makes make_ml() more readable
