+++ src/ms3/bs4_measures.py
@@ -6,12 +6,11 @@ import numpy as np
 import pandas as pd
 from numpy.typing import NDArray
 
-from .logger import LoggedClass, function_logger
+from .logger import LoggedClass
 
 # region helper functions
 
 
-@function_logger
 def get_volta_structure(
     measures, mc, volta_start, volta_length, frac_col=None, logger=None
 ) -> Dict[int, Dict[int, List[int]]]:
@@ -54,7 +53,6 @@ def get_volta_structure(
     return res
 
 
-@function_logger
 def keep_one_row_each(
     df,
     compress_col,
@@ -167,9 +165,10 @@ def keep_one_row_each(
     return result.drop(columns=differentiating_col) if drop_differentiating else result
 
 
-def make_actdur_col(df, len_col, timesig_col="timesig", name="act_dur"):
-    actdur = df[len_col]
-    actdur = actdur.fillna(df[timesig_col])
+def make_actdur_col(
+    len_col: pd.Series, timesig_col: pd.Series, name: str = "act_dur"
+) -> pd.Series:
+    actdur = len_col.fillna(timesig_col)
     try:
         return actdur.map(frac).rename(name)
     except Exception:
@@ -177,24 +176,34 @@ def make_actdur_col(df, len_col, timesig_col="timesig", name="act_dur"):
         raise
 
 
-def make_keysig_col(df, keysig_col, name="keysig"):
+def make_keysig_col(
+    df: pd.DataFrame, keysig_col: str = "keysig_col", name: str = "keysig"
+) -> pd.Series:
     if keysig_col in df:
         return df[keysig_col].fillna(method="ffill").fillna(0).astype(int).rename(name)
     return pd.Series(0, index=df.index).rename(name)
 
 
-def make_mn_col(df, dont_count, numbering_offset, name="mn"):
+def make_mn_col(
+    df: pd.DataFrame,
+    dont_count: str = "dont_count",
+    numbering_offset: str = "numbering_offset",
+    name="mn",
+) -> pd.Series:
     """Compute measure numbers where one or two columns can influence the counting.
 
-    Parameters
-    ----------
-    df : :obj:`pd.DataFrame`
-        If no other parameters are given, every row is counted, starting from 1.
-    dont_count : :obj:`str`, optional
-        This column has notna() for measures where the option "Exclude from bar count" is activated, NaN otherwise.
-    numbering_offset : :obj:`str`, optional
-        This column has values of the MuseScore option "Add to bar number", which adds
-        notna() values to this and all subsequent measures.
+    Args:
+        df: If no other parameters are given, every row is counted, starting from 1.
+        dont_count:
+            This column has notna() for measures where the option "Exclude from bar count" is activated,
+            NaN otherwise.
+        numbering_offset:
+            This column has values of the MuseScore option "Add to bar number", which adds
+            notna() values to this and all subsequent measures.
+        name:
+
+    Returns:
+
     """
     if dont_count is None:
         mn = pd.Series(range(1, len(df) + 1), index=df.index)
@@ -209,23 +218,33 @@ def make_mn_col(df, dont_count, numbering_offset, name="mn"):
     return mn.rename(name)
 
 
-@function_logger
-def make_next_col(df, volta_structure={}, sections=True, name="next", logger=None):
+def make_next_col(
+    df: pd.DataFrame,
+    volta_structure: Optional[Dict[int, Dict[int, List[int]]]] = None,
+    sections: bool = True,
+    name="next",
+    logger=None,
+) -> pd.Series:
     """Uses a `NextColumnMaker` object to create a column with all MCs that can follow each MC
     (e.g. due to repetitions).
 
-    Parameters
-    ----------
-    df : :obj:`pandas.DataFrame`
[…]
