### 2023-08-11 804fbe3
captures, ignores, and warns about use of deprecated 'only_metadata_fnames' argument
