+++ src/ms3/logger.py
@@ -1,28 +1,31 @@
-import logging, sys, os
+import logging
+import os
+import sys
 from contextlib import contextmanager
-from functools import wraps
 from enum import Enum, unique
-from inspect import stack
-from typing import Iterable, Tuple, List, Set
+from functools import wraps
+from typing import Iterable, List, Set, Tuple
 
 LEVELS = {
-    'DEBUG': logging.DEBUG,
-    'INFO': logging.INFO,
-    'WARNING': logging.WARNING,
-    'ERROR': logging.ERROR,
-    'CRITICAL': logging.CRITICAL,
-    'NOTSET': logging.NOTSET,
-    'D': logging.DEBUG,
-    'I': logging.INFO,
-    'W': logging.WARNING,
-    'E': logging.ERROR,
-    'C': logging.CRITICAL,
-    'N': logging.NOTSET
+    "DEBUG": logging.DEBUG,
+    "INFO": logging.INFO,
+    "WARNING": logging.WARNING,
+    "ERROR": logging.ERROR,
+    "CRITICAL": logging.CRITICAL,
+    "NOTSET": logging.NOTSET,
+    "D": logging.DEBUG,
+    "I": logging.INFO,
+    "W": logging.WARNING,
+    "E": logging.ERROR,
+    "C": logging.CRITICAL,
+    "N": logging.NOTSET,
 }
+
 
 @unique
 class MessageType(Enum):
     """Enumerated constants of message types."""
+
     NO_TYPE = 0  # 0 is reserved as no type message
     MCS_NOT_EXCLUDED_FROM_BARCOUNT_WARNING = 1
     INCORRECT_VOLTA_MN_WARNING = 2
@@ -46,8 +49,10 @@ class MessageType(Enum):
     UNUSED_FINE_MARKER_WARNING = 20
     PLAY_UNTIL_IS_MISSING_LABEL_WARNING = 21
     JUMP_TO_IS_MISSING_LABEL_WARNING = 22
-    MISSING_TIME_SIGNATURE_WARNING = 23 # no timesig through the piece
-    BEGINNING_WITHOUT_TIME_SIGNATURE_WARNING = 24 # no timesig in more than only the first measure (which could be an incipit)
+    MISSING_TIME_SIGNATURE_WARNING = 23  # no timesig through the piece
+    BEGINNING_WITHOUT_TIME_SIGNATURE_WARNING = (
+        24  # no timesig in more than only the first measure (which could be an incipit)
+    )
     INVALID_REPEAT_STRUCTURE = 25
     UNFOLDING_REPEATS_FAILED_WARNING = 26
     DCML_DEFAULT_CORRECTION_WARNING = 27
@@ -58,24 +63,46 @@ class MessageType(Enum):
     FRICTIONLESS_VALIDATION_ERROR_WARNING = 32
 
 
-
 class CustomFormatter(logging.Formatter):
     """Formats message depending on whether there is a specified message type"""
+
     def format(self, record):
         if not hasattr(record, "_message_type"):
-            raise ValueError(f"Logger {record.name} has not been correctly defined and is missing the default _message_type field.")
-        msg = record.msg.replace('\n', '\n\t')
+            raise ValueError(
+                f"Logger {record.name} has not been correctly defined and is missing the default _message_type field."
+            )
+        msg = record.msg.replace("\n", "\n\t")
         if record._message_type == 0:  # if there is no message type
-            record.msg = '%-8s %s -- %s (line %s) %s():\n\t%s' % (record.levelname, record.name, record.pathname, record.lineno, record.funcName, msg)
+            record.msg = "%-8s %s -- %s (line %s) %s():\n\t%s" % (
+                record.levelname,
+                record.name,
+                record.pathname,
+                record.lineno,
+                record.funcName,
+                msg,
[…]
