### 2023-08-11 0230b53
changes loglevel from DEBUG to INFO during test
