+++ tests/test_metarepo_files/test_parse.py
@@ -1,11 +1,10 @@
 import os
 
-from ms3 import Parse, first_level_subdirs, capture_parse_logs, ignored_warnings2dict
-from ms3.logger import iter_ms3_loggers, get_logger
+from ms3 import Parse, capture_parse_logs, first_level_subdirs, ignored_warnings2dict
+from ms3.logger import get_logger, iter_ms3_loggers
 
 
-class TestEquivalence():
-
+class TestEquivalence:
     def test_parallel(self, directory):
         a = Parse(directory)
         b = Parse(directory)
@@ -24,28 +23,32 @@ class TestEquivalence():
         assert a.count_extensions() == c.count_extensions()
         assert len(b.count_extensions()) == 1
 
+
 def assert_all_loggers_level(level):
     for name, logger in iter_ms3_loggers():
-        if not name.startswith('ms3.Parse'):
+        if not name.startswith("ms3.Parse"):
             continue
         eff_level = logger.getEffectiveLevel()
         if eff_level != level:
-            head = get_logger('ms3.Parse')
-            print(f"LOGGER '{logger.name}' SHOULD HAVE LEVEL {level}, not {eff_level}. ms3.Parse: {head}")
+            head = get_logger("ms3.Parse")
+            print(
+                f"LOGGER '{logger.name}' SHOULD HAVE LEVEL {level}, not {eff_level}. ms3.Parse: {head}"
+            )
         assert eff_level == level
         for h in logger.handlers:
             h_level = h.level
             if eff_level != level:
-                print(f"THE {h.__class__} of {logger.name} SHOULD HAVE LEVEL {level}, not {h_level}.")
+                print(
+                    f"THE {h.__class__} of {logger.name} SHOULD HAVE LEVEL {level}, not {h_level}."
+                )
             assert h_level == level
 
 
-class TestLogging():
-
+class TestLogging:
     def test_parallel_log_capture(self, small_directory):
         """Compare log messages emitted when parsing the same thing in parallel or iteratively."""
-        a = Parse(small_directory, level='d')
-        b = Parse(small_directory, level='d')
+        a = Parse(small_directory, level="i")
+        b = Parse(small_directory, level="i")
         with capture_parse_logs(a.logger) as captured_msgs:
             a.parse_scores(parallel=False)
             non_parallel_msgs = captured_msgs.content_list
@@ -62,61 +65,59 @@ class TestLogging():
         assert_all_loggers_level(30)
 
     def test_debug(self, small_directory):
-        p = Parse(small_directory, level='d')
+        p = Parse(small_directory, level="d")
         p.parse()
         _ = p.get_dataframes(expanded=True)
         assert_all_loggers_level(10)
 
-
     def test_info(self, small_directory):
-        p = Parse(small_directory, level='i')
+        p = Parse(small_directory, level="i")
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
             all_msgs = captured_msgs.content_list
         assert_all_loggers_level(20)
-        wrong_msgs = [msg for msg in all_msgs if msg.startswith('DEBUG')]
+        wrong_msgs = [msg for msg in all_msgs if msg.startswith("DEBUG")]
         if len(wrong_msgs) > 0:
-            print('\n'.join(wrong_msgs))
+            print("\n".join(wrong_msgs))
             assert False
 
     def test_warning(self, small_directory):
-        p = Parse(small_directory, level='w')
+        p = Parse(small_directory, level="w")
         with capture_parse_logs(p.logger) as captured_msgs:
             p.parse()
             _ = p.get_dataframes(expanded=True)
[…]
