### 2023-08-11 abf5b45
includes logger parent in Corpus.parse_musescore_file because multiprocessing was changing the logger's parent (bug unresolved after a whole day's search!)
