+++ src/ms3/corpus.py
@@ -1,30 +1,77 @@
-from functools import cache
-from logging import Logger
-from pprint import pformat
-from typing import Literal, Collection, Dict, List, Union, Tuple, Iterator, Optional, Set
-
-import os, re
-from itertools import zip_longest
-import pathos.multiprocessing as mp
+import os
+import re
 from collections import Counter, defaultdict
+from functools import cache
+from itertools import zip_longest
+from logging import Logger
+from typing import (
+    Collection,
+    Dict,
+    Iterator,
+    List,
+    Literal,
+    Optional,
+    Set,
+    Tuple,
+    Union,
+)
 
-import pandas as pd
 import numpy as np
-
+import pandas as pd
+import pathos.multiprocessing as mp
 from ms3.utils.frictionless_helpers import store_dataframe_resource
 from ms3.utils.functions import compute_path_from_file
 
-from .logger import LoggedClass, get_logger, get_log_capture_handler, normalize_logger_name
+from ._typing import (
+    AnnotationsFacet,
+    Facet,
+    FacetArguments,
+    FileDataframeTuple,
+    FileDict,
+    FileList,
+    FileParsedTuple,
+    FileScoreTuple,
+    ParsedFile,
+    ScoreFacet,
+    ScoreFacets,
+)
+from .logger import (
+    LoggedClass,
+    get_log_capture_handler,
+    get_logger,
+    normalize_logger_name,
+)
 from .piece import Piece
 from .score import Score, compare_two_score_objects
-from ._typing import FileDict, FileList, ParsedFile, FileParsedTuple, ScoreFacets, FacetArguments, \
-    Facet, ScoreFacet, FileScoreTuple, FileDataframeTuple, AnnotationsFacet
-from .utils import File, column_order, get_musescore, join_tsvs, load_tsv, path2type, \
-    pretty_dict, resolve_dir, update_labels_cfg, write_metadata, available_views2str, prepare_metadata_for_writing, \
-    files2disambiguation_dict, ask_user_to_choose, resolve_paths_argument, make_file_path, resolve_facets_param, check_argument_against_literal_type, \
-    convert, string2identifier, write_markdown, parse_ignored_warnings_file, parse_tsv_file_at_git_revision, disambiguate_files, enforce_piece_index_for_metadata, \
-    scan_directory
-from .utils.constants import METADATA_COLUMN_ORDER, LATEST_MUSESCORE_VERSION
+from .utils import (
+    File,
+    ask_user_to_choose,
+    available_views2str,
+    check_argument_against_literal_type,
+    column_order,
+    convert,
+    disambiguate_files,
+    enforce_piece_index_for_metadata,
+    files2disambiguation_dict,
+    get_musescore,
+    join_tsvs,
+    load_tsv,
+    make_file_path,
+    parse_ignored_warnings_file,
+    parse_tsv_file_at_git_revision,
+    path2type,
+    prepare_metadata_for_writing,
+    pretty_dict,
+    resolve_dir,
+    resolve_facets_param,
+    resolve_paths_argument,
+    scan_directory,
+    string2identifier,
+    update_labels_cfg,
+    write_markdown,
+    write_metadata,
+)
+from .utils.constants import LATEST_MUSESCORE_VERSION, METADATA_COLUMN_ORDER
 from .view import DefaultView, View, create_view_from_parameters
 
 
@@ -33,24 +80,26 @@ class Corpus(LoggedClass):
     Collection of scores and TSV files that can be matched to each other based on their file names.
     """
 
-    default_count_index = pd.MultiIndex.from_product([
-        ('scores',) + Score.dataframe_types,
-        ('found', 'parsed')
-    ])
+    default_count_index = pd.MultiIndex.from_product(
[…]
