### 2023-08-13 67c57fb
adds 'profile=black' option to isort pre-commit hook for improved compatibility
