### 2023-08-13 9ed7b79
use Fraction rather than 'frac' alias
