### 2023-08-13 b2ded48
adds handler to 'ms3' top-level logger, too, as fallback
