### 2023-08-13 bb0c6f3
replaces @function_logger with module_logger; avoids bare except:
