### 2023-08-13 374e024
eliminates function_logger entirely
