+++ setup.cfg
@@ -39,7 +39,7 @@ install_requires =
   beautifulsoup4
   frictionless[zenodo,pandas,visidata]==5.15.1
   lxml
-  pandas==1.5.3
+  pandas>=1.5.3
   pathos
   webcolors
   tqdm
