### 2023-08-13 01871f8
adds pre-commit package to ms3[dev]
