+++ setup.cfg
@@ -66,6 +66,7 @@ testing =
     pytest-cov
 dev =
     ms3[testing]
+    pre-commit
 
 [options.entry_points]
 # Add here console scripts like:
