+++ tests/test_local_files/test_score_object.py
@@ -1,25 +1,40 @@
 #!/usr/bin/env python
 """Tests for `ms3` package."""
 
-import pytest
 import os
 import tempfile
 
 import ms3
-from ms3.utils import assert_all_lines_equal, assert_dfs_equal, decode_harmonies, load_tsv, write_tsv, no_collections_no_booleans
+import pytest
+from ms3.utils import (
+    assert_all_lines_equal,
+    assert_dfs_equal,
+    decode_harmonies,
+    load_tsv,
+    no_collections_no_booleans,
+)
 
 
 @pytest.fixture(
-    params=['Did03M-Son_regina-1762-Sarti.mscx', 'D973deutscher01.mscx', '05_symph_fant.mscx', 'BWV_0815.mscx', 'K281-3.mscx', '76CASM34A33UM.mscx', 'stabat_03_coloured.mscx'],
-ids=['sarti', "schubert", "berlioz", 'bach', 'mozart', 'monty', 'pergolesi'])
+    params=[
+        "Did03M-Son_regina-1762-Sarti.mscx",
+        "D973deutscher01.mscx",
+        "05_symph_fant.mscx",
+        "BWV_0815.mscx",
+        "K281-3.mscx",
+        "76CASM34A33UM.mscx",
+        "stabat_03_coloured.mscx",
+    ],
+    ids=["sarti", "schubert", "berlioz", "bach", "mozart", "monty", "pergolesi"],
+)
 def score_object(request):
     test_folder = os.path.dirname(os.path.realpath(__file__))
-    mscx_path = os.path.join(test_folder, 'MS3', request.param)
+    mscx_path = os.path.join(test_folder, "MS3", request.param)
     s = ms3.Score(mscx_path)
     return s
 
-class TestBasic:
 
+class TestBasic:
     def test_init(self):
         s = ms3.Score()
         assert isinstance(s, ms3.score.Score)
@@ -28,91 +43,122 @@ class TestBasic:
 
 
 class TestScore:
-
     test_folder = os.path.dirname(os.path.realpath(__file__))
-    test_results = os.path.join(test_folder, 'test_results')
+    test_results = os.path.join(test_folder, "test_results")
 
     def test_parse_and_write_back(self, score_object):
-        original_mscx = score_object.full_paths['mscx']
+        original_mscx = score_object.full_paths["mscx"]
         try:
-            tmp_file = tempfile.NamedTemporaryFile(mode='r', suffix='.mscx', dir=self.test_folder, encoding='utf-8', delete=False)
+            tmp_file = tempfile.NamedTemporaryFile(
+                mode="r",
+                suffix=".mscx",
+                dir=self.test_folder,
+                encoding="utf-8",
+                delete=False,
+            )
             if score_object.mscx.has_annotations:
-                score_object.detach_labels('labels')
-                score_object.attach_labels('labels')
+                score_object.detach_labels("labels")
+                score_object.attach_labels("labels")
             score_object.store_score(tmp_file.name)
-            original = open(original_mscx, encoding='utf-8').read()
+            original = open(original_mscx, encoding="utf-8").read()
             after_parsing = tmp_file.read()
-            assert_all_lines_equal(original, after_parsing, original=original_mscx, tmp_file=tmp_file)
+            assert_all_lines_equal(
+                original, after_parsing, original=original_mscx, tmp_file=tmp_file
+            )
         finally:
             tmp_file.close()
             os.remove(tmp_file.name)
 
-
-
     def test_store_and_load_labels(self, score_object):
         if score_object.mscx.has_annotations:
-            piece_name = score_object.fnames['mscx'] + '_labels.tsv'
+            piece_name = score_object.fnames["mscx"] + "_labels.tsv"
[…]
