### 2023-08-13 56eca56
replaces test_parse_to_measurelist() by the equivalent but simpler test_parse_to_measures_table() by factoring out the fixture target_measures_table() and replacing the NamedTemporaryFile with PyTest's the tmp_path fixture
