+++ tests/test_local_files/test_score_object.py
@@ -6,6 +6,7 @@ import tempfile
 
 import ms3
 import pytest
+from ms3.bs4_measures import MeasureList
 from ms3.utils import (
     assert_all_lines_equal,
     assert_dfs_equal,
@@ -45,6 +46,22 @@ class TestBasic:
 class TestScore:
     test_folder = os.path.dirname(os.path.realpath(__file__))
     test_results = os.path.join(test_folder, "test_results")
+
+    @pytest.fixture()
+    def measure_list_object(self, score_object):
+        return MeasureList(
+            score_object._measures,
+            sections=True,
+            secure=True,
+            reset_index=True,
+            logger_cfg=dict(score_object.logger_cfg),
+        )
+
+    @pytest.fixture()
+    def target_measures_table(self, score_object):
+        piece_name = score_object.fnames["mscx"] + "_measures.tsv"
+        target_path = os.path.join(self.test_results, piece_name)
+        return load_tsv(target_path)
 
     def test_parse_and_write_back(self, score_object):
         original_mscx = score_object.full_paths["mscx"]
@@ -119,28 +136,16 @@ class TestScore:
             finally:
                 os.remove(tmp_file.name)
 
-    def test_parse_to_measurelist(self, score_object):
-        piece_name = score_object.fnames["mscx"] + "_measures.tsv"
-        target_path = os.path.join(self.test_results, piece_name)
-        target_measurelist = load_tsv(target_path)
-        try:
-            extracted_measurelist = no_collections_no_booleans(
-                score_object.mscx.measures()
-            )
-            with tempfile.NamedTemporaryFile(
-                mode="r+",
-                suffix=".tsv",
-                dir=self.test_folder,
-                encoding="utf-8",
-                delete=False,
-            ) as tmp_file:
-                extracted_measurelist.to_csv(tmp_file, sep="\t", index=False)
-                new_path = tmp_file.name
-            new_measurelist = load_tsv(new_path)
-            assert len(new_measurelist) > 0
-            assert_dfs_equal(target_measurelist, new_measurelist)
-        finally:
-            os.remove(tmp_file.name)
+    def test_parse_to_measures_table(
+        self, score_object, target_measures_table, tmp_path
+    ):
+        extracted_measurelist = no_collections_no_booleans(score_object.mscx.measures())
+
+        tmp_file = tmp_path / (score_object.fnames["mscx"] + "_measures.tsv")
+        extracted_measurelist.to_csv(tmp_file, sep="\t", index=False)
+        new_measurelist = load_tsv(tmp_file)
+        assert len(new_measurelist) > 0
+        assert_dfs_equal(target_measures_table, new_measurelist)
 
     def test_parse_to_notelist(self, score_object):
         piece_name = score_object.fnames["mscx"] + "_notes.tsv"
