### 2023-08-13 86a10c8
makes function make_offset_col() readable before refactoring it
