+++ src/ms3/bs4_measures.py
@@ -296,6 +296,7 @@ def make_offset_col(
     """If one MN is composed of two MCs, the resulting column indicates the second MC's offset from the MN's beginning.
 
     Args:
+        df: Raw measures table that comes with the indicated columns.
         mc_col, timesig, act_dur, next_col: Names of the required columns.
         section_breaks:
             If you pass the name of a column, the string 'section' is taken into account
@@ -304,45 +305,54 @@ def make_offset_col(
     """
     if logger is None:
         logger = module_logger
-    nom_dur = df[timesig].map(Fraction)
-    sel = df["act_dur"] < nom_dur
-    if sel.sum() == 0:
+    nominal_duration = df[timesig].map(Fraction)
+    shorter_than_nominal_mask = df["act_dur"] < nominal_duration
+    if shorter_than_nominal_mask.sum() == 0:
+        logger.debug(
+            "Actual durations do not diverge from nominal durations, hence mc_offset=0 everywhere."
+        )
         return pd.Series(0, index=df.index, name=name)
 
     if (
         section_breaks is not None
-        and len(df[df[section_breaks].fillna("") == "section"]) == 0
+        and (df[section_breaks].fillna("") == "section").sum() == 0
     ):
+        logger.debug(
+            f"No section breaks in column {section_breaks!r} to be taken into account."
+        )
         section_breaks = None
-    cols = [mc_col, next_col]
-    if section_breaks is not None:
-        cols.append(section_breaks)
+
+    columns_to_display = [mc_col, next_col]
+    if section_breaks is None:
+        mc2mc_offset = {}
+    else:
+        columns_to_display.append(section_breaks)
         last_mc = df[mc_col].max()
-        offsets = {
+        mc2mc_offset = {
             m: 0
             for m in df[df[section_breaks].fillna("") == "section"].mc + 1
             if m <= last_mc
         }
-        # offset == 0 is a neutral value but the presence of mc in offsets indicates that it could potentially be an
-        # (incomplete) pickup measure which can be offset even if the previous measure is complete
-    else:
-        offsets = {}
-    nom_durs = dict(df[[mc_col]].join(nom_dur).itertuples(index=False))
-    act_durs = dict(df[[mc_col, act_dur]].itertuples(index=False))
+        # offset == 0 is a neutral value but the presence of mc in mc2mc_offset indicates that it could potentially be
+        # an (incomplete) pickup measure which can be offset even if the previous measure is complete
 
-    def missing(mc):
-        return nom_durs[mc] - act_durs[mc]
+    mc_indexed = df.set_index(mc_col)
+    nominal_duration.index = mc_indexed.index
+    actual_duration = mc_indexed[act_dur]
+
+    def expected_completion(mc) -> Fraction:
+        return nominal_duration[mc] - actual_duration[mc]
 
     def add_offset(mc, val=None):
         if val is None:
-            val = missing(mc)
-        offsets[mc] = val
+            val = expected_completion(mc)
+        mc2mc_offset[mc] = val
 
-    irregular = df.loc[sel, cols]
+    irregular = df.loc[shorter_than_nominal_mask, columns_to_display]
     if irregular[mc_col].iloc[0] == 1:
         # Check whether first MC is an anacrusis and mark accordingly
         if len(irregular) > 1 and irregular[mc_col].iloc[1] == 2:
-            if not missing(1) + act_durs[2] == nom_durs[1]:
+            if not expected_completion(1) + actual_duration[2] == nominal_duration[1]:
                 add_offset(1)
             else:
                 # regular divided measure, no anacrusis
@@ -350,35 +360,35 @@ def make_offset_col(
         else:
             # is anacrusis
             add_offset(1)
-    for t in irregular.itertuples(index=False):
[…]
