### 2023-08-13 5b397f4
makes is_range_index_equivalent() compatible with pandas 2
