### 2023-08-13 3800a11
removes wrong argument 'flat' from operations._transform() generator
