### 2023-08-13 79c37d1
renames validate_descriptor_path() => validate_descriptor_at_path()
