### 2023-08-13 e13efb4
small bug when directory is empty string
