### 2023-08-14 4a777cf
removes PYTHONWARNINGS env
