### 2025-07-13 #5112 @allorens
Hello,

Hope this finds you well.

I would like to request a certificate for my review activities for JOSS in 2024.

Best regards,

Dr. Ana Llorens
Profesora Ayudante Doctora / Assistant Professor
Departamento de Musicología
Universidad Complutense de Madrid

> On 27 Feb 2023, at 23:26, Fabian-Robert Stöter ***@***.***> wrote:
>
>
> @johentsch <https://github.com/johentsch> sorry if it looked like to be a slow response time. One of the reviewers I wanted to have on board @TGabor <https://github.com/TGabor> wasn't available earlier (and still will only able to finish the review end of march). Thanks for your patience.
>
