### 2025-07-15 #5112 @faroit
@allorens not sure if this is something we could provide easily (@arfon, please correct me). We have a new `reviewed_by` section on our website that you could link to, though. Would that be sufficient for you?

https://joss.theoj.org/papers/reviewed_by/@allorens
